Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)

Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)

calculate_pathway_gsea(
  mat,
  pathway_list,
  n_perm = 1000,
  scale = TRUE,
  no_warnings = TRUE
)

Arguments

mat
expression matrix

pathway_list
a list of vectors, each named for a specific pathway, or dataframe

n_perm
Number of permutation for fgsea function. Defaults to 1000.

scale
convert expr_mat into zscores prior to running GSEA?, default = FALSE

no_warnings
suppress warnings from gsea ties

Value

matrix of GSEA NES values, cell types as row names, pathways as column names

Examples

gl <- list(
    "n" = c("PPBP", "LYZ", "S100A9"),
    "a" = c("IGLL5", "GNLY", "FTL")
)

pbmc_avg <- average_clusters(
    mat = pbmc_matrix_small,
    metadata = pbmc_meta,
    cluster_col = "classified"
)

calculate_pathway_gsea(
    mat = pbmc_avg,
    pathway_list = gl
)
#>                       n          a
#> Naive CD4 T  -1.1373073 -1.1376514
#> Memory CD4 T -1.2681728 -1.3000704
#> CD14+ Mono    1.4582318  0.9602883
#> B            -1.0977842  1.1884500
#> CD8 T        -1.3809367 -0.5480161
#> FCGR3A+ Mono  0.9694219  1.0574108
#> NK           -1.3359662  1.0513766
#> DC            1.0786510 -0.8376655
#> Platelet      1.8171964  1.2979486