Returns a list of variable genes based on PCA

Extract genes, i.e. "features", based on the top loadings of principal components formed from the bulk expression data set

feature_select_PCA(
  mat = NULL,
  pcs = NULL,
  n_pcs = 10,
  percentile = 0.99,
  if_log = TRUE
)

Arguments

mat
Expression matrix. Rownames are genes, colnames are single cell cluster name, and values are average single cell expression (log transformed).

pcs
Precalculated pcs if available, will skip over processing on mat.

n_pcs
Number of PCs to selected gene loadings from. See the explore_PCA_corr.Rmd vignette for details.

percentile
Select the percentile of absolute values of PCA loadings to select genes from. E.g. 0.999 would select the top point 1 percent of genes with the largest loadings.

if_log
whether the data is already log transformed

Value

vector of genes

Examples

feature_select_PCA(
  cbmc_ref,
  if_log = FALSE
)

#>    [1] "AHNAK"          "ALDH2"          "AMICA1"         "ANXA2"         
#>    [5] "ANXA5"          "AP1S2"          "AP2S1"          "APLP2"         
#>    [9] "ATF7IP2"        "ATP1B3"         "ATP6V0E2"       "BCL2A1"        
#>   [13] "BID"            "BLVRB"          "BRI3"           "BTK"           
#>   [17] "C19orf38"       "C5AR1"          "CAMK4"          "CAPZA2"        
#>   [21] "CCR7"           "CD14"           "CD2"            "CD27"          
#>   [25] "CD300A"         "CD36"           "CD3D"           "CD3E"          
#>   [29] "CD3G"           "CD6"            "CD68"           "CD69"          
#>   [33] "CD7"            "CD74"           "CD86"           "CD96"          
#>   [37] "CDKN1A"         "CEBPB"          "CEBPD"          "CFD"           
#>   [41] "CFP"            "CHI3L2"         "CHRM3-AS2"      "CLEC10A"       
#>   [45] "CLEC12A"        "CLEC4A"         "CLEC7A"         "CMTM6"         
#>   [49] "CNPY3"          "CPPED1"         "CPVL"           "CSF2RA"        
#>   [53] "CSF3R"          "CST3"           "CSTA"           "CTSA"          
#>   [57] "CTSH"           "CTSS"           "CTSZ"           "CXCL16"        
#>   [61] "CYBB"           "EFHD2"          "ETS1"           "EVL"           
#>   [65] "F13A1"          "FCER1G"         "FCGR3A"         "FCN1"          
#>   [69] "FGL2"           "FGR"            "FLT3LG"         "G0S2"          
#>   [73] "GCA"            "GLIPR1"         "GPR171"         "GPX1"          
#>   [77] "GRN"            "GSN"            "GSTP1"          "GYPC"          
#>   [81] "H2AFJ"          "HCK"            "HERPUD1"        "HHEX"          
#>   [85] "HLA-DMA"        "HLA-DMB"        "HLA-DPA1"       "HLA-DPB1"      
#>   [89] "HLA-DQA1"       "HLA-DQB1"       "HLA-DRA"        "HLA-DRB1"      
#>   [93] "HLA-DRB5"       "HSBP1"          "IFI27L2"        "IFI30"         
#>   [97] "IFITM1"         "IFNGR1"         "IGSF6"          "IL32"          
#>  [101] "IL7R"           "IRF8"           "ITM2A"          "JUN"           
#>  [105] "KLF2"           "KLF4"           "KLRB1"          "LBH"           
#>  [109] "LCK"            "LDHB"           "LEF1"           "LGALS1"        
#>  [113] "LGALS2"         "LGALS3"         "LILRA5"         "LINC00861"     
#>  [117] "LST1"           "LTB"            "LY86"           "LY96"          
#>  [121] "LYN"            "LYZ"            "MAP3K8"         "MEF2C"         
#>  [125] "MGAT4A"         "MGST1"          "MGST2"          "MNDA"          
#>  [129] "MPEG1"          "MS4A6A"         "MS4A7"          "MYC"           
#>  [133] "MYO1F"          "NAMPT"          "NCF2"           "NEAT1"         
#>  [137] "NOSIP"          "NPC2"           "NRGN"           "OCIAD2"        
#>  [141] "PCED1B-AS1"     "PF4"            "PHACTR1"        "PIK3IP1"       
#>  [145] "PILRA"          "PLAUR"          "PLBD1"          "PLEK"          
#>  [149] "PRKCQ-AS1"      "PSAP"           "PTPRCAP"        "PTPRE"         
#>  [153] "PYCARD"         "RAB32"          "RCAN3"          "RGS1"          
#>  [157] "RGS18"          "RGS2"           "RHOC"           "RHOH"          
#>  [161] "RNASE6"         "RNF130"         "RP11-1143G9.4"  "RP11-290F20.3" 
#>  [165] "RP11-664D1.1"   "S100A11"        "S100A12"        "S100A9"        
#>  [169] "SAT1"           "SATB1"          "SERINC5"        "SERPINA1"      
#>  [173] "SH3YL1"         "SKAP2"          "SLC7A7"         "SOX4"          
#>  [177] "SPI1"           "SRGN"           "STK17A"         "STMN3"         
#>  [181] "STX11"          "STXBP2"         "TAX1BP3"        "TCF7"          
#>  [185] "TGFBI"          "THEMIS2"        "TIMP1"          "TMEM167A"      
#>  [189] "TNFAIP3"        "TNFRSF1B"       "TNFRSF25"       "TNFSF13B"      
#>  [193] "TPM4"           "TRAC"           "TRBC2"          "TSPO"          
#>  [197] "TYMP"           "TYROBP"         "VCAN"           "YWHAH"         
#>  [201] "ZEB2"           "ZYX"            "ADD3"           "AGTRAP"        
#>  [205] "AHSP"           "ALAS2"          "ANXA1"          "ARRB2"         
#>  [209] "ATP6V1B2"       "BCL11A"         "BCL11B"         "BCL2A1"        
#>  [213] "BIN2"           "BLOC1S1"        "BTG1"           "C10orf54"      
#>  [217] "C11orf21"       "C1orf162"       "C5AR1"          "CA1"           
#>  [221] "CAMK4"          "CARD16"         "CASP1"          "CAT"           
#>  [225] "CD14"           "CD3D"           "CD3G"           "CD44"          
#>  [229] "CD48"           "CD52"           "CD59"           "CD7"           
#>  [233] "CD97"           "CEBPB"          "CEBPD"          "CFD"           
#>  [237] "CFP"            "CHPT1"          "CKS2"           "CLEC12A"       
#>  [241] "CLEC7A"         "CORO1A"         "COTL1"          "CST3"          
#>  [245] "CSTA"           "CTSS"           "CTSW"           "CXXC5"         
#>  [249] "CYB561A3"       "CYTL1"          "DANCR"          "DDAH2"         
#>  [253] "DOK2"           "DUSP1"          "EBPL"           "EMP3"          
#>  [257] "EVI2A"          "EVL"            "FAM213A"        "FCER1G"        
#>  [261] "FCGR3A"         "FCN1"           "FGL2"           "FKBP4"         
#>  [265] "FOS"            "FYB"            "G0S2"           "GADD45GIP1"    
#>  [269] "GIMAP4"         "GIMAP7"         "GLIPR2"         "GPSM3"         
#>  [273] "GRAP2"          "GYPA"           "GYPB"           "GZMA"          
#>  [277] "GZMM"           "HBA1"           "HBA2"           "HBB"           
#>  [281] "HBE1"           "HBEGF"          "HBG1"           "HBG2"          
#>  [285] "HBM"            "HBZ"            "HCST"           "HEMGN"         
#>  [289] "HIST1H4C"       "HLA-A"          "HLA-B"          "HLA-C"         
#>  [293] "HLA-E"          "HLA-F"          "HMBS"           "HMGA1"         
#>  [297] "HSPD1"          "ICAM3"          "ID2"            "IER2"          
#>  [301] "IFI30"          "IFITM1"         "IFITM2"         "IFITM3"        
#>  [305] "IGJ"            "IL32"           "IL7R"           "IL8"           
#>  [309] "IRS2"           "ISG15"          "ITGB2"          "ITM2C"         
#>  [313] "JUNB"           "KIAA0101"       "KIAA0125"       "KLF2"          
#>  [317] "KLRB1"          "LCK"            "LGALS1"         "LGALS2"        
#>  [321] "LILRA4"         "LINC00861"      "LINC00936"      "LY96"          
#>  [325] "LYST"           "LYZ"            "MCL1"           "MFSD2B"        
#>  [329] "MNDA"           "MPC2"           "MS4A7"          "MT2A"          
#>  [333] "MYB"            "MYBL2"          "MYL6B"          "MYO1F"         
#>  [337] "MZB1"           "NAMPT"          "NASP"           "NFKBIA"        
#>  [341] "NGLY1"          "NUCKS1"         "NUP214"         "OSTC"          
#>  [345] "PA2G4"          "PARP1"          "PCED1B-AS1"     "PEBP1"         
#>  [349] "PELI1"          "PLAUR"          "PLD4"           "POU2F2"        
#>  [353] "PPP1R14B"       "PRDX2"          "PRSS57"         "PTPRC"         
#>  [357] "PTTG1"          "RARRES3"        "RASSF3"         "RGS18"         
#>  [361] "RP11-1143G9.4"  "RP11-290F20.3"  "RP11-620J15.3"  "RPL22L1"       
#>  [365] "S100A10"        "S100A11"        "S100A12"        "S100A8"        
#>  [369] "S100A9"         "SAMHD1"         "SAMSN1"         "SAT2"          
#>  [373] "SERPINA1"       "SLC11A1"        "SLC2A1"         "SLC2A3"        
#>  [377] "SMARCB1"        "SNCA"           "SNRPD1"         "SOCS3"         
#>  [381] "STMN1"          "STXBP2"         "TC2N"           "TCF4"          
#>  [385] "TCF7"           "TIMP1"          "TMEM14C"        "TMEM71"        
#>  [389] "TNFAIP3"        "TOB1"           "TRAC"           "TRBC2"         
#>  [393] "TRDC"           "TSC22D3"        "TSPAN13"        "TUBA4A"        
#>  [397] "TXN"            "TXNIP"          "TYMP"           "TYMS"          
#>  [401] "TYROBP"         "UPP1"           "VCAN"           "WIPF1"         
#>  [405] "ABHD13"         "ACO2"           "ACTN4"          "ADD1"          
#>  [409] "AHNAK"          "AKAP13"         "AKIRIN2"        "ALOX5AP"       
#>  [413] "ANXA1"          "ANXA2"          "APLP2"          "ARF3"          
#>  [417] "ATF6B"          "BAG1"           "BIN2"           "C1orf56"       
#>  [421] "CARD16"         "CASP8"          "CAV1"           "CAV2"          
#>  [425] "CCDC88B"        "CCNG2"          "CD200"          "CD247"         
#>  [429] "CD320"          "CD40LG"         "CD58"           "CENPK"         
#>  [433] "CLIC1"          "COA1"           "COMMD8"         "COQ10B"        
#>  [437] "CSRP1"          "CST7"           "CTA-445C9.15"   "DCTN3"         
#>  [441] "DCTN6"          "DDB1"           "DDX3X"          "DHX16"         
#>  [445] "DKC1"           "DLST"           "DUSP5"          "EBPL"          
#>  [449] "EGFL7"          "EIF3I"          "EIF4EBP1"       "ELAVL1"        
#>  [453] "EPHB6"          "EXD3"           "FABP4"          "FAM129C"       
#>  [457] "FGFR1"          "FOSL2"          "GATA3"          "GIMAP1"        
#>  [461] "GLMN"           "GLUD1"          "GMPR2"          "GNAI1"         
#>  [465] "GNAI2"          "GNB4"           "GNG11"          "GPCPD1"        
#>  [469] "GZMB"           "H2AFZ"          "HLA-DMA"        "HLA-DMB"       
#>  [473] "HLA-DPA1"       "HLA-DPB1"       "HLA-DQA1"       "HLA-DQB1"      
#>  [477] "HOTAIRM1"       "ID1"            "ID3"            "IFI27"         
#>  [481] "IFRD1"          "IGHD"           "IGHM"           "IGJ"           
#>  [485] "IL11RA"         "IL32"           "ILF2"           "IRF8"          
#>  [489] "ISCU"           "ITGAE"          "ITPA"           "KANK3"         
#>  [493] "KLRB1"          "KLRG1"          "LAS1L"          "LGALS9"        
#>  [497] "LINC00152"      "LMNA"           "LMO4"           "LPXN"          
#>  [501] "MACF1"          "MARCKS"         "MDM4"           "METAP1"        
#>  [505] "MGST3"          "MIEN1"          "MPO"            "MS4A1"         
#>  [509] "MT2A"           "NCR3"           "NDN"            "NDUFA12"       
#>  [513] "NDUFS3"         "NFE2L2"         "NOP58"          "NPEPPS"        
#>  [517] "NR4A2"          "NRD1"           "NUP54"          "OAZ2"          
#>  [521] "OXCT1"          "OXLD1"          "PCBD1"          "PHACTR2"       
#>  [525] "PHF15"          "PLAGL1"         "PNPT1"          "PPP1R7"        
#>  [529] "PPP2CA"         "PPP2R2A"        "PRSS57"         "PSMD6"         
#>  [533] "PTGES3"         "PTPRA"          "RAD23B"         "RAF1"          
#>  [537] "RAMP2"          "RANGRF"         "RASSF1"         "RBM4"          
#>  [541] "RC3H1"          "RCBTB2"         "RCN2"           "RHOC"          
#>  [545] "RIT1"           "RNASE2"         "RNASE3"         "RNASE6"        
#>  [549] "RNF125"         "RP11-192H23.6"  "RP11-428G5.5"   "RP11-544A12.8" 
#>  [553] "RP13-582O9.5"   "RPS3A"          "RTKN2"          "RTN4"          
#>  [557] "SAFB"           "SCGB3A1"        "SCN1B"          "SEPT1"         
#>  [561] "SERPINB6"       "SKAP1"          "SLC30A5"        "SLC35B3"       
#>  [565] "SMIM12"         "SMIM14"         "SPNS3"          "SRSF6"         
#>  [569] "STARD7"         "STAT5A"         "STOML2"         "STRBP"         
#>  [573] "SUPT5H"         "TBC1D4"         "TBRG1"          "TCEB1"         
#>  [577] "TCL1A"          "TES"            "TFPI2"          "TIMP1"         
#>  [581] "TLE4"           "TMEM107"        "TPM1"           "TPST2"         
#>  [585] "TRABD2A"        "TRAV3"          "TUBB4B"         "TXK"           
#>  [589] "TXN"            "UBE2A"          "UQCRFS1"        "USB1"          
#>  [593] "UVRAG"          "VAMP5"          "VDAC1"          "VDAC3"         
#>  [597] "VPS29"          "VTI1B"          "WARS"           "XCL2"          
#>  [601] "ZBTB16"         "ZCCHC7"         "ZDHHC12"        "ZFAND5"        
#>  [605] "ZFP36L1"        "ZMYM2"          "ACAP1"          "ACRBP"         
#>  [609] "AHSP"           "ALAS2"          "ALOX5AP"        "AMICA1"        
#>  [613] "ANKRD9"         "APBB1IP"        "ARHGAP9"        "ARID3A"        
#>  [617] "ARL4C"          "BCL11A"         "BIN1"           "BIRC5"         
#>  [621] "BLNK"           "BLVRB"          "BNIP3L"         "BPGM"          
#>  [625] "BST2"           "C12orf75"       "C19orf10"       "C2orf88"       
#>  [629] "C9orf142"       "CA1"            "CA2"            "CAPG"          
#>  [633] "CCDC50"         "CD37"           "CD4"            "CD48"          
#>  [637] "CD53"           "CD7"            "CD74"           "CDK2AP2"       
#>  [641] "CENPF"          "CLEC4C"         "CLIC3"          "CLN8"          
#>  [645] "CNPY3"          "COMMD6"         "CORO1A"         "CPOX"          
#>  [649] "CXCR4"          "CXXC5"          "CYB561A3"       "CYBA"          
#>  [653] "CYTH4"          "DBI"            "DDIT4"          "DERL3"         
#>  [657] "DMTN"           "DNASE1L3"       "DUSP5"          "EMB"           
#>  [661] "EPB42"          "ERN1"           "FAM129C"        "FAM210B"       
#>  [665] "FAM49B"         "FAM65B"         "FCER1A"         "FECH"          
#>  [669] "FKBP2"          "FNBP1"          "GAPT"           "GAS6"          
#>  [673] "GATA1"          "GLIPR1"         "GMPR"           "GNG7"          
#>  [677] "GP9"            "GPR183"         "GPSM3"          "GRN"           
#>  [681] "GYPA"           "GYPB"           "GZMB"           "H1F0"          
#>  [685] "HBA1"           "HBA2"           "HBB"            "HBD"           
#>  [689] "HBE1"           "HBG1"           "HBG2"           "HBM"           
#>  [693] "HBQ1"           "HBZ"            "HEMGN"          "HERPUD1"       
#>  [697] "HLA-B"          "HLA-C"          "HLA-DMA"        "HLA-DPA1"      
#>  [701] "HLA-DQA1"       "HLA-DQB1"       "HMBS"           "HMGB2"         
#>  [705] "IFIT1B"         "IGJ"            "IGLC3"          "IL2RG"         
#>  [709] "IL3RA"          "IRF4"           "IRF7"           "IRF8"          
#>  [713] "ITM2C"          "KIAA0101"       "KLF1"           "LAMP5"         
#>  [717] "LAPTM4B"        "LILRA4"         "LILRB4"         "LIMS1"         
#>  [721] "LINC00996"      "LMNA"           "LTB"            "LYL1"          
#>  [725] "MAPKAPK2"       "MIF4GD"         "MINPP1"         "MIR4435-1HG"   
#>  [729] "MPP1"           "MYL4"           "MZB1"           "NCOA4"         
#>  [733] "NFE2"           "NGFRAP1"        "NPC2"           "OFD1"          
#>  [737] "OPN3"           "PCED1B-AS1"     "PLA2G16"        "PLAC8"         
#>  [741] "PLD4"           "PLP2"           "PPIB"           "PPP1R14B"      
#>  [745] "PRDX2"          "PRKAR2B"        "PTGDS"          "PTPRCAP"       
#>  [749] "PTTG1"          "RABAC1"         "RASD1"          "RFESD"         
#>  [753] "RGCC"           "RGS1"           "RGS19"          "RGS2"          
#>  [757] "RHAG"           "RHCE"           "RNASE6"         "RNASET2"       
#>  [761] "RNU1-60P"       "RRBP1"          "S100A12"        "SAMHD1"        
#>  [765] "SCT"            "SEC61B"         "SEL1L3"         "SELENBP1"      
#>  [769] "SELL"           "SEPT9"          "SERPINF1"       "SLC15A4"       
#>  [773] "SLC25A37"       "SLC38A1"        "SMIM1"          "SMIM14"        
#>  [777] "SMOX"           "SMPD3"          "SNCA"           "SPIB"          
#>  [781] "SPNS3"          "STOM"           "STRADB"         "TAL1"          
#>  [785] "TAP1"           "TAPBP"          "TBCC"           "TCF4"          
#>  [789] "TCL1A"          "TESC"           "TFRC"           "TGFBI"         
#>  [793] "TMCC2"          "TMIGD2"         "TPM1"           "TPM2"          
#>  [797] "TRAF4"          "TRIM22"         "TRIM58"         "TSC22D1"       
#>  [801] "UBAC1"          "UGCG"           "UNC93B1"        "UROD"          
#>  [805] "VIMP"           "YBX3"           "YWHAH"          "ZFP36"         
#>  [809] "AC147651.3"     "ACAP1"          "ACRBP"          "ALOX12"        
#>  [813] "ANXA5"          "AP001189.4"     "ARPC5L"         "BET1"          
#>  [817] "BIN1"           "BIN2"           "BTN3A2"         "BZRAP1-AS1"    
#>  [821] "C12orf75"       "C19orf33"       "C2orf88"        "C6orf25"       
#>  [825] "CALD1"          "CCL3"           "CCL4"           "CCL5"          
#>  [829] "CD160"          "CD247"          "CD3D"           "CD3G"          
#>  [833] "CD69"           "CD7"            "CD9"            "CD99"          
#>  [837] "CHST12"         "CLDN5"          "CLEC10A"        "CLEC1B"        
#>  [841] "CLIC3"          "CLU"            "CMC1"           "CMTM5"         
#>  [845] "CNST"           "CPVL"           "CSF1R"          "CST3"          
#>  [849] "CST7"           "CTSA"           "CTSW"           "CTSZ"          
#>  [853] "CTTN"           "CYFIP2"         "DAB2"           "DNAJB6"        
#>  [857] "DYNLRB1"        "ERV3-1"         "ESAM"           "FCER1A"        
#>  [861] "FCRL6"          "FERMT3"         "FGFBP2"         "FKBP11"        
#>  [865] "GCHFR"          "GMPR"           "GNG11"          "GNG2"          
#>  [869] "GNLY"           "GP1BA"          "GP9"            "GPR56"         
#>  [873] "GRAP2"          "GSAP"           "GSN"            "GZMA"          
#>  [877] "GZMB"           "GZMH"           "GZMM"           "HIST1H1C"      
#>  [881] "HIST1H2AC"      "HIST1H2BJ"      "HIST1H2BK"      "HLA-DMA"       
#>  [885] "HLA-DPA1"       "HLA-DPB1"       "HLA-DQA1"       "HLA-DQB1"      
#>  [889] "HLA-E"          "HOPX"           "IFITM1"         "IGHM"          
#>  [893] "IL2RB"          "IL2RG"          "IL32"           "ITGA2B"        
#>  [897] "JAK1"           "JUN"            "KIF2A"          "KIFAP3"        
#>  [901] "KIFC3"          "KLRB1"          "KLRC1"          "KLRD1"         
#>  [905] "KLRF1"          "LBH"            "LCK"            "LCN2"          
#>  [909] "LDLRAP1"        "LGALS3"         "LGALSL"         "LIMS1"         
#>  [913] "LINC00861"      "LPCAT1"         "LY6G6F"         "LY96"          
#>  [917] "MAGED2"         "MAP3K7CL"       "MARCH2"         "MATK"          
#>  [921] "MAX"            "MIB2"           "MLH3"           "MLLT3"         
#>  [925] "MMD"            "MNDA"           "MYBL1"          "MYL9"          
#>  [929] "MYOM2"          "NCR3"           "NGFRAP1"        "NKG7"          
#>  [933] "NRGN"           "NT5C3A"         "ORMDL3"         "PARP8"         
#>  [937] "PCED1B-AS1"     "PDLIM1"         "PDZK1IP1"       "PF4"           
#>  [941] "PF4V1"          "PGRMC1"         "PIM1"           "PLA2G12A"      
#>  [945] "PPBP"           "PPM1A"          "PRDX6"          "PRF1"          
#>  [949] "PRKAR2B"        "PRKCB"          "PRKCH"          "PRSS23"        
#>  [953] "PTCRA"          "PTGDR"          "PTGDS"          "PTPRCAP"       
#>  [957] "PYHIN1"         "RAB27B"         "RAB37"          "RAC2"          
#>  [961] "RAP1B"          "RARRES3"        "RASGRP1"        "RGS18"         
#>  [965] "RNF11"          "RNF125"         "RP11-367G6.3"   "RPA1"          
#>  [969] "RSU1"           "RUFY1"          "SAMD3"          "SDPR"          
#>  [973] "SEPT1"          "SH2D1A"         "SIDT1-AS1"      "SKAP1"         
#>  [977] "SLA2"           "SMIM5"          "SOCS2"          "SPARC"         
#>  [981] "SPON2"          "STARD3NL"       "STAT4"          "STK17A"        
#>  [985] "STOM"           "SYNE2"          "TBC1D10C"       "TBPL1"         
#>  [989] "TBX21"          "TLK1"           "TMED4"          "TMEM40"        
#>  [993] "TMIGD2"         "TNFSF4"         "TNIK"           "TNNC2"         
#>  [997] "TPM4"           "TPST2"          "TRBC2"          "TREML1"        
#> [1001] "TSC22D1"        "TTC38"          "TUBA4A"         "TUBB1"         
#> [1005] "TXK"            "TYMP"           "UBE2F"          "XCL2"          
#> [1009] "XPNPEP1"        "ZAP70"          "ABI3"           "ACRBP"         
#> [1013] "ADAM19"         "AKIRIN2"        "AKNA"           "ALOX5AP"       
#> [1017] "AOAH"           "AP001189.4"     "APP"            "ARL2BP"        
#> [1021] "ATG2A"          "ATP2A3"         "BASP1"          "C12orf75"      
#> [1025] "C19orf33"       "C19orf66"       "C20orf27"       "C2orf88"       
#> [1029] "C6orf25"        "CA2"            "CASP1"          "CASP3"         
#> [1033] "CASP4"          "CBFA2T3"        "CD2AP"          "CD36"          
#> [1037] "CD69"           "CD79A"          "CD79B"          "CDKN2D"        
#> [1041] "CEBPB"          "CLDN5"          "CLEC1B"         "CLEC4C"        
#> [1045] "CLIC3"          "CLU"            "CMTM5"          "CORO1C"        
#> [1049] "CTTN"           "DAB2"           "DERL3"          "DHRS1"         
#> [1053] "DMTN"           "DNASE1L3"       "DPYSL2"         "DRD4"          
#> [1057] "EFHD2"          "ERO1LB"         "EVL"            "FABP4"         
#> [1061] "FAM129C"        "FAM213A"        "FAM26F"         "FCGR3A"        
#> [1065] "FGFBP2"         "FGR"            "FHL1"           "FMNL1"         
#> [1069] "FXR1"           "FYN"            "GAS6"           "GLTP"          
#> [1073] "GMPR"           "GNG11"          "GNG2"           "GNG7"          
#> [1077] "GP9"            "GPR183"         "GPX1"           "GRAP2"         
#> [1081] "GYPC"           "GZMB"           "HBA1"           "HBD"           
#> [1085] "HBG2"           "HIGD1A"         "HIST1H1C"       "HIST1H2AC"     
#> [1089] "HIST1H2BJ"      "HIST2H2BE"      "HOPX"           "ICAM3"         
#> [1093] "IFI27"          "IFI6"           "IFITM3"         "IGHM"          
#> [1097] "IGJ"            "IGLC2"          "IL16"           "IL3RA"         
#> [1101] "ILK"            "INPP5D"         "IRF1"           "IRF4"          
#> [1105] "ITGA2B"         "ITGB2"          "ITM2C"          "LAMP5"         
#> [1109] "LAT2"           "LCN2"           "LGALSL"         "LGMN"          
#> [1113] "LILRA4"         "LILRB4"         "LINC00493"      "LINC00996"     
#> [1117] "LTB"            "LY6E"           "LY6G6F"         "LY96"          
#> [1121] "MAP1A"          "MAP3K7CL"       "MAPKAPK2"       "MARCH2"        
#> [1125] "MLH3"           "MMD"            "MPEG1"          "MRPL42"        
#> [1129] "MRPL52"         "MS4A6A"         "MYBL2"          "MYL9"          
#> [1133] "MYO1F"          "MZB1"           "NCOA4"          "NDUFA12"       
#> [1137] "NGFRAP1"        "NPC2"           "NRBP1"          "NRGN"          
#> [1141] "ODC1"           "P2RX1"          "PARL"           "PARVB"         
#> [1145] "PF4"            "PF4V1"          "PIK3R1"         "PKIG"          
#> [1149] "PKN1"           "PLD4"           "PLP2"           "PMAIP1"        
#> [1153] "PPBP"           "PPP1R14A"       "PPP1R14B"       "PRKAR2B"       
#> [1157] "PRKCH"          "PTCRA"          "PTGDS"          "PTMS"          
#> [1161] "PTPN6"          "PTPRS"          "RARRES3"        "RASD1"         
#> [1165] "RCSD1"          "RGL4"           "RGS18"          "RIN3"          
#> [1169] "RNASE6"         "RP11-73G16.2"   "RPF1"           "RRBP1"         
#> [1173] "RUNX3"          "SCT"            "SDPR"           "SEPT11"        
#> [1177] "SERPINF1"       "SH3BP1"         "SIGIRR"         "SKAP1"         
#> [1181] "SMARCC1"        "SMIM5"          "SMPD3"          "SNCA"          
#> [1185] "SNRPN"          "SNX9"           "SPARC"          "SPIB"          
#> [1189] "SPNS3"          "SSX2IP"         "SVIP"           "SYTL1"         
#> [1193] "TAGLN2"         "TAX1BP3"        "TCF4"           "TCL1A"         
#> [1197] "TLE4"           "TMEM40"         "TMEM8B"         "TNNC2"         
#> [1201] "TPM2"           "TREML1"         "TSC22D1"        "TSPAN13"       
#> [1205] "TUBA1C"         "TUBA4A"         "TUBB1"          "UBE2J1"        
#> [1209] "UGCG"           "UNC93B1"        "UTRN"           "ZFAT"          
#> [1213] "ADAM28"         "AFF3"           "AHSP"           "ALAS2"         
#> [1217] "ALOX5AP"        "ANKRD28"        "APP"            "ARL4C"         
#> [1221] "ATG2A"          "AVP"            "BANK1"          "BASP1"         
#> [1225] "BCL11A"         "BIRC3"          "BLVRB"          "BPGM"          
#> [1229] "BSG"            "BTK"            "C12orf75"       "C16orf74"      
#> [1233] "C19orf77"       "C1QTNF4"        "C6orf48"        "C9orf142"      
#> [1237] "C9orf89"        "CA1"            "CCL3"           "CCL4"          
#> [1241] "CCL5"           "CD160"          "CD1C"           "CD22"          
#> [1245] "CD247"          "CD34"           "CD36"           "CD37"          
#> [1249] "CD52"           "CD69"           "CD7"            "CD74"          
#> [1253] "CD79A"          "CD79B"          "CD82"           "CD9"           
#> [1257] "CDCA7L"         "CEBPB"          "CHST2"          "CLIC3"         
#> [1261] "CMC1"           "CRHBP"          "CSF3R"          "CST7"          
#> [1265] "CTSC"           "CTSW"           "CXCR4"          "DAPP1"         
#> [1269] "DUSP2"          "EAF2"           "EBPL"           "EGFL7"         
#> [1273] "EIF3E"          "FAM26F"         "FCER1G"         "FCGR3A"        
#> [1277] "FCRL6"          "FCRLA"          "FECH"           "FGFBP2"        
#> [1281] "FGR"            "FHL1"           "GAS5"           "GIMAP4"        
#> [1285] "GNG11"          "GNLY"           "GUK1"           "GYPA"          
#> [1289] "GYPB"           "GZMA"           "GZMB"           "GZMH"          
#> [1293] "GZMK"           "GZMM"           "HAGH"           "HBA1"          
#> [1297] "HBA2"           "HBB"            "HBD"            "HBE1"          
#> [1301] "HBG1"           "HBG2"           "HBM"            "HBZ"           
#> [1305] "HCST"           "HES6"           "HHEX"           "HIST1H2AC"     
#> [1309] "HLA-DMA"        "HLA-DOB"        "HLA-DPA1"       "HLA-DPB1"      
#> [1313] "HLA-DQA1"       "HLA-DQB1"       "HLA-DRA"        "HLA-DRB1"      
#> [1317] "HLA-DRB5"       "HSPB1"          "ICAM3"          "IFIT1B"        
#> [1321] "IFT57"          "IGHD"           "IGHM"           "IGLC2"         
#> [1325] "IGLC3"          "IGLL1"          "IL2RB"          "INPP4A"        
#> [1329] "ITM2C"          "JUP"            "KIAA0125"       "KLRB1"         
#> [1333] "KLRC1"          "KLRD1"          "KLRF1"          "LAPTM4B"       
#> [1337] "LIMS1"          "LINC00861"      "LINC00926"      "LTA4H"         
#> [1341] "LYL1"           "MARCKS"         "MDK"            "MEF2C"         
#> [1345] "MEG3"           "MGST3"          "MLLT3"          "MPO"           
#> [1349] "MS4A1"          "MSI2"           "MSRB3"          "MYBL2"         
#> [1353] "MYL4"           "MYOM2"          "MZB1"           "NAP1L1"        
#> [1357] "NGFRAP1"        "NKG7"           "NPR3"           "NRIP1"         
#> [1361] "ORAI2"          "P2RX5"          "PARP1"          "PAWR"          
#> [1365] "PDE4B"          "PDLIM1"         "PKM"            "PLA2G16"       
#> [1369] "PPP1R14A"       "PRDX1"          "PRF1"           "PRKCH"         
#> [1373] "PTGDS"          "PYHIN1"         "RBM38"          "RCSD1"         
#> [1377] "RHOC"           "RNA5SP151"      "RP5-887A10.1"   "RSL1D1"        
#> [1381] "S1PR5"          "SDPR"           "SELENBP1"       "SELL"          
#> [1385] "SH2D1B"         "SNHG7"          "SNHG8"          "SOCS2"         
#> [1389] "SOX4"           "SPI1"           "SPINK2"         "SPINT2"        
#> [1393] "SPON2"          "SPRY1"          "SSBP2"          "STRBP"         
#> [1397] "SYPL1"          "TCEAL4"         "TCL1A"          "TESC"          
#> [1401] "TMEM123"        "TRDC"           "TSC22D1"        "TSC22D3"       
#> [1405] "TSTA3"          "TTC38"          "TUBA1A"         "TYROBP"        
#> [1409] "VPREB3"         "XCL1"           "XCL2"           "YBX3"          
#> [1413] "ZAP70"          "ZFAS1"          "ADA"            "AIF1"          
#> [1417] "ALDH1A1"        "ALOX5"          "ANKRD28"        "ANXA1"         
#> [1421] "AP001171.1"     "APMAP"          "AREGB"          "ARHGAP24"      
#> [1425] "ATP8B4"         "AVP"            "BAALC"          "BANK1"         
#> [1429] "BEX4"           "BIRC3"          "BLK"            "BST2"          
#> [1433] "BTG1"           "C11orf74"       "C14orf1"        "C19orf77"      
#> [1437] "C1QTNF4"        "C1orf21"        "CCR7"           "CD164"         
#> [1441] "CD19"           "CD1C"           "CD22"           "CD34"          
#> [1445] "CD40"           "CD63"           "CD72"           "CD79A"         
#> [1449] "CD79B"          "CD83"           "CD9"            "CD99"          
#> [1453] "CDCA7"          "CDK6"           "CEBPB"          "CHD7"          
#> [1457] "CHI3L2"         "CHST12"         "CLEC2D"         "COL19A1"       
#> [1461] "COMT"           "CRHBP"          "CRYGD"          "CSF3R"         
#> [1465] "CTA-250D10.23"  "CTSW"           "CYB561A3"       "CYTL1"         
#> [1469] "DDIT4"          "DDOST"          "DSTN"           "EBF1"          
#> [1473] "ECI2"           "EGFL7"          "EML4"           "FCER2"         
#> [1477] "FCRL1"          "FCRLA"          "FLT3"           "GADD45A"       
#> [1481] "GATA2"          "GBP2"           "GCHFR"          "GGA2"          
#> [1485] "GLUL"           "GNA15"          "GNAQ"           "GPR183"        
#> [1489] "GPR56"          "GSN"            "GUCY1A3"        "GZMB"          
#> [1493] "H2AFY"          "HBA2"           "HBG1"           "HLA-DOB"       
#> [1497] "HLA-DQA1"       "HMGN3"          "HOPX"           "HOXA9"         
#> [1501] "HPGDS"          "ID2"            "IFITM3"         "IGF2BP2"       
#> [1505] "IGFBP7"         "IGHD"           "IGHM"           "IGHV1OR15-1"   
#> [1509] "IGJ"            "IGLC2"          "IGLC3"          "IGLC7"         
#> [1513] "IGLL1"          "IGLL5"          "IL4R"           "IL7R"          
#> [1517] "IQGAP2"         "ISG20"          "ISYNA1"         "KIAA1407"      
#> [1521] "KLF2"           "KLF6"           "LAPTM4B"        "LGALS1"        
#> [1525] "LINC00926"      "LTB"            "LY86"           "LY9"           
#> [1529] "MAP3K8"         "MAPKAPK3"       "MARCH1"         "MATK"          
#> [1533] "MBOAT7"         "MDK"            "MEG3"           "MEST"          
#> [1537] "MGST2"          "MLLT3"          "MPO"            "MPST"          
#> [1541] "MS4A1"          "MSRB3"          "MTSS1"          "MYB"           
#> [1545] "NEAT1"          "NGFRAP1"        "NPDC1"          "NPR3"          
#> [1549] "NRIP1"          "NUCB2"          "P2RX5"          "PCDH9"         
#> [1553] "PDLIM1"         "PGD"            "PLAGL1"         "POU2AF1"       
#> [1557] "PPP1R14A"       "PRDX3"          "PROM1"          "PRSS57"        
#> [1561] "PTPN7"          "PTPRE"          "RAB34"          "RALGPS2"       
#> [1565] "RBM38"          "RNA5SP151"      "RNASE6"         "RNF130"        
#> [1569] "RP11-1275H24.1" "RP11-164H13.1"  "RP11-373D23.2"  "RP11-620J15.3" 
#> [1573] "RP11-693J15.5"  "RP11-796E2.4"   "RP5-887A10.1"   "RUNX1"         
#> [1577] "S100A11"        "SCCPDH"         "SDF2"           "SEC13"         
#> [1581] "SEC63"          "SEPT11"         "SEPW1"          "SERPINB1"      
#> [1585] "SERPINB6"       "SMIM14"         "SOCS2"          "SORL1"         
#> [1589] "SOX4"           "SP140"          "SPIB"           "SPINK2"        
#> [1593] "SPN"            "SRGN"           "STARD3NL"       "STMN1"         
#> [1597] "TAGAP"          "TBXAS1"         "TCEAL4"         "TCEAL8"        
#> [1601] "TCL1A"          "TFPI"           "TMEM173"        "TMIGD2"        
#> [1605] "TNFRSF13C"      "TNFSF13B"       "TRDC"           "TSC22D1"       
#> [1609] "TSPAN2"         "TSPO"           "TUBA1B"         "VPREB3"        
#> [1613] "XBP1"           "YWHAE"          "ZFP36L1"        "ZFP36L2"       
#> [1617] "AC013264.2"     "AC017104.6"     "ACTN1"          "ADAM8"         
#> [1621] "AIF1"           "AMICA1"         "ANXA1"          "APBA2"         
#> [1625] "APLP2"          "APMAP"          "APOBEC3C"       "APOBEC3G"      
#> [1629] "AQP3"           "ATG2A"          "ATP1A1"         "BANK1"         
#> [1633] "BASP1"          "BCL11B"         "BLK"            "BST2"          
#> [1637] "C1orf21"        "C1orf228"       "C5orf56"        "CAMK4"         
#> [1641] "CCL3"           "CCL4"           "CCL5"           "CCR7"          
#> [1645] "CD160"          "CD19"           "CD22"           "CD27"          
#> [1649] "CD38"           "CD3D"           "CD3E"           "CD4"           
#> [1653] "CD40"           "CD63"           "CD69"           "CD72"          
#> [1657] "CD74"           "CD79A"          "CD79B"          "CD81"          
#> [1661] "CD8B"           "CFP"            "CHST12"         "CHST2"         
#> [1665] "CITED4"         "CLIC1"          "CMC1"           "COL19A1"       
#> [1669] "CORO1B"         "CST7"           "CTSW"           "CXCR5"         
#> [1673] "CXXC5"          "DDIT4"          "DPP4"           "DUSP1"         
#> [1677] "FAM105A"        "FAM134B"        "FARS2"          "FCER2"         
#> [1681] "FCGR3A"         "FCGRT"          "FCRL1"          "FCRL6"         
#> [1685] "FCRLA"          "FGFBP2"         "FYB"            "GNLY"          
#> [1689] "GPR56"          "GSAP"           "GZMB"           "GZMH"          
#> [1693] "HLA-DOB"        "HLA-DPA1"       "HLA-DPB1"       "HLA-DQA1"      
#> [1697] "HLA-DQB1"       "HLA-DRA"        "HLA-DRB1"       "HLA-DRB5"      
#> [1701] "HOPX"           "HSH2D"          "ID2"            "IGFBP7"        
#> [1705] "IGHD"           "IGHM"           "IGJ"            "IGLC2"         
#> [1709] "IGLC3"          "IGLC7"          "IGLL5"          "IL2RB"         
#> [1713] "IL6ST"          "IL7R"           "IRF8"           "KLRB1"         
#> [1717] "KLRC1"          "KLRD1"          "KLRF1"          "LAT2"          
#> [1721] "LDLRAP1"        "LEF1"           "LEF1-AS1"       "LEPROTL1"      
#> [1725] "LGALS9"         "LINC00926"      "LRRN3"          "LYN"           
#> [1729] "MAL"            "MAP3K8"         "MARCH1"         "MCTP2"         
#> [1733] "MEF2C"          "MGAT4A"         "MIR142"         "MLLT3"         
#> [1737] "MS4A1"          "MTSS1"          "MYOM2"          "NANS"          
#> [1741] "NCR3"           "NDFIP1"         "NEIL1"          "NELL2"         
#> [1745] "NFKBIZ"         "NKG7"           "NOSIP"          "NT5C"          
#> [1749] "NUCB2"          "ORAI2"          "OXNAD1"         "P2RX5"         
#> [1753] "PAG1"           "PARP14"         "PASK"           "PCDH9"         
#> [1757] "PDLIM1"         "PIK3IP1"        "PKIG"           "POU2AF1"       
#> [1761] "PPA1"           "PPAPDC1B"       "PRDM2"          "PRF1"          
#> [1765] "PRKCA"          "PRKCQ-AS1"      "PRSS57"         "PTGDR"         
#> [1769] "PTGER4"         "PTPRCAP"        "RALGPS2"        "RAP1A"         
#> [1773] "RCAN3"          "REG4"           "RGS1"           "RGS10"         
#> [1777] "RP11-164H13.1"  "RP11-415F23.2"  "RP11-664D1.1"   "RP4-594I10.3"  
#> [1781] "RP5-887A10.1"   "S100A8"         "S100B"          "S1PR5"         
#> [1785] "SAMHD1"         "SELL"           "SERINC5"        "SERPINF1"      
#> [1789] "SGK1"           "SH2D1B"         "SH3BP5"         "SIRPG"         
#> [1793] "SMIM14"         "SOCS3"          "SOX4"           "SP140"         
#> [1797] "SPON2"          "TCL1A"          "TLE1"           "TMEM123"       
#> [1801] "TMEM66"         "TNFAIP3"        "TNFRSF18"       "TOB1"          
#> [1805] "TPD52"          "TRABD2A"        "TRAC"           "TRAT1"         
#> [1809] "TRDC"           "TSPAN13"        "TTC38"          "USP10"         
#> [1813] "VAV3"           "VPREB3"         "XCL1"           "XCL2"          
#> [1817] "ZEB2"           "ZYX"            "ACSF2"          "ACTN1"         
#> [1821] "ADAM28"         "ADAM8"          "AGPAT9"         "AIM1"          
#> [1825] "ALCAM"          "ALDH2"          "ALOX5AP"        "ANXA1"         
#> [1829] "APOBEC3A"       "AREGB"          "ARF6"           "ARL4C"         
#> [1833] "ASGR1"          "BASP1"          "BCL2A1"         "BIN1"          
#> [1837] "BLVRA"          "C12orf75"       "C1QA"           "C1QB"          
#> [1841] "C1orf228"       "C1orf54"        "C5AR1"          "CACNA2D3"      
#> [1845] "CAPN2"          "CCDC47"         "CD14"           "CD1C"          
#> [1849] "CD1D"           "CD1E"           "CD2"            "CD52"          
#> [1853] "CD68"           "CD79B"          "CDA"            "CDKN1C"        
#> [1857] "CEBPD"          "CHCHD10"        "CHURC1"         "CIITA"         
#> [1861] "CKB"            "CLEC10A"        "CLEC4A"         "CLEC9A"        
#> [1865] "COL9A2"         "CREM"           "CSF2RA"         "CSRNP1"        
#> [1869] "CST3"           "CST7"           "CTD-2319I12.1"  "CTSD"          
#> [1873] "CTSH"           "CTSW"           "CYBB"           "DDIT4"         
#> [1877] "DEDD2"          "DMXL2"          "DUSP2"          "DUSP5"         
#> [1881] "EML4"           "ENHO"           "ETV3"           "FAM105A"       
#> [1885] "FAM129A"        "FCER1A"         "FCGR2B"         "FCGR3A"        
#> [1889] "FLT3"           "FOSB"           "GIMAP2"         "GIMAP4"        
#> [1893] "GNLY"           "GPBAR1"         "GPR183"         "GPR65"         
#> [1897] "GRINA"          "GSN"            "GSTM4"          "HAVCR2"        
#> [1901] "HBEGF"          "HES4"           "HGSNAT"         "HIST1H4C"      
#> [1905] "HLA-DMB"        "HLA-DQA1"       "HMGA1"          "HMGN1"         
#> [1909] "HMOX1"          "ID1"            "ID2"            "ID3"           
#> [1913] "IFITM3"         "IFNGR2"         "IGFBP7"         "IGHM"          
#> [1917] "IKZF1"          "IL13RA1"        "IL1R2"          "IL8"           
#> [1921] "IRS2"           "ITGB7"          "KCNK6"          "KIAA0125"      
#> [1925] "KLF4"           "KLF6"           "KLF9"           "KLRB1"         
#> [1929] "KLRF1"          "LAMP1"          "LGMN"           "LILRA3"        
#> [1933] "LILRA5"         "LILRB2"         "LIMA1"          "LMNA"          
#> [1937] "LTB"            "LY86"           "LYN"            "MAFB"          
#> [1941] "MAFF"           "MAGEF1"         "MAP3K8"         "MAT2A"         
#> [1945] "MFSD2B"         "MIF4GD"         "MNDA"           "MPST"          
#> [1949] "MS4A7"          "MSLN"           "MT2A"           "MTSS1"         
#> [1953] "MYO9B"          "MZB1"           "NDRG2"          "NEAT1"         
#> [1957] "NINJ1"          "NKG7"           "NPC2"           "NPL"           
#> [1961] "NR4A2"          "NR4A3"          "NUCB2"          "NUDT16"        
#> [1965] "NUMB"           "PAK1"           "PDLIM1"         "PEBP1"         
#> [1969] "PHACTR1"        "PHLDA2"         "PIK3AP1"        "PILRA"         
#> [1973] "PKIB"           "PLBD1"          "PLD4"           "PON2"          
#> [1977] "PPM1N"          "PPP1R14A"       "PPP1R16A"       "PRCP"          
#> [1981] "RARRES3"        "RBP7"           "RGS1"           "RNASE6"        
#> [1985] "RNU1-60P"       "RNU7-18P"       "RP11-290F20.3"  "RP13-516M14.4" 
#> [1989] "RP6-91H8.3"     "RRAS"           "RXRA"           "S100B"         
#> [1993] "SCARB2"         "SERPINA1"       "SERPINF1"       "SIGLEC10"      
#> [1997] "SIVA1"          "SLC11A1"        "SLC7A7"         "SOD2"          
#> [2001] "SPINK2"         "SRSF7"          "STXBP2"         "SULT1A1"       
#> [2005] "SUSD3"          "TCF7L2"         "TCTN3"          "TESC"          
#> [2009] "TIPARP"         "TMEM109"        "TMEM141"        "TMEM176B"      
#> [2013] "TNFSF10"        "TOB1"           "UPK3A"          "UTRN"          
#> [2017] "VIMP"           "VMP1"           "ZFP36"          "ZNF703"