takes files with positive and negative markers, as described in garnett, and returns list of markers

takes files with positive and negative markers, as described in garnett, and returns list of markers

Usage

file_marker_parse(filename)

Arguments

filename

txt file to load

Value

list of positive and negative gene markers

Examples

marker_file <- system.file(
    "extdata",
    "hsPBMC_markers.txt",
    package = "clustifyr"
)

file_marker_parse(marker_file)
#> $pos
#> $pos$`CD34+`
#> [1] "CD34"   "THY1"   "ENG"    "KIT"    "PROM1 "
#> 
#> $pos$`NK cells`
#> [1] "NCAM1"  "FCGR3A"
#> 
#> $pos$Monocytes
#> [1] "CD14"    "FCGR1A"  "CD68"    "S100A12"
#> 
#> $pos$`B cells`
#> [1] "CD19"  "MS4A1" "CD79A"
#> 
#> $pos$`T cells`
#> [1] "CD3D" "CD3E" "CD3G"
#> 
#> $pos$`CD4 T cells`
#> [1] "CD4"   "FOXP3" "IL2RA" "IL7R" 
#> 
#> $pos$`CD8 T cells`
#> [1] "CD8A" "CD8B"
#> 
#> $pos$`Dendritic cells`
#> [1] "\tIL3RA" "CD1C"    "BATF3"   "THBD"    "CD209 " 
#> 
#> 
#> $neg
#> NULL
#>