Generate variable gene list from marker matrix

Variable gene list is required for clustify main function. This function parses variables genes from a matrix input.

get_vargenes(marker_mat)

Arguments

marker_mat
matrix or dataframe of candidate genes for each cluster

Value

vector of marker gene names

Examples

get_vargenes(cbmc_m)
#>  [1] "ITM2A"      "TXNIP"      "AES"        "CD8B"       "CD8A"      
#>  [6] "S100B"      "ADCY2"      "PTGDR2"     "CD200R1"    "S100A8"    
#> [11] "S100A9"     "LYZ"        "CDHR3"      "DICER1-AS1" "RAD9A"     
#> [16] "GNLY"       "NKG7"       "CST7"       "IGHM"       "CD79A"     
#> [21] "MS4A1"      "CDKN1C"     "HES4"       "CKB"        "SPINK2"    
#> [26] "C1QTNF4"    "KIAA0125"   "HBM"        "AHSP"       "CA1"       
#> [31] "PF4"        "SDPR"       "TUBB1"      "ENHO"       "CD1E"      
#> [36] "NDRG2"      "LILRA4"     "TPM2"       "SCT"