convert gmt format of pathways to list of vectors

convert gmt format of pathways to list of vectors

gmt_to_list(
  path,
  cutoff = 0,
  sep = "\thttp://www.broadinstitute.org/gsea/msigdb/cards/.*?\t"
)

Arguments

path
gmt file path

cutoff
remove pathways with less genes than this cutoff

sep
sep used in file to split path and genes

Value

list of genes in each pathway

Examples

gmt_file <- system.file(
  "extdata",
  "c2.cp.reactome.v6.2.symbols.gmt",
  package = "clustifyr"
)

gmt_list <- gmt_to_list(path = gmt_file)

#> Parsed with column specification:
#> cols(
#>   X1 = col_character()
#> )