decide for one gene whether it is a marker for a certain cell type

decide for one gene whether it is a marker for a certain cell type

Usage

marker_select(row1, cols, cut = 1, compto = 1)

Arguments

row1

a numeric vector of expression values (row)

cols

a vector of cell types (column)

cut

an expression minimum cutoff

compto

compare max expression to the value of next 1 or more

Value

vector of cluster name and ratio value

Examples

pbmc_avg <- average_clusters(
    mat = pbmc_matrix_small,
    metadata = pbmc_meta,
    cluster_col = "classified",
    if_log = FALSE
)

marker_select(
    row1 = pbmc_avg["PPBP", ],
    cols = names(pbmc_avg["PPBP", ])
)
#> [1] "Platelet"          "0.050236218204724"