Convert candidate genes list into matrix

Convert candidate genes list into matrix

matrixize_markers(
  marker_df,
  ranked = FALSE,
  n = NULL,
  step_weight = 1,
  background_weight = 0,
  unique = FALSE,
  metadata = NULL,
  cluster_col = "classified",
  remove_rp = FALSE
)

Arguments

marker_df
dataframe of candidate genes, must contain "gene" and "cluster" columns, or a matrix of gene names to convert to ranked

ranked
unranked gene list feeds into hyperp, the ranked gene list feeds into regular corr_coef

n
number of genes to use

step_weight
ranked genes are tranformed into pseudo expression by descending weight

background_weight
ranked genes are tranformed into pseudo expression with added weight

unique
whether to use only unique markers to 1 cluster

metadata
vector or dataframe of cluster names, should have column named cluster

cluster_col
column for cluster names to replace original cluster, if metadata is dataframe

remove_rp
do not include rps, rpl, rp1-9 in markers

Value

matrix of unranked gene marker names, or matrix of ranked expression

Examples

pbmc_mm <- matrixize_markers(pbmc_markers)