Plot a tSNE colored by feature.

Plot a tSNE colored by feature.

plot_tsne(data, x = "tSNE_1", y = "tSNE_2", feature,
  legend_name = "", c_cols = pretty_palette, d_cols = NULL,
  pt_size = 0.25, scale_limits = NULL, do.label = FALSE,
  do.legend = TRUE)

Arguments

data
input data

x
x variable

y
y variable

feature
feature to color by

legend_name
legend name to display, defaults to no name

c_cols
character vector of colors to built color gradient for continuous values. defaults to clustifyR::pretty_palette

d_cols
character vector of colors for discrete values. defaults to RColorBrewer paired palette

pt_size
point size

scale_limits
defaults to min = 0, max = max(data$x), otherwise a two-element numeric vector indicating min and max to plot

do.label
whether to label each cluster at median center

do.legend
whether to draw legend