refine_clusters: refine clustering using reference data

refine_clusters: refine clustering using reference data

refine_clusters(expr_mat, sc_cluster, bulk_mat, lambda = 0.5, epsilon = 1,
  if_compute_sigma = TRUE, num_iteration = 100, default_similiarity = -5,
  disagreement_freq = 0.001, if_plot = FALSE, tsne_coord = NULL,
  cluster_names = NULL, compute_method, ...)

Arguments

expr_mat
input single cell data after gene subsetting

sc_cluster
initial clustering assignment. Note that the number of bulk cell types must be no more than the number of clusters. for additional clusters, there is an assignment of default_similiarity

bulk_mat
input bulk data after gene subsetting

lambda
control parameters

epsilon
control parameters

if_compute_sigma
whether compute the standard deviation of each gene's expression within each cluster. If not, the function assumes that all standard deviations equal to 1.

num_iteration
maximal number of iteration before exiting the algorithm

default_similiarity
default similarity score if there is no reference data to compare to

disagreement_freq
maximal fraction of changes in cluster assignment after one iteration to be considered as converged.

if_plot
whether to plot the clustering result at each iteration.

tsne_coord
t-SNE coordinate of every cells for plotting

compute_method
corr_coef

...
control parameters for calculating similarity score

cluster_name
Name of each cluster for plotting