generate negative markers from a list of exclusive positive markers

generate negative markers from a list of exclusive positive markers

Usage

reverse_marker_matrix(mat)

Arguments

mat

matrix or dataframe of markers

Value

matrix of gene names

Examples

reverse_marker_matrix(cbmc_m)
#>         CD4 T      CD8 T Memory CD4 T CD14+ Mono Naive CD4 T         NK
#> 1        CD8B      ITM2A        ITM2A      ITM2A       ITM2A      ITM2A
#> 2       ADCY2      ADCY2         CD8B       CD8B        CD8B       CD8B
#> 3      S100A8     S100A8       S100A8      ADCY2       ADCY2      ADCY2
#> 4       CDHR3      CDHR3        CDHR3      CDHR3      S100A8     S100A8
#> 5        GNLY       GNLY         GNLY       GNLY        GNLY      CDHR3
#> 6        IGHM       IGHM         IGHM       IGHM        IGHM       IGHM
#> 7      CDKN1C     CDKN1C       CDKN1C     CDKN1C      CDKN1C     CDKN1C
#> 8      SPINK2     SPINK2       SPINK2     SPINK2      SPINK2     SPINK2
#> 9         HBM        HBM          HBM        HBM         HBM        HBM
#> 10        PF4        PF4          PF4        PF4         PF4        PF4
#> 11       ENHO       ENHO         ENHO       ENHO        ENHO       ENHO
#> 12     LILRA4     LILRA4       LILRA4     LILRA4      LILRA4     LILRA4
#> 13       CD8A      TXNIP        TXNIP      TXNIP       TXNIP      TXNIP
#> 14     PTGDR2     PTGDR2         CD8A       CD8A        CD8A       CD8A
#> 15     S100A9     S100A9       S100A9     PTGDR2      PTGDR2     PTGDR2
#> 16 DICER1-AS1 DICER1-AS1   DICER1-AS1 DICER1-AS1      S100A9     S100A9
#> 17       NKG7       NKG7         NKG7       NKG7        NKG7 DICER1-AS1
#> 18      CD79A      CD79A        CD79A      CD79A       CD79A      CD79A
#> 19       HES4       HES4         HES4       HES4        HES4       HES4
#> 20    C1QTNF4    C1QTNF4      C1QTNF4    C1QTNF4     C1QTNF4    C1QTNF4
#> 21       AHSP       AHSP         AHSP       AHSP        AHSP       AHSP
#> 22       SDPR       SDPR         SDPR       SDPR        SDPR       SDPR
#> 23       CD1E       CD1E         CD1E       CD1E        CD1E       CD1E
#> 24       TPM2       TPM2         TPM2       TPM2        TPM2       TPM2
#> 25      S100B        AES          AES        AES         AES        AES
#> 26    CD200R1    CD200R1        S100B      S100B       S100B      S100B
#> 27        LYZ        LYZ          LYZ    CD200R1     CD200R1    CD200R1
#> 28      RAD9A      RAD9A        RAD9A      RAD9A         LYZ        LYZ
#> 29       CST7       CST7         CST7       CST7        CST7      RAD9A
#> 30      MS4A1      MS4A1        MS4A1      MS4A1       MS4A1      MS4A1
#> 31        CKB        CKB          CKB        CKB         CKB        CKB
#> 32   KIAA0125   KIAA0125     KIAA0125   KIAA0125    KIAA0125   KIAA0125
#> 33        CA1        CA1          CA1        CA1         CA1        CA1
#> 34      TUBB1      TUBB1        TUBB1      TUBB1       TUBB1      TUBB1
#> 35      NDRG2      NDRG2        NDRG2      NDRG2       NDRG2      NDRG2
#> 36        SCT        SCT          SCT        SCT         SCT        SCT
#>             B CD16+ Mono      CD34+      Eryth         Mk         DC       pDCs
#> 1       ITM2A      ITM2A      ITM2A      ITM2A      ITM2A      ITM2A      ITM2A
#> 2        CD8B       CD8B       CD8B       CD8B       CD8B       CD8B       CD8B
#> 3       ADCY2      ADCY2      ADCY2      ADCY2      ADCY2      ADCY2      ADCY2
#> 4      S100A8     S100A8     S100A8     S100A8     S100A8     S100A8     S100A8
#> 5       CDHR3      CDHR3      CDHR3      CDHR3      CDHR3      CDHR3      CDHR3
#> 6        GNLY       GNLY       GNLY       GNLY       GNLY       GNLY       GNLY
#> 7      CDKN1C       IGHM       IGHM       IGHM       IGHM       IGHM       IGHM
#> 8      SPINK2     SPINK2     CDKN1C     CDKN1C     CDKN1C     CDKN1C     CDKN1C
#> 9         HBM        HBM        HBM     SPINK2     SPINK2     SPINK2     SPINK2
#> 10        PF4        PF4        PF4        PF4        HBM        HBM        HBM
#> 11       ENHO       ENHO       ENHO       ENHO       ENHO        PF4        PF4
#> 12     LILRA4     LILRA4     LILRA4     LILRA4     LILRA4     LILRA4       ENHO
#> 13      TXNIP      TXNIP      TXNIP      TXNIP      TXNIP      TXNIP      TXNIP
#> 14       CD8A       CD8A       CD8A       CD8A       CD8A       CD8A       CD8A
#> 15     PTGDR2     PTGDR2     PTGDR2     PTGDR2     PTGDR2     PTGDR2     PTGDR2
#> 16     S100A9     S100A9     S100A9     S100A9     S100A9     S100A9     S100A9
#> 17 DICER1-AS1 DICER1-AS1 DICER1-AS1 DICER1-AS1 DICER1-AS1 DICER1-AS1 DICER1-AS1
#> 18       NKG7       NKG7       NKG7       NKG7       NKG7       NKG7       NKG7
#> 19       HES4      CD79A      CD79A      CD79A      CD79A      CD79A      CD79A
#> 20    C1QTNF4    C1QTNF4       HES4       HES4       HES4       HES4       HES4
#> 21       AHSP       AHSP       AHSP    C1QTNF4    C1QTNF4    C1QTNF4    C1QTNF4
#> 22       SDPR       SDPR       SDPR       SDPR       AHSP       AHSP       AHSP
#> 23       CD1E       CD1E       CD1E       CD1E       CD1E       SDPR       SDPR
#> 24       TPM2       TPM2       TPM2       TPM2       TPM2       TPM2       CD1E
#> 25        AES        AES        AES        AES        AES        AES        AES
#> 26      S100B      S100B      S100B      S100B      S100B      S100B      S100B
#> 27    CD200R1    CD200R1    CD200R1    CD200R1    CD200R1    CD200R1    CD200R1
#> 28        LYZ        LYZ        LYZ        LYZ        LYZ        LYZ        LYZ
#> 29      RAD9A      RAD9A      RAD9A      RAD9A      RAD9A      RAD9A      RAD9A
#> 30       CST7       CST7       CST7       CST7       CST7       CST7       CST7
#> 31        CKB      MS4A1      MS4A1      MS4A1      MS4A1      MS4A1      MS4A1
#> 32   KIAA0125   KIAA0125        CKB        CKB        CKB        CKB        CKB
#> 33        CA1        CA1        CA1   KIAA0125   KIAA0125   KIAA0125   KIAA0125
#> 34      TUBB1      TUBB1      TUBB1      TUBB1        CA1        CA1        CA1
#> 35      NDRG2      NDRG2      NDRG2      NDRG2      NDRG2      TUBB1      TUBB1
#> 36        SCT        SCT        SCT        SCT        SCT        SCT      NDRG2