Function to convert labelled seurat object to avg expression matrix

Function to convert labelled seurat object to avg expression matrix

use_seurat_comp(seurat_object, cluster_col = "classified",
  var.genes_only = FALSE, assay_name = NULL, method = "mean")

Arguments

seurat_object
seurat_object after tsne projections and clustering

cluster_col
column name where classified cluster names are stored in seurat meta data, cannot be "rn"

var.genes_only
whether to keep only var.genes in the final matrix output, could also look up genes used for PCA

assay_name
any additional assay data, such as ADT, to include. If more than 1, pass a vector of names

method
whether to take mean (default) or median

seurat3
if using newest version