Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)

Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)

calculate_pathway_gsea(
  mat,
  pathway_list,
  n_perm = 1000,
  scale = TRUE,
  no_warnings = TRUE
)

Arguments

mat
expression matrix

pathway_list
a list of vectors, each named for a specific pathway, or dataframe

n_perm
Number of permutation for fgsea function. Defaults to 1000.

scale
convert expr_mat into zscores prior to running GSEA?, default = FALSE

no_warnings
suppress warnings from gsea ties

Value

matrix of GSEA NES values, cell types as row names, pathways as column names

Examples

gl <- list(
    "n" = c("PPBP", "LYZ", "S100A9"),
    "a" = c("IGLL5", "GNLY", "FTL")
)

pbmc_avg <- average_clusters(
    mat = pbmc_matrix_small,
    metadata = pbmc_meta,
    cluster_col = "classified"
)

calculate_pathway_gsea(
    mat = pbmc_avg,
    pathway_list = gl
)

#>                       n          a
#> Naive CD4 T  -1.1774274 -1.1565701
#> Memory CD4 T -1.2699026 -1.3115100
#> CD14+ Mono    1.4836403  0.9923194
#> B            -1.1131379  1.2192091
#> CD8 T        -1.3906078 -0.5494581
#> FCGR3A+ Mono  0.9893043  1.0798741
#> NK           -1.3801696  1.0742091
#> DC            1.1088866 -0.8547561
#> Platelet      1.8248440  1.2914442