Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)

Convert expression matrix to GSEA pathway scores (would take a similar place in workflow before average_clusters/binarize)

Usage

calculate_pathway_gsea(
  mat,
  pathway_list,
  n_perm = 1000,
  scale = TRUE,
  no_warnings = TRUE
)

Arguments

mat

expression matrix

pathway_list

a list of vectors, each named for a specific pathway, or dataframe

n_perm

Number of permutation for fgsea function. Defaults to 1000.

scale

convert expr_mat into zscores prior to running GSEA?, default = FALSE

no_warnings

suppress warnings from gsea ties

Value

matrix of GSEA NES values, cell types as row names, pathways as column names

Examples

gl <- list(
    "n" = c("PPBP", "LYZ", "S100A9"),
    "a" = c("IGLL5", "GNLY", "FTL")
)

pbmc_avg <- average_clusters(
    mat = pbmc_matrix_small,
    metadata = pbmc_meta,
    cluster_col = "classified"
)

calculate_pathway_gsea(
    mat = pbmc_avg,
    pathway_list = gl
)
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>   |                                                                        |                                                                |   0%  |                                                                        |================================================================| 100%
#> 
#>                       n          a
#> Naive CD4 T  -1.1309372 -1.1489896
#> Memory CD4 T -1.2839630 -1.3221002
#> CD14+ Mono    1.4593450  0.9853898
#> B            -1.1100514  1.1981402
#> CD8 T        -1.3948764 -0.5489829
#> FCGR3A+ Mono  0.9957008  1.0827227
#> NK           -1.4090702  1.0794074
#> DC            1.0849890 -0.8633225
#> Platelet      1.8335770  1.3112985