Compare scRNA-seq data to reference data.

Compare scRNA-seq data to reference data.

clustify(input, ...)

# S3 method for default
clustify(
  input,
  ref_mat,
  metadata = NULL,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  verbose = FALSE,
  lookuptable = NULL,
  rm0 = FALSE,
  obj_out = FALSE,
  seurat_out = FALSE,
  rename_prefix = NULL,
  threshold = "auto",
  low_threshold_cell = 10,
  exclude_genes = rownames(ref_mat)[stringr::str_detect(rownames(ref_mat),
    "RP[0-9,L,S]|Rp[0-9,l,s]")],
  ...
)

# S3 method for seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  ...
)

# S3 method for Seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  ...
)

# S3 method for SingleCellExperiment
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  ...
)

Arguments

input
single-cell expression matrix or Seurat object

...
additional arguments to pass to compute_method function

ref_mat
reference expression matrix

metadata
cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set. Not required if running correlation per cell.

cluster_col
column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes
A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

per_cell
if true run per cell, otherwise per cluster.

n_perm
number of permutations, set to 0 by default

compute_method
method(s) for computing similarity scores

verbose
whether to report certain variables chosen

lookuptable
if not supplied, will look in built-in table for object parsing

rm0
consider 0 as missing data, recommended for per_cell

obj_out
whether to output object instead of cor matrix

seurat_out
output cor matrix or called seurat object

rename_prefix
prefix to add to type and r column names

threshold
identity calling minimum correlation score threshold, only used when obj_out = TRUE

low_threshold_cell
option to remove clusters with too few cells

exclude_genes
a vector of gene names to throw out of query

use_var_genes
if providing a seurat object, use the variable genes (stored in seurat_object@var.genes) as the query_genes.

dr
stored dimension reduction

Value

single cell object with identity assigned in metadata, or matrix of correlation values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Annotate a matrix and metadata
clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = cbmc_ref,
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  verbose = TRUE
)

#> using # of genes: 1881

#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T     CD8 T
#> B            0.8912913  0.6851783  0.7317660 0.7317217 0.7814900 0.7659276
#> CD14+ Mono   0.6989945  0.9081472  0.8847734 0.6710434 0.6670570 0.6521402
#> CD8 T        0.7525971  0.6651899  0.7126303 0.7136651 0.8400829 0.8368499
#> DC           0.7026957  0.7823684  0.7882898 0.6814873 0.6658926 0.6524832
#> FCGR3A+ Mono 0.7039327  0.8496499  0.9059093 0.6726128 0.6757769 0.6596604
#> Memory CD4 T 0.7758978  0.6820664  0.7241914 0.7427243 0.8948977 0.8702287
#> Naive CD4 T  0.7871487  0.6792475  0.7175664 0.7548765 0.9158677 0.8936029
#> NK           0.7192197  0.6526278  0.6958540 0.7067672 0.7940021 0.7937825
#> Platelet     0.3164496  0.3823818  0.3918461 0.3271915 0.2858296 0.2797719
#>                     DC     Eryth Memory CD4 T        Mk Naive CD4 T        NK
#> B            0.7253935 0.6300171    0.5959394 0.5450655   0.5953671 0.7474430
#> CD14+ Mono   0.8121193 0.6055132    0.5708809 0.6549038   0.5384300 0.7573710
#> CD8 T        0.6840079 0.6525508    0.6542797 0.5673165   0.6121911 0.8662208
#> DC           0.8327092 0.6049950    0.5657932 0.5748673   0.5260392 0.7190055
#> FCGR3A+ Mono 0.7784995 0.6097144    0.5710313 0.6356263   0.5300822 0.7552514
#> Memory CD4 T 0.6965506 0.6643080    0.6860640 0.5514448   0.6554955 0.8182608
#> Naive CD4 T  0.6958788 0.6677468    0.6913118 0.5530940   0.6622109 0.8080934
#> NK           0.6741498 0.6420863    0.6259194 0.5642024   0.5880876 0.8960291
#> Platelet     0.3476335 0.3599792    0.2945712 0.6304742   0.2899873 0.3402365
#>                   pDCs
#> B            0.6889127
#> CD14+ Mono   0.6405745
#> CD8 T        0.6317819
#> DC           0.7012226
#> FCGR3A+ Mono 0.6374282
#> Memory CD4 T 0.6434303
#> Naive CD4 T  0.6452324
#> NK           0.6263499
#> Platelet     0.3349940

# Annotate using a different method
clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = cbmc_ref,
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  compute_method = "cosine"
)

#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T     CD8 T
#> B            0.9043453  0.6692736  0.7304611 0.7126145 0.6886051 0.6911548
#> CD14+ Mono   0.7163914  0.9237414  0.9034916 0.6848786 0.6584317 0.6560087
#> CD8 T        0.7304785  0.6900086  0.7260586 0.7189075 0.8108622 0.8232917
#> DC           0.8138921  0.8702626  0.8990999 0.7737429 0.7120145 0.7137373
#> FCGR3A+ Mono 0.7395710  0.8895103  0.9478961 0.7209188 0.6963324 0.6950678
#> Memory CD4 T 0.7822091  0.7381052  0.7693954 0.7691244 0.8911277 0.8828845
#> Naive CD4 T  0.7803030  0.7156062  0.7458398 0.7608143 0.8904659 0.8846681
#> NK           0.6640022  0.6547077  0.6966375 0.6789784 0.7356862 0.7528583
#> Platelet     0.4850807  0.5178941  0.5281099 0.4847617 0.4812028 0.4839748
#>                     DC     Eryth Memory CD4 T        Mk Naive CD4 T        NK
#> B            0.7608264 0.6030349    0.6327392 0.6147951   0.6747284 0.6350225
#> CD14+ Mono   0.8501651 0.6562285    0.6967433 0.7612545   0.6907615 0.7202606
#> CD8 T        0.7091008 0.6722061    0.7688782 0.6968347   0.7590352 0.8458730
#> DC           0.9375263 0.6849344    0.7203505 0.7385450   0.7164560 0.7428729
#> FCGR3A+ Mono 0.8543528 0.6627492    0.7150458 0.7540976   0.7091690 0.7490253
#> Memory CD4 T 0.7447723 0.7072160    0.8328531 0.6970886   0.8263857 0.7795909
#> Naive CD4 T  0.7222539 0.7003026    0.8159739 0.6813145   0.8230363 0.7593719
#> NK           0.6693547 0.6337988    0.7039878 0.6815097   0.7039110 0.9043371
#> Platelet     0.4995469 0.5025886    0.4786084 0.8040026   0.4623411 0.4846219
#>                   pDCs
#> B            0.7253193
#> CD14+ Mono   0.7118965
#> CD8 T        0.6746757
#> DC           0.8116601
#> FCGR3A+ Mono 0.7252533
#> Memory CD4 T 0.7131557
#> Naive CD4 T  0.6998970
#> NK           0.6617692
#> Platelet     0.5034934

# Annotate a Seurat object
clustify(
  s_small,
  cbmc_ref,
  cluster_col = "res.1",
  seurat_out = TRUE,
  per_cell = FALSE,
  dr = "tsne"
)

#> using threshold of 0.61

#> seurat not loaded, returning cor_mat instead

#>            B CD14+ Mono CD16+ Mono      CD34+       CD4 T      CD8 T
#> 0  0.5224110  0.3852326  0.4119037 0.41720039  0.61062598  0.5825024
#> 1  0.4136741  0.7251465  0.6939952 0.53311242  0.42401600  0.3648038
#> 2  0.2311589  0.6184823  0.6094641 0.42129146  0.46805261  0.4812869
#> 3 -0.1762557  0.1362385  0.1859978 0.01798528 -0.08048411 -0.1647901
#>            DC     Eryth Memory CD4 T          Mk Naive CD4 T        NK
#> 0  0.36476066 0.1523530    0.7047302 -0.10828917   0.6398803 0.4346006
#> 1  0.71146653 0.4359801    0.5349202  0.08053672   0.4304222 0.6174695
#> 2  0.67077257 0.4137920    0.6331816  0.18844401   0.5864320 0.8099383
#> 3 -0.05333518 0.4352437    0.1083785  0.74534070  -0.1097800 0.1124080
#>        pDCs
#> 0 0.3456669
#> 1 0.5116362
#> 2 0.2859966
#> 3 0.1901284

# Annotate (and return) a Seurat object per-cell
clustify(
  input = s_small,
  ref_mat = cbmc_ref,
  cluster_col = "res.1",
  seurat_out = TRUE,
  per_cell = TRUE,
  dr = "tsne"
)

#> using threshold of NA

#> seurat not loaded, returning cor_mat instead

#>                           B   CD14+ Mono   CD16+ Mono        CD34+       CD4 T
#> ATGCCAGAACGACT  0.136109346  0.187465419  0.223788486  0.224013298  0.17139695
#> CATGGCCTGTGCAT  0.152939576  0.296785448  0.244802052  0.206675103  0.34996984
#> GAACCTGATGAACC  0.236952133  0.382726194  0.469936740  0.304019287  0.48637543
#> TGACTGGATTCTCA  0.334623228  0.489606618  0.466899648  0.442054896  0.44029372
#> AGTCAGACTGCACA  0.180905688  0.297314565  0.267271409  0.261657792  0.37623139
#> TCTGATACACGTGT  0.324270809  0.203995818  0.324401642  0.237012482  0.50114580
#> TGGTATCTAAACAG  0.231687026  0.217627636  0.321152254  0.216456020  0.59488793
#> GCAGCTCTGTTTCT  0.235985880  0.093890243  0.111894243  0.083493002  0.63076600
#> GATATAACACGCAT  0.273082649  0.259238155  0.388840998  0.280697121  0.52159132
#> AATGTTGACAGTCA  0.300425525  0.199360717  0.257265238  0.188977346  0.47002058
#> AGGTCATGAGTGTC  0.442331591  0.006576135  0.161353433  0.108679280  0.04187960
#> AGAGATGATCTCGC  0.403107368  0.062601580  0.134033138  0.290775562  0.08015342
#> GGGTAACTCTAGTG  0.305617210 -0.102795370  0.041550240 -0.024573977  0.04984018
#> CATGAGACACGGGA  0.488773840  0.078581003  0.162335243  0.123372175  0.17287821
#> TACGCCACTCCGAA  0.411772330  0.162500191  0.260420870  0.462257826  0.08046126
#> CTAAACCTGTGCAT  0.414596949  0.137182814  0.260059388  0.143556965  0.10697488
#> GTAAGCACTCATTC  0.300687477  0.203995818  0.206437409  0.119095824  0.14739582
#> TTGGTACTGAATCC  0.545680062 -0.130635162 -0.006739509  0.095779592 -0.09372926
#> CATCATACGGAGCA  0.418604135  0.180412486  0.206437409  0.095512493  0.11202082
#> TACATCACGCTAAC  0.377766448 -0.120986085 -0.034986651  0.121931289  0.25741050
#> TTACCATGAATCGC  0.109788758  0.439155033  0.353906509  0.207378766  0.25617377
#> ATAGGAGAAACAGA  0.310485803  0.303859581  0.350384444  0.335097482  0.29534015
#> GCGCACGACTTTAC  0.007572824  0.255582825  0.183715087  0.169441947  0.06436901
#> ACTCGCACGAAAGT  0.304224977  0.465770798  0.464779080  0.393841637  0.32780831
#> ATTACCTGCCTTAT  0.215058310  0.503003651  0.501265856  0.301525054  0.36471229
#> CCCAACTGCAATCG -0.048795004  0.402558780  0.280684473  0.158583762  0.12198751
#> AAATTCGAATCACG  0.021224998  0.453979132  0.335018423  0.158008321  0.15093332
#> CCATCCGATTCGCC  0.142886902  0.285773803  0.265468454  0.122474487  0.12247449
#> TCCACTCTGAGCTT  0.183009924  0.468568513  0.438769709  0.334466414  0.23980611
#> CATCAGGATGCACA  0.315415090  0.443863350  0.507853412  0.461418921  0.50443007
#> CTAAACCTCTGACA  0.501617966  0.613581313  0.631018301  0.458661731  0.53681436
#> GATAGAGAAGGGTG -0.158583762  0.024397502  0.000000000 -0.134186260 -0.10978876
#> CTAACGGAACCGAT  0.180135930  0.369351883  0.378584591  0.217041828  0.28763168
#> AGATATACCCGTAA  0.329561055  0.327985715  0.358061578  0.196917456  0.35634183
#> TACTCTGAATCGAC  0.105218156  0.281389346  0.278732862  0.172017841  0.30215609
#> GCGCATCTTGCTCC  0.091546869  0.318253063  0.194173389  0.144589046  0.15166134
#> GTTGACGATATCGG  0.075569219  0.332153081  0.322617364  0.194781128  0.22231410
#> ACAGGTACTGGTGT  0.079605842  0.201783504  0.203942429  0.096219235  0.24400921
#> GGCATATGCTTATC  0.079252870  0.311727955  0.243140194  0.088058744  0.14265517
#> CATTACACCAACTG -0.006818326  0.366241488  0.337643290  0.214777255  0.24789484
#> TAGGGACTGAACTC  0.236001666  0.477253955  0.354974709  0.354555196  0.39683653
#> GCTCCATGAGAAGT  0.232897126  0.414039336  0.371059889  0.345032780  0.43129097
#> TACAATGATGCTAG  0.172595139  0.283549156  0.202616828  0.209579811  0.32933970
#> CTTCATGACCGAAT  0.368944546  0.393204777  0.370354185  0.265287197  0.38249247
#> CTGCCAACAGGAGC -0.030253686  0.073473238  0.055815023 -0.002750335  0.04125503
#> TTGCATTGAGCTAC  0.411054154  0.278929604  0.367160716  0.293610110  0.39637365
#> AAGCAAGAGCTTAG  0.261893512  0.369175192  0.325131511  0.235073093  0.29186928
#> CGGCACGAACTCAG  0.303980009  0.313785816  0.323722182  0.235339362  0.48048453
#> GGTGGAGATTACTC  0.332844400  0.620277064  0.596630591  0.514491344  0.47170456
#> GGCCGATGTACTCT  0.166670935  0.210150310  0.159173501  0.209520174  0.47638271
#> CGTAGCCTGTATGC  0.190898283  0.457524811  0.446661250  0.265048856  0.18932061
#> TGAGCTGAATGCTG  0.253113368  0.554202473  0.553322741  0.502642342  0.39483480
#> CCTATAACGAGACG  0.142886902  0.285773803  0.265468454  0.122474487  0.12247449
#> ATAAGTTGGTACGT           NA           NA           NA           NA          NA
#> AAGCGACTTTGACG  0.167657419  0.258419329  0.388100663  0.288673300  0.30978805
#> ACCAGTGAATACCG  0.404663096  0.563694118  0.615903148  0.464181637  0.45536408
#> ATTGCACTTGCTTT  0.268853044  0.712460567  0.626435827  0.463893707  0.39985780
#> CTAGGTGATGGTTG  0.354451422  0.487095020  0.493595449  0.372725362  0.37083495
#> GCACTAGACCTTTA  0.364442177  0.484345231  0.497167116  0.462257826  0.41177233
#> CATGCGCTAGTCAC  0.314445066  0.547497461  0.547718360  0.465519232  0.42980015
#> TTGAGGACTACGCA  0.126940701  0.451878552  0.407031865  0.394729950  0.22935302
#> ATACCACTCTAAGC  0.251162481  0.254699981  0.283114160  0.211070817  0.28535831
#> CATATAGACTAAGC  0.156892908  0.284368396  0.215814788  0.039223227  0.19611614
#> TTTAGCTGTACTCT  0.334026073  0.531242128  0.559050262  0.388639750  0.40436187
#> GACATTCTCCACCT  0.333261479  0.530790633  0.618493626  0.526652453  0.33657202
#> ACGTGATGCCATGA  0.123852191  0.513056913  0.455253743  0.340357166  0.21682012
#> ATTGTAGATTCCCG  0.367755778  0.565473938  0.660908829  0.529357421  0.35773806
#> GATAGAGATCACGA  0.124636069  0.265048856  0.299878578  0.137257443  0.09466031
#> AATGCGTGGACGGA  0.188113775  0.450149093  0.441500700  0.331267807  0.29182460
#> GCGTAAACACGGTT  0.383864558  0.582823790  0.563210127  0.476730571  0.46956583
#> ATTCAGCTCATTGG -0.138542016 -0.029812839  0.061988940 -0.233241621 -0.16250920
#> GGCATATGGGGAGT -0.149861695 -0.028020359  0.051968816 -0.168436989 -0.10798857
#> ATCATCTGACACCA -0.144048893  0.012339853  0.039553848 -0.047092907 -0.13322004
#> GTCATACTTCGCCT -0.112552701 -0.023679919  0.024859324 -0.052329697 -0.11898428
#> TTACGTACGTTCAG -0.094427331  0.063438795  0.136872511 -0.095596710 -0.06519286
#> GAGTTGTGGTAGCT -0.254601826 -0.035469338  0.010693702 -0.093289219 -0.20601369
#> GACGCTCTCTCTCG -0.026439683  0.062322110  0.075257545  0.063581142 -0.01416412
#> AGTCTTACTTCGGA -0.252293457 -0.106121118 -0.019302534 -0.287374818 -0.22422837
#> GGAACACTTCAGAC -0.151286175  0.032241316  0.110271195 -0.145223706 -0.12951640
#> CTTGATTGATCTTC -0.085115355  0.183771789  0.226697244  0.250371791  0.02348963
#>                      CD8 T          DC       Eryth Memory CD4 T           Mk
#> ATGCCAGAACGACT  0.15186274  0.15627377  0.15910930  0.255327083  0.268230988
#> CATGGCCTGTGCAT  0.35520561  0.17929926  0.08294561  0.312355912 -0.001654068
#> GAACCTGATGAACC  0.47806132  0.29017797  0.23030085  0.531237787  0.131970436
#> TGACTGGATTCTCA  0.43148785  0.48128616  0.42796550  0.503191455  0.269568476
#> AGTCAGACTGCACA  0.38278595  0.26336542  0.22233047  0.371957153 -0.017835580
#> TCTGATACACGTGT  0.47756246  0.23371070  0.05070416  0.471665348  0.044826409
#> TGGTATCTAAACAG  0.55095233  0.23309373  0.09900154  0.592805554 -0.032232434
#> GCAGCTCTGTTTCT  0.60209481  0.08452345 -0.12823265  0.458934064 -0.162010256
#> GATATAACACGCAT  0.48974898  0.28444513  0.10279537  0.522493387  0.001038756
#> AATGTTGACAGTCA  0.43887047  0.22935770  0.18932346  0.489410530  0.099374325
#> AGGTCATGAGTGTC  0.05537798  0.12905702 -0.03391901  0.104282920 -0.060594100
#> AGAGATGATCTCGC  0.09302478  0.01405565  0.19599560  0.315476455  0.208365927
#> GGGTAACTCTAGTG  0.05987744 -0.07033174 -0.30388665  0.062569752 -0.226795061
#> CATGAGACACGGGA  0.14930391  0.09891537  0.01650201  0.242069653 -0.009826589
#> TACGCCACTCCGAA  0.13725744  0.16898131  0.08046126  0.181943989  0.116794815
#> CTAAACCTGTGCAT  0.07177848  0.22470760 -0.09755222  0.101928551 -0.148882613
#> GTAAGCACTCATTC  0.18277082  0.17469284  0.06249583  0.153138100  0.021233562
#> TTGGTACTGAATCC -0.06824662 -0.12651566 -0.21323408 -0.058124158  0.008497642
#> CATCATACGGAGCA  0.08843749  0.08616606 -0.07900416  0.189891244 -0.014155708
#> TACATCACGCTAAC  0.29238304 -0.10707355 -0.19786266  0.084127000 -0.140261798
#> TTACCATGAATCGC  0.23177627  0.30527667  0.08539126  0.228131864  0.134240400
#> ATAGGAGAAACAGA  0.26599546  0.42545387  0.39568008  0.381590636  0.203601771
#> GCGCACGACTTTAC  0.04922336  0.21320072  0.22623813  0.181452248  0.022727640
#> ACTCGCACGAAAGT  0.30304581  0.40722318  0.33488331  0.480241081  0.246545248
#> ATTACCTGCCTTAT  0.31288767  0.39531893  0.31815327  0.507915152  0.325212859
#> CCCAACTGCAATCG  0.08539126  0.32969880  0.17078251  0.177435894  0.073222036
#> AAATTCGAATCACG  0.11437916  0.36000890  0.11084166  0.149462785  0.093191744
#> CCATCCGATTCGCC  0.10206207  0.24519627  0.08164966  0.169661166  0.081682601
#> TCCACTCTGAGCTT  0.20036431  0.41218805  0.23980611  0.299961711  0.071023874
#> CATCAGGATGCACA  0.45820040  0.48765893  0.39324479  0.696445081  0.209581005
#> CTAAACCTCTGACA  0.47473568  0.62446520  0.39547450  0.515689364  0.287507020
#> GATAGAGAAGGGTG -0.12198751  0.02442213 -0.07319251 -0.126739924 -0.073222036
#> CTAACGGAACCGAT  0.22319281  0.43921310  0.35089893  0.345701801  0.209217800
#> AGATATACCCGTAA  0.30624603  0.37215550  0.23945163  0.314248717  0.288403471
#> TACTCTGAATCGAC  0.39145307  0.43966001  0.09587312  0.135208199  0.034625200
#> GCGCATCTTGCTCC  0.16305558  0.28278391  0.28092709  0.233905415  0.128138714
#> GTTGACGATATCGG  0.15728942  0.32603803  0.23549478  0.318922084  0.202771314
#> ACAGGTACTGGTGT  0.31980782  0.35443041  0.07960584  0.231040607  0.012118820
#> GGCATATGCTTATC  0.12152107  0.26973180  0.15146104  0.234212750  0.142712723
#> CATTACACCAACTG  0.28393456  0.46801435  0.31607809  0.400243477  0.108162785
#> TAGGGACTGAACTC  0.35068632  0.53333792  0.42253694  0.420622890  0.160899751
#> GCTCCATGAGAAGT  0.43129097  0.37128471  0.23289713  0.600445022  0.000000000
#> TACAATGATGCTAG  0.34519028  0.22742093  0.11095402  0.440063800 -0.070475419
#> CTTCATGACCGAAT  0.38375274  0.38130170  0.26087625  0.420635002  0.189432225
#> CTGCCAACAGGAGC  0.08172424 -0.02792442 -0.04007631  0.150630189  0.056994213
#> TTGCATTGAGCTAC  0.38169314  0.32329719  0.23488809  0.427068089  0.190923572
#> AAGCAAGAGCTTAG  0.27924790  0.40902952  0.32973340  0.388475003  0.149939289
#> CGGCACGAACTCAG  0.48048453  0.31410261  0.16669871  0.458452056  0.098097631
#> GGTGGAGATTACTC  0.44256755  0.54417720  0.41054309  0.566988918  0.251572230
#> GGCCGATGTACTCT  0.49465665  0.16179303 -0.00756163  0.487085512 -0.098340856
#> CGTAGCCTGTATGC  0.17827691  0.39797467  0.07257290  0.180304854  0.101011732
#> TGAGCTGAATGCTG  0.38959607  0.53240591  0.51504986  0.702410424  0.209633741
#> CCTATAACGAGACG  0.10206207  0.24519627  0.08164966  0.169661166  0.081682601
#> ATAAGTTGGTACGT          NA          NA          NA           NA           NA
#> AAGCGACTTTGACG  0.29749737  0.27697712  0.07405920  0.376209061 -0.053281082
#> ACCAGTGAATACCG  0.41537014  0.56867645  0.40686749  0.578456644  0.229349100
#> ATTGCACTTGCTTT  0.37834956  0.59867963  0.42747634  0.559288991  0.147684325
#> CTAGGTGATGGTTG  0.34027336  0.48443293  0.36201305  0.492159319  0.185334690
#> GCACTAGACCTTTA  0.42123836  0.46746231  0.37390821  0.486823105  0.202023463
#> CATGCGCTAGTCAC  0.37534318  0.50526126  0.40930559  0.579777664  0.288796953
#> TTGAGGACTACGCA  0.20533038  0.46575040  0.24351374  0.333393308 -0.026056111
#> ATACCACTCTAAGC  0.27238748  0.26439998  0.21578748  0.352830182  0.064880328
#> CATATAGACTAAGC  0.13728129  0.30428691  0.33339743  0.147723440  0.255053841
#> TTTAGCTGTACTCT  0.36436700  0.49008659  0.31306325  0.556380243  0.110099237
#> GACATTCTCCACCT  0.29988016  0.64510329  0.47120083  0.448570392  0.154278419
#> ACGTGATGCCATGA  0.21366867  0.45552957  0.36935055  0.467887510  0.102463620
#> ATTGTAGATTCCCG  0.31529456  0.65312866  0.40993565  0.471646401  0.140832129
#> GATAGAGATCACGA  0.06468454  0.37270644  0.44648111  0.168830908  0.130999589
#> AATGCGTGGACGGA  0.26782768  0.47628134  0.40712014  0.431864551  0.017384090
#> GCGTAAACACGGTT  0.42244391  0.57899868  0.48113964  0.497736712  0.241493916
#> ATTCAGCTCATTGG -0.22681140 -0.16735451  0.25136315 -0.119645801  0.636263838
#> GGCATATGGGGAGT -0.16654798 -0.13677655  0.32333605  0.010467229  0.542050497
#> ATCATCTGACACCA -0.19995598 -0.18200752  0.33242052 -0.062533089  0.684760246
#> GTCATACTTCGCCT -0.20698003 -0.14134504  0.37829401 -0.106914334  0.624115252
#> TTACGTACGTTCAG -0.12950869 -0.04009166  0.41892992  0.063784120  0.591359437
#> GAGTTGTGGTAGCT -0.31582288 -0.18044397  0.31728052 -0.114591964  0.740781935
#> GACGCTCTCTCTCG -0.09820454 -0.08318004  0.29115127 -0.014388904  0.599226392
#> AGTCTTACTTCGGA -0.28854420 -0.24376898  0.23826091 -0.120278626  0.616511223
#> GGAACACTTCAGAC -0.19895923 -0.09902836  0.34308067 -0.001717814  0.663556914
#> CTTGATTGATCTTC -0.06052033  0.03568495  0.32498590  0.196386387  0.561213909
#>                Naive CD4 T           NK         pDCs
#> ATGCCAGAACGACT  0.13486507  0.171712022  0.153432667
#> CATGGCCTGTGCAT  0.30834768  0.198408099  0.138614282
#> GAACCTGATGAACC  0.46645269  0.441479237  0.329071463
#> TGACTGGATTCTCA  0.38882976  0.508979542  0.255886780
#> AGTCAGACTGCACA  0.30835262  0.277650903  0.177593551
#> TCTGATACACGTGT  0.45941430  0.305404143  0.265849007
#> TGGTATCTAAACAG  0.50622795  0.284702641  0.277353980
#> GCAGCTCTGTTTCT  0.58757963  0.174862701  0.167323690
#> GATATAACACGCAT  0.51925702  0.364110199  0.296871312
#> AATGTTGACAGTCA  0.33280635  0.220819683  0.320107735
#> AGGTCATGAGTGTC  0.05214146  0.007268359  0.236872788
#> AGAGATGATCTCGC  0.05045192  0.076643055  0.301915985
#> GGGTAACTCTAGTG  0.10428292 -0.127023235  0.047166470
#> CATGAGACACGGGA  0.18287893  0.117871505  0.173227809
#> TACGCCACTCCGAA  0.10326551  0.080461260  0.327238476
#> CTAAACCTGTGCAT  0.06190576  0.129422978  0.188001165
#> GTAAGCACTCATTC  0.11638496  0.187487486  0.242217984
#> TTGGTACTGAATCC -0.04047389 -0.144987456  0.326661355
#> CATCATACGGAGCA  0.20214229  0.116737491  0.147693893
#> TACATCACGCTAAC  0.17905630 -0.119410746  0.231096115
#> TTACCATGAATCGC  0.13941392  0.195180015  0.268915236
#> ATAGGAGAAACAGA  0.19964665  0.368228589  0.338620684
#> GCGCACGACTTTAC  0.07867848  0.166602138  0.152711289
#> ACTCGCACGAAAGT  0.38590801  0.295970811  0.391093428
#> ATTACCTGCCTTAT  0.20011396  0.403788606  0.479583474
#> CCCAACTGCAATCG  0.08871795  0.134186260  0.158904457
#> AAATTCGAATCACG  0.06738076  0.248804148  0.202045245
#> CCATCCGATTCGCC  0.19086881  0.265361389  0.184083241
#> TCCACTCTGAGCTT  0.24423112  0.377063551  0.282974334
#> CATCAGGATGCACA  0.47361913  0.423381851  0.389642257
#> CTAAACCTCTGACA  0.58565727  0.602772970  0.445426689
#> GATAGAGAAGGGTG -0.11406593  0.268372520 -0.134457618
#> CTAACGGAACCGAT  0.40352164  0.607775699  0.219535214
#> AGATATACCCGTAA  0.36367742  0.620998891  0.254142586
#> TACTCTGAATCGAC  0.31680432  0.300771637  0.054449528
#> GCGCATCTTGCTCC  0.31718064  0.325325353 -0.027558970
#> GTTGACGATATCGG  0.25714615  0.573213033  0.127670880
#> ACAGGTACTGGTGT  0.36714780  0.240894200  0.049940968
#> GGCATATGCTTATC  0.30557445  0.545964214 -0.005294209
#> CATTACACCAACTG  0.30688707  0.663812696  0.094673578
#> TAGGGACTGAACTC  0.35358163  0.505165159  0.241186346
#> GCTCCATGAGAAGT  0.59148316  0.457168433  0.285227680
#> TACAATGATGCTAG  0.42634040  0.368085551  0.180003114
#> CTTCATGACCGAAT  0.43438338  0.521437425  0.210890990
#> CTGCCAACAGGAGC  0.11021721  0.337112504 -0.095275295
#> TTGCATTGAGCTAC  0.35080593  0.337651626  0.397175213
#> AAGCAAGAGCTTAG  0.40322722  0.478034544  0.145439323
#> CGGCACGAACTCAG  0.50939117  0.460872918  0.127733275
#> GGTGGAGATTACTC  0.46976355  0.583002691  0.328783018
#> GGCCGATGTACTCT  0.54076967  0.314752862  0.086503191
#> CGTAGCCTGTATGC  0.12293513  0.228762405  0.308268130
#> TGAGCTGAATGCTG  0.44931923  0.518909977  0.384305762
#> CCTATAACGAGACG  0.19086881  0.265361389  0.184083241
#> ATAAGTTGGTACGT          NA           NA           NA
#> AAGCGACTTTGACG  0.41419013  0.186566150  0.232731943
#> ACCAGTGAATACCG  0.47081398  0.557710773  0.529808025
#> ATTGCACTTGCTTT  0.54798894  0.623494651  0.346052347
#> CTAGGTGATGGTTG  0.37349769  0.469766284  0.457772359
#> GCACTAGACCTTTA  0.35897057  0.459102483  0.327238476
#> CATGCGCTAGTCAC  0.41338817  0.517633963  0.394877989
#> TTGAGGACTACGCA  0.23855093  0.293329111  0.210559851
#> ATACCACTCTAAGC  0.29157494  0.287716645  0.321381911
#> CATATAGACTAAGC  0.20375647  0.313785816  0.314420369
#> TTTAGCTGTACTCT  0.46396066  0.530414648  0.490029611
#> GACATTCTCCACCT  0.35570342  0.445268234  0.440087071
#> ACGTGATGCCATGA  0.38406721  0.438052278  0.243468559
#> ATTGTAGATTCCCG  0.38126120  0.502204127  0.463067794
#> GATAGAGATCACGA  0.05245232  0.239806108  0.132792425
#> AATGCGTGGACGGA  0.32955822  0.372641804  0.354601236
#> GCGTAAACACGGTT  0.48098802  0.492713445  0.432134166
#> ATTCAGCTCATTGG -0.21013933 -0.162509200  0.179238766
#> GGCATATGGGGAGT -0.19462505 -0.132230907  0.222723351
#> ATCATCTGACACCA -0.16078068 -0.093682148  0.126928529
#> GTCATACTTCGCCT -0.20167931 -0.124246488  0.232884017
#> TTACGTACGTTCAG -0.16173830 -0.071039757  0.283561923
#> GAGTTGTGGTAGCT -0.28092700 -0.017005847  0.097372784
#> GACGCTCTCTCTCG -0.12443132 -0.011016535  0.351665032
#> AGTCTTACTTCGGA -0.28292813 -0.226567125  0.079678557
#> GGAACACTTCAGAC -0.21086171 -0.112982390  0.200466073
#> CTTGATTGATCTTC -0.03646355  0.137068883  0.215710865