Compare scRNA-seq data to reference data.

Compare scRNA-seq data to reference data.

Usage

clustify(input, ...)

# S3 method for default
clustify(
  input,
  ref_mat,
  metadata = NULL,
  cluster_col = NULL,
  query_genes = NULL,
  n_genes = 1000,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  pseudobulk_method = "mean",
  verbose = TRUE,
  lookuptable = NULL,
  rm0 = FALSE,
  obj_out = TRUE,
  seurat_out = TRUE,
  vec_out = FALSE,
  rename_prefix = NULL,
  threshold = "auto",
  low_threshold_cell = 0,
  exclude_genes = c(),
  if_log = TRUE,
  organism = "hsapiens",
  plot_name = NULL,
  rds_name = NULL,
  expand_unassigned = FALSE,
  ...
)

# S3 method for Seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  n_genes = 1000,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  pseudobulk_method = "mean",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  vec_out = FALSE,
  threshold = "auto",
  verbose = TRUE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  metadata = NULL,
  organism = "hsapiens",
  plot_name = NULL,
  rds_name = NULL,
  expand_unassigned = FALSE,
  ...
)

# S3 method for SingleCellExperiment
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  pseudobulk_method = "mean",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  vec_out = FALSE,
  threshold = "auto",
  verbose = TRUE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  metadata = NULL,
  organism = "hsapiens",
  plot_name = NULL,
  rds_name = NULL,
  expand_unassigned = FALSE,
  ...
)

Arguments

input

single-cell expression matrix or Seurat object

...

additional arguments to pass to compute_method function

ref_mat

reference expression matrix

metadata

cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set. Not required if running correlation per cell.

cluster_col

column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes

A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

n_genes

number of genes limit for Seurat variable genes, by default 1000, set to 0 to use all variable genes (generally not recommended)

per_cell

if true run per cell, otherwise per cluster.

n_perm

number of permutations, set to 0 by default

compute_method

method(s) for computing similarity scores

pseudobulk_method

method used for summarizing clusters, options are mean (default), median, truncate (10% truncated mean), or trimean, max, min

verbose

whether to report certain variables chosen and steps

lookuptable

if not supplied, will look in built-in table for object parsing

rm0

consider 0 as missing data, recommended for per_cell

obj_out

whether to output object instead of cor matrix

seurat_out

output cor matrix or called seurat object (deprecated, use obj_out instead)

vec_out

only output a result vector in the same order as metadata

rename_prefix

prefix to add to type and r column names

threshold

identity calling minimum correlation score threshold, only used when obj_out = TRUE

low_threshold_cell

option to remove clusters with too few cells

exclude_genes

a vector of gene names to throw out of query

if_log

input data is natural log, averaging will be done on unlogged data

organism

for GO term analysis, organism name: human - 'hsapiens', mouse - 'mmusculus'

plot_name

name for saved pdf, if NULL then no file is written (default)

rds_name

name for saved rds of rank_diff, if NULL then no file is written (default)

expand_unassigned

test all ref clusters for unassigned results

use_var_genes

if providing a seurat object, use the variable genes (stored in seurat_object@var.genes) as the query_genes.

dr

stored dimension reduction

Value

single cell object with identity assigned in metadata, or matrix of correlation values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Annotate a matrix and metadata
clustify(
    input = pbmc_matrix_small,
    metadata = pbmc_meta,
    ref_mat = cbmc_ref,
    query_genes = pbmc_vargenes,
    cluster_col = "classified",
    verbose = TRUE
)
#> using # of genes: 599
#> similarity computation completed, matrix of 9 x 13, preparing output
#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T
#> B            0.9093577  0.5347126  0.5908755 0.6248639 0.6411407
#> CD14+ Mono   0.5279272  0.9145428  0.8919954 0.5103056 0.4636934
#> CD8 T        0.5791953  0.4853982  0.5498177 0.5663773 0.7657468
#> DC           0.6088123  0.7415535  0.7520484 0.5694074 0.4951613
#> FCGR3A+ Mono 0.5380921  0.8429769  0.9294914 0.5331927 0.4857454
#> Memory CD4 T 0.6268758  0.5373744  0.5866177 0.6265186 0.8610826
#> NK           0.5042912  0.4399003  0.5092065 0.5350808 0.6755032
#> Naive CD4 T  0.6443057  0.5213433  0.5625255 0.6480900 0.8891488
#> Platelet     0.1456751  0.2571662  0.2867421 0.2455947 0.1422081
#>                  CD8 T        DC     Eryth Memory CD4 T        Mk
#> B            0.6269530 0.6036723 0.4817829    0.4803367 0.2241017
#> CD14+ Mono   0.4381450 0.7643448 0.4559350    0.4553454 0.4227047
#> CD8 T        0.7692801 0.5187820 0.5007841    0.6323214 0.2961827
#> DC           0.4789794 0.8493399 0.4470039    0.4764929 0.2948557
#> FCGR3A+ Mono 0.4643569 0.7235944 0.4604308    0.4614474 0.3937726
#> Memory CD4 T 0.8347482 0.5541873 0.5395170    0.7128879 0.2635539
#> NK           0.6926741 0.4817286 0.4578726    0.5665250 0.2783269
#> Naive CD4 T  0.8707703 0.5419174 0.5528441    0.7277404 0.2668581
#> Platelet     0.1208273 0.1898078 0.3068445    0.1457308 0.7319575
#>              Naive CD4 T        NK      pDCs
#> B              0.5407149 0.5484213 0.5823418
#> CD14+ Mono     0.4564707 0.5917870 0.4738376
#> CD8 T          0.6036230 0.8257838 0.4536589
#> DC             0.4475783 0.5643991 0.6825388
#> FCGR3A+ Mono   0.4574654 0.6028156 0.4797367
#> Memory CD4 T   0.6871755 0.6999959 0.4696025
#> NK             0.5505877 0.8940904 0.4185162
#> Naive CD4 T    0.6967073 0.6880495 0.4631464
#> Platelet       0.1609963 0.1396979 0.2152614

# Annotate using a different method
clustify(
    input = pbmc_matrix_small,
    metadata = pbmc_meta,
    ref_mat = cbmc_ref,
    query_genes = pbmc_vargenes,
    cluster_col = "classified",
    compute_method = "cosine"
)
#> using # of genes: 599
#> similarity computation completed, matrix of 9 x 13, preparing output
#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T
#> B            0.9669867  0.7149343  0.7720125 0.8004738 0.7635724
#> CD14+ Mono   0.7527223  0.9627275  0.9460084 0.7577338 0.7096110
#> CD8 T        0.7510063  0.7070527  0.7467587 0.7677927 0.8562280
#> DC           0.8230905  0.8842235  0.9086642 0.8037140 0.7246647
#> FCGR3A+ Mono 0.7590829  0.9172636  0.9762851 0.7746692 0.7286717
#> Memory CD4 T 0.7865337  0.7486511  0.7797697 0.8075621 0.9318837
#> NK           0.6784059  0.6671718  0.7107553 0.7238537 0.7777913
#> Naive CD4 T  0.7986227  0.7404425  0.7704809 0.8194404 0.9493149
#> Platelet     0.5463488  0.5932935  0.6068613 0.5953661 0.5558004
#>                  CD8 T        DC     Eryth Memory CD4 T        Mk
#> B            0.7657181 0.8062767 0.6514282    0.7123319 0.6594086
#> CD14+ Mono   0.7061112 0.8919577 0.6951238    0.7557225 0.7911713
#> CD8 T        0.8674614 0.7351337 0.6826181    0.8232194 0.7188642
#> DC           0.7260570 0.9560634 0.6874076    0.7519049 0.7448677
#> FCGR3A+ Mono 0.7274581 0.8800476 0.6880012    0.7601400 0.7722167
#> Memory CD4 T 0.9235744 0.7594695 0.7101214    0.8855087 0.7122113
#> NK           0.7973584 0.6864934 0.6500106    0.7578212 0.6960715
#> Naive CD4 T  0.9444251 0.7508430 0.7181980    0.8877208 0.7125300
#> Platelet     0.5516889 0.5730797 0.5898023    0.5655621 0.8775257
#>              Naive CD4 T        NK      pDCs
#> B              0.7494337 0.6957714 0.7863873
#> CD14+ Mono     0.7542713 0.7675894 0.7495702
#> CD8 T          0.8123971 0.9018633 0.7085898
#> DC             0.7491931 0.7595771 0.8503681
#> FCGR3A+ Mono   0.7561998 0.7776798 0.7477358
#> Memory CD4 T   0.8731255 0.8102022 0.7241459
#> NK             0.7544227 0.9480188 0.6812692
#> Naive CD4 T    0.8847456 0.8097839 0.7246787
#> Platelet       0.5455913 0.5569351 0.5756617

# Annotate a Seurat object
clustify(
    s_small3,
    cbmc_ref,
    cluster_col = "RNA_snn_res.1",
    obj_out = TRUE,
    per_cell = FALSE,
    dr = "tsne"
)
#> object data retrieval complete, moving to similarity computation
#> using # of genes: 19
#> similarity computation completed, matrix of 3 x 13, preparing output
#> using threshold of 0.55
#> An object of class Seurat 
#> 230 features across 80 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

# Annotate (and return) a Seurat object per-cell
clustify(
    input = s_small3,
    ref_mat = cbmc_ref,
    cluster_col = "RNA_snn_res.1",
    obj_out = TRUE,
    per_cell = TRUE,
    dr = "tsne"
)
#> object data retrieval complete, moving to similarity computation
#> using # of genes: 19
#> similarity computation completed, matrix of 80 x 13, preparing output
#> using threshold of NA
#> An object of class Seurat 
#> 230 features across 80 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne