Compare scRNA-seq data to reference data.

Compare scRNA-seq data to reference data.

clustify(input, ...)

# S3 method for default
clustify(
  input,
  ref_mat,
  metadata = NULL,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  verbose = FALSE,
  lookuptable = NULL,
  rm0 = FALSE,
  obj_out = FALSE,
  seurat_out = FALSE,
  rename_prefix = NULL,
  threshold = "auto",
  exclude_genes = rownames(ref_mat)[stringr::str_detect(rownames(ref_mat),
    "RP[0-9,L,S]|Rp[0-9,l,s]")],
  ...
)

# S3 method for seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  ...
)

# S3 method for Seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  ...
)

# S3 method for SingleCellExperiment
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  ...
)

Arguments

input
single-cell expression matrix or Seurat object

...
additional arguments to pass to compute_method function

ref_mat
reference expression matrix

metadata
cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set. Not required if running correlation per cell.

cluster_col
column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes
A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

per_cell
if true run per cell, otherwise per cluster.

n_perm
number of permutations, set to 0 by default

compute_method
method(s) for computing similarity scores

verbose
whether to report certain variables chosen

lookuptable
if not supplied, will look in built-in table for object parsing

rm0
consider 0 as missing data, recommended for per_cell

obj_out
whether to output object instead of cor matrix

seurat_out
output cor matrix or called seurat object

rename_prefix
prefix to add to type and r column names

threshold
identity calling minimum correlation score threshold, only used when obj_out = TRUE

exclude_genes
a vector of gene names to throw out of query

use_var_genes
if providing a seurat object, use the variable genes (stored in seurat_object@var.genes) as the query_genes.

dr
stored dimension reduction

Value

single cell object with identity assigned in metadata, or matrix of correlation values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Annotate a matrix and metadata
clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = cbmc_ref,
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  verbose = TRUE
)

#> using # of genes: 1881

#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T     CD8 T
#> B            0.8912913  0.6851783  0.7317660 0.7317217 0.7814900 0.7659276
#> CD14+ Mono   0.6989945  0.9081472  0.8847734 0.6710434 0.6670570 0.6521402
#> CD8 T        0.7525971  0.6651899  0.7126303 0.7136651 0.8400829 0.8368499
#> DC           0.7026957  0.7823684  0.7882898 0.6814873 0.6658926 0.6524832
#> FCGR3A+ Mono 0.7039327  0.8496499  0.9059093 0.6726128 0.6757769 0.6596604
#> Memory CD4 T 0.7758978  0.6820664  0.7241914 0.7427243 0.8948977 0.8702287
#> Naive CD4 T  0.7871487  0.6792475  0.7175664 0.7548765 0.9158677 0.8936029
#> NK           0.7192197  0.6526278  0.6958540 0.7067672 0.7940021 0.7937825
#> Platelet     0.3164496  0.3823818  0.3918461 0.3271915 0.2858296 0.2797719
#>                     DC     Eryth Memory CD4 T        Mk Naive CD4 T        NK
#> B            0.7253935 0.6300171    0.5959394 0.5450655   0.5953671 0.7474430
#> CD14+ Mono   0.8121193 0.6055132    0.5708809 0.6549038   0.5384300 0.7573710
#> CD8 T        0.6840079 0.6525508    0.6542797 0.5673165   0.6121911 0.8662208
#> DC           0.8327092 0.6049950    0.5657932 0.5748673   0.5260392 0.7190055
#> FCGR3A+ Mono 0.7784995 0.6097144    0.5710313 0.6356263   0.5300822 0.7552514
#> Memory CD4 T 0.6965506 0.6643080    0.6860640 0.5514448   0.6554955 0.8182608
#> Naive CD4 T  0.6958788 0.6677468    0.6913118 0.5530940   0.6622109 0.8080934
#> NK           0.6741498 0.6420863    0.6259194 0.5642024   0.5880876 0.8960291
#> Platelet     0.3476335 0.3599792    0.2945712 0.6304742   0.2899873 0.3402365
#>                   pDCs
#> B            0.6889127
#> CD14+ Mono   0.6405745
#> CD8 T        0.6317819
#> DC           0.7012226
#> FCGR3A+ Mono 0.6374282
#> Memory CD4 T 0.6434303
#> Naive CD4 T  0.6452324
#> NK           0.6263499
#> Platelet     0.3349940

# Annotate using a different method
res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = cbmc_ref,
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  compute_method = "cosine"
)

# Annotate a Seurat object
clustify(
  s_small,
  cbmc_ref,
  cluster_col = "res.1",
  seurat_out = TRUE,
  per_cell = FALSE,
  dr = "tsne"
)

#> using threshold of 0.61

#> seurat not loaded, returning cor_mat instead

#>            B CD14+ Mono CD16+ Mono      CD34+       CD4 T      CD8 T
#> 0  0.5224110  0.3852326  0.4119037 0.41720039  0.61062598  0.5825024
#> 1  0.4136741  0.7251465  0.6939952 0.53311242  0.42401600  0.3648038
#> 2  0.2311589  0.6184823  0.6094641 0.42129146  0.46805261  0.4812869
#> 3 -0.1762557  0.1362385  0.1859978 0.01798528 -0.08048411 -0.1647901
#>            DC     Eryth Memory CD4 T          Mk Naive CD4 T        NK
#> 0  0.36476066 0.1523530    0.7047302 -0.10828917   0.6398803 0.4346006
#> 1  0.71146653 0.4359801    0.5349202  0.08053672   0.4304222 0.6174695
#> 2  0.67077257 0.4137920    0.6331816  0.18844401   0.5864320 0.8099383
#> 3 -0.05333518 0.4352437    0.1083785  0.74534070  -0.1097800 0.1124080
#>        pDCs
#> 0 0.3456669
#> 1 0.5116362
#> 2 0.2859966
#> 3 0.1901284

# Annotate (and return) a Seurat object per-cell
res <- clustify(
  input = s_small,
  ref_mat = cbmc_ref,
  cluster_col = "res.1",
  seurat_out = TRUE,
  per_cell = TRUE,
  dr = "tsne"
)

#> using threshold of NA

#> seurat not loaded, returning cor_mat instead