Compare scRNA-seq data to reference data.

Compare scRNA-seq data to reference data.

clustify(input, ...)

# S3 method for default
clustify(
  input,
  ref_mat,
  metadata = NULL,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  verbose = FALSE,
  lookuptable = NULL,
  rm0 = FALSE,
  obj_out = TRUE,
  seurat_out = TRUE,
  rename_prefix = NULL,
  threshold = "auto",
  low_threshold_cell = 0,
  exclude_genes = c(),
  if_log = TRUE,
  ...
)

# S3 method for seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  if_log = TRUE,
  ...
)

# S3 method for Seurat
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  if_log = TRUE,
  ...
)

# S3 method for SingleCellExperiment
clustify(
  input,
  ref_mat,
  cluster_col = NULL,
  query_genes = NULL,
  per_cell = FALSE,
  n_perm = 0,
  compute_method = "spearman",
  use_var_genes = TRUE,
  dr = "umap",
  seurat_out = TRUE,
  obj_out = TRUE,
  threshold = "auto",
  verbose = FALSE,
  rm0 = FALSE,
  rename_prefix = NULL,
  exclude_genes = c(),
  if_log = TRUE,
  ...
)

Arguments

input
single-cell expression matrix or Seurat object

...
additional arguments to pass to compute_method function

ref_mat
reference expression matrix

metadata
cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set. Not required if running correlation per cell.

cluster_col
column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes
A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

per_cell
if true run per cell, otherwise per cluster.

n_perm
number of permutations, set to 0 by default

compute_method
method(s) for computing similarity scores

verbose
whether to report certain variables chosen

lookuptable
if not supplied, will look in built-in table for object parsing

rm0
consider 0 as missing data, recommended for per_cell

obj_out
whether to output object instead of cor matrix

seurat_out
output cor matrix or called seurat object (deprecated, use obj_out instead)

rename_prefix
prefix to add to type and r column names

threshold
identity calling minimum correlation score threshold, only used when obj_out = TRUE

low_threshold_cell
option to remove clusters with too few cells

exclude_genes
a vector of gene names to throw out of query

if_log
input data is natural log, averaging will be done on unlogged data

use_var_genes
if providing a seurat object, use the variable genes (stored in seurat_object@var.genes) as the query_genes.

dr
stored dimension reduction

Value

single cell object with identity assigned in metadata, or matrix of correlation values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Annotate a matrix and metadata
clustify(
    input = pbmc_matrix_small,
    metadata = pbmc_meta,
    ref_mat = cbmc_ref,
    query_genes = pbmc_vargenes,
    cluster_col = "classified",
    verbose = TRUE
)

#> using # of genes: 599

#> The following clusters have less than 10 cells for this analysis: . Classification is likely inaccurate.

#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T     CD8 T
#> B            0.9093577  0.5347126  0.5908755 0.6248639 0.6411407 0.6269530
#> CD14+ Mono   0.5279272  0.9145428  0.8919954 0.5103056 0.4636934 0.4381450
#> CD8 T        0.5791953  0.4853982  0.5498177 0.5663773 0.7657468 0.7692801
#> DC           0.6088123  0.7415535  0.7520484 0.5694074 0.4951613 0.4789794
#> FCGR3A+ Mono 0.5380921  0.8429769  0.9294914 0.5331927 0.4857454 0.4643569
#> Memory CD4 T 0.6268758  0.5373744  0.5866177 0.6265186 0.8610826 0.8347482
#> Naive CD4 T  0.6443057  0.5213433  0.5625255 0.6480900 0.8891488 0.8707703
#> NK           0.5042912  0.4399003  0.5092065 0.5350808 0.6755032 0.6926741
#> Platelet     0.1456751  0.2571662  0.2867421 0.2455947 0.1422081 0.1208273
#>                     DC     Eryth Memory CD4 T        Mk Naive CD4 T        NK
#> B            0.6036723 0.4817829    0.4803367 0.2241017   0.5407149 0.5484213
#> CD14+ Mono   0.7643448 0.4559350    0.4553454 0.4227047   0.4564707 0.5917870
#> CD8 T        0.5187820 0.5007841    0.6323214 0.2961827   0.6036230 0.8257838
#> DC           0.8493399 0.4470039    0.4764929 0.2948557   0.4475783 0.5643991
#> FCGR3A+ Mono 0.7235944 0.4604308    0.4614474 0.3937726   0.4574654 0.6028156
#> Memory CD4 T 0.5541873 0.5395170    0.7128879 0.2635539   0.6871755 0.6999959
#> Naive CD4 T  0.5419174 0.5528441    0.7277404 0.2668581   0.6967073 0.6880495
#> NK           0.4817286 0.4578726    0.5665250 0.2783269   0.5505877 0.8940904
#> Platelet     0.1898078 0.3068445    0.1457308 0.7319575   0.1609963 0.1396979
#>                   pDCs
#> B            0.5823418
#> CD14+ Mono   0.4738376
#> CD8 T        0.4536589
#> DC           0.6825388
#> FCGR3A+ Mono 0.4797367
#> Memory CD4 T 0.4696025
#> Naive CD4 T  0.4631464
#> NK           0.4185162
#> Platelet     0.2152614

# Annotate using a different method
clustify(
    input = pbmc_matrix_small,
    metadata = pbmc_meta,
    ref_mat = cbmc_ref,
    query_genes = pbmc_vargenes,
    cluster_col = "classified",
    compute_method = "cosine"
)

#> The following clusters have less than 10 cells for this analysis: . Classification is likely inaccurate.

#>                      B CD14+ Mono CD16+ Mono     CD34+     CD4 T     CD8 T
#> B            0.9669867  0.7149343  0.7720125 0.8004738 0.7635724 0.7657181
#> CD14+ Mono   0.7527223  0.9627275  0.9460084 0.7577338 0.7096110 0.7061112
#> CD8 T        0.7510063  0.7070527  0.7467587 0.7677927 0.8562280 0.8674614
#> DC           0.8230905  0.8842235  0.9086642 0.8037140 0.7246647 0.7260570
#> FCGR3A+ Mono 0.7590829  0.9172636  0.9762851 0.7746692 0.7286717 0.7274581
#> Memory CD4 T 0.7865337  0.7486511  0.7797697 0.8075621 0.9318837 0.9235744
#> Naive CD4 T  0.7986227  0.7404425  0.7704809 0.8194404 0.9493149 0.9444251
#> NK           0.6784059  0.6671718  0.7107553 0.7238537 0.7777913 0.7973584
#> Platelet     0.5463488  0.5932935  0.6068613 0.5953661 0.5558004 0.5516889
#>                     DC     Eryth Memory CD4 T        Mk Naive CD4 T        NK
#> B            0.8062767 0.6514282    0.7123319 0.6594086   0.7494337 0.6957714
#> CD14+ Mono   0.8919577 0.6951238    0.7557225 0.7911713   0.7542713 0.7675894
#> CD8 T        0.7351337 0.6826181    0.8232194 0.7188642   0.8123971 0.9018633
#> DC           0.9560634 0.6874076    0.7519049 0.7448677   0.7491931 0.7595771
#> FCGR3A+ Mono 0.8800476 0.6880012    0.7601400 0.7722167   0.7561998 0.7776798
#> Memory CD4 T 0.7594695 0.7101214    0.8855087 0.7122113   0.8731255 0.8102022
#> Naive CD4 T  0.7508430 0.7181980    0.8877208 0.7125300   0.8847456 0.8097839
#> NK           0.6864934 0.6500106    0.7578212 0.6960715   0.7544227 0.9480188
#> Platelet     0.5730797 0.5898023    0.5655621 0.8775257   0.5455913 0.5569351
#>                   pDCs
#> B            0.7863873
#> CD14+ Mono   0.7495702
#> CD8 T        0.7085898
#> DC           0.8503681
#> FCGR3A+ Mono 0.7477358
#> Memory CD4 T 0.7241459
#> Naive CD4 T  0.7246787
#> NK           0.6812692
#> Platelet     0.5756617

# Annotate a Seurat object
clustify(
    s_small,
    cbmc_ref,
    cluster_col = "res.1",
    obj_out = TRUE,
    per_cell = FALSE,
    dr = "tsne"
)

#> The following clusters have less than 10 cells for this analysis: . Classification is likely inaccurate.

#> using threshold of 0.59

#> seurat not loaded, returning cor_mat instead

#>             B CD14+ Mono CD16+ Mono      CD34+       CD4 T      CD8 T        DC
#> 0  0.66641984  0.4000028  0.3435278 0.24453004  0.53962648  0.4400031 0.3605451
#> 1  0.49928480  0.5282509  0.4731645 0.52825094  0.34835596  0.2873746 0.4992739
#> 2  0.04411189  0.3881846  0.4854133 0.27172921  0.41288725  0.4446478 0.5601817
#> 3 -0.33741199  0.3510691  0.3407536 0.08194291 -0.08997653 -0.1831665 0.0941350
#>         Eryth Memory CD4 T          Mk Naive CD4 T        NK      pDCs
#> 0 -0.05358529   0.65833004 -0.30461152   0.6063574 0.3441534 0.4155303
#> 1  0.40476370   0.50267909  0.02137564   0.3693701 0.5633152 0.3852524
#> 2  0.26467131   0.46802523  0.03887696   0.3960427 0.7287283 0.1487763
#> 3  0.53825246   0.07327631  0.78848121  -0.2140216 0.1261278 0.2326462

# Annotate (and return) a Seurat object per-cell
clustify(
    input = s_small,
    ref_mat = cbmc_ref,
    cluster_col = "res.1",
    obj_out = TRUE,
    per_cell = TRUE,
    dr = "tsne"
)

#> using threshold of NA

#> seurat not loaded, returning cor_mat instead

#>                          B   CD14+ Mono  CD16+ Mono       CD34+       CD4 T
#> ATGCCAGAACGACT  0.30783491  0.489737351  0.55991036  0.35949951  0.34335432
#> CATGGCCTGTGCAT  0.31929826  0.324136109  0.32909771  0.15239235  0.51039342
#> GAACCTGATGAACC  0.34846603  0.426142664  0.51264280  0.23302991  0.55344604
#> TGACTGGATTCTCA  0.43045803  0.610624678  0.63567707  0.43800993  0.53294804
#> AGTCAGACTGCACA  0.45852332  0.425156601  0.47807731  0.25832300  0.58876117
#> TCTGATACACGTGT  0.39103715  0.364403041  0.46021675  0.28813264  0.59805682
#> TGGTATCTAAACAG  0.29127763  0.346759078  0.41032074  0.19220360  0.70540704
#> GCAGCTCTGTTTCT  0.27344430  0.293259106  0.24976045  0.13870363  0.74107369
#> GATATAACACGCAT  0.35316322  0.338649666  0.43315066  0.26850081  0.59747477
#> AATGTTGACAGTCA  0.36595758  0.373492002  0.42423977  0.20450571  0.57584502
#> AGGTCATGAGTGTC  0.55205245  0.134381188  0.21860296  0.05811078  0.18401748
#> AGAGATGATCTCGC  0.44132518  0.258291371  0.28766636  0.40973559  0.14308227
#> GGGTAACTCTAGTG  0.34348752 -0.036283893  0.02298844 -0.08466242  0.16206805
#> CATGAGACACGGGA  0.57328551  0.164486981  0.23472410  0.16206805  0.21286550
#> TACGCCACTCCGAA  0.39103715  0.249392114  0.24161380  0.58716105  0.13801311
#> CTAAACCTGTGCAT  0.33875645  0.143486011  0.13922299  0.11867264  0.07659780
#> GTAAGCACTCATTC  0.52098807  0.326158738  0.35659961  0.18328389  0.24678382
#> TTGGTACTGAATCC  0.57859229 -0.120870307 -0.06045789  0.10303698 -0.20607397
#> CATCATACGGAGCA  0.46614722  0.051954489  0.08229222  0.07360219  0.08226127
#> TACATCACGCTAAC  0.43592570 -0.009907402  0.02081337  0.21003693  0.29920355
#> TTACCATGAATCGC  0.34684399  0.520265982  0.52046168  0.43355498  0.37574765
#> ATAGGAGAAACAGA  0.29585195  0.427181356  0.46054767  0.36945415  0.29296559
#> GCGCACGACTTTAC  0.09946434  0.377964473  0.33830595  0.33817874  0.13925007
#> ACTCGCACGAAAGT  0.33626100  0.520988072  0.51829660  0.43584044  0.36656778
#> ATTACCTGCCTTAT  0.32425373  0.699234275  0.71342754  0.48638060  0.44261629
#> CCCAACTGCAATCG  0.09946434  0.377964473  0.33830595  0.33817874  0.13925007
#> AAATTCGAATCACG  0.09946434  0.377964473  0.33830595  0.33817874  0.13925007
#> CCATCCGATTCGCC          NA           NA          NA          NA          NA
#> TCCACTCTGAGCTT  0.09946434  0.377964473  0.33830595  0.33817874  0.13925007
#> CATCAGGATGCACA  0.33090723  0.614258937  0.65809896  0.51518492  0.59642561
#> CTAAACCTCTGACA  0.51089064  0.521786414  0.56437107  0.34624342  0.51452257
#> GATAGAGAAGGGTG -0.37796447 -0.258607271 -0.21890385 -0.25860727 -0.17903580
#> CTAACGGAACCGAT  0.07384912  0.176753634  0.21920851  0.04721501  0.19249197
#> AGATATACCCGTAA  0.28738434  0.344430664  0.41024202  0.19051321  0.38210277
#> TACTCTGAATCGAC -0.09669308  0.076488554  0.15592210  0.03752269  0.20781796
#> GCGCATCTTGCTCC  0.17903580  0.298393005  0.25870455  0.13925007  0.21882154
#> GTTGACGATATCGG -0.14755450  0.215284430  0.18632738  0.08466242  0.10885168
#> ACAGGTACTGGTGT -0.09669308  0.076488554  0.15592210  0.03752269  0.20781796
#> GGCATATGCTTATC -0.12411350  0.049068129  0.04619914 -0.07215901  0.04329541
#> CATTACACCAACTG -0.13100390  0.297313809  0.35876968  0.20811967  0.24714210
#> TAGGGACTGAACTC  0.20805545  0.645962625  0.50546757  0.47555531  0.45970346
#> GCTCCATGAGAAGT  0.32783314  0.424968887  0.32795646  0.32783314  0.49782070
#> TACAATGATGCTAG  0.33776893  0.422513825  0.33789598  0.32687316  0.49515230
#> CTTCATGACCGAAT  0.51089064  0.521786414  0.56437107  0.34624342  0.51452257
#> CTGCCAACAGGAGC  0.02742042  0.079374914  0.13715370  0.01876134  0.16452255
#> TTGCATTGAGCTAC  0.49136232  0.404651319  0.49154714  0.31794032  0.46245865
#> AAGCAAGAGCTTAG  0.34203372  0.378113227  0.40279875  0.22946566  0.34492008
#> CGGCACGAACTCAG  0.43355498  0.476910483  0.54214758  0.32516624  0.65033248
#> GGTGGAGATTACTC  0.33689624  0.689937675  0.65358767  0.51126427  0.47574486
#> GGCCGATGTACTCT  0.24433050  0.395018918  0.30256695  0.30568221  0.61674616
#> CGTAGCCTGTATGC  0.34684399  0.520265982  0.52046168  0.43355498  0.37574765
#> TGAGCTGAATGCTG  0.38132031  0.659231724  0.67887616  0.54935977  0.48257486
#> CCTATAACGAGACG          NA           NA          NA          NA          NA
#> ATAAGTTGGTACGT          NA           NA          NA          NA          NA
#> AAGCGACTTTGACG  0.23122933  0.260132991  0.31805992  0.20232566  0.52026598
#> ACCAGTGAATACCG  0.45762242  0.575054633  0.61765933  0.47941397  0.46125435
#> ATTGCACTTGCTTT  0.38861587  0.621059003  0.59464849  0.45641178  0.43825216
#> CTAGGTGATGGTTG  0.46685275  0.561190875  0.61221853  0.46685275  0.46443383
#> GCACTAGACCTTTA  0.48304589  0.571422709  0.62129262  0.47215012  0.48304589
#> CATGCGCTAGTCAC  0.27144836  0.689392653  0.64223840  0.52566190  0.43302476
#> TTGAGGACTACGCA -0.08499378  0.473536760  0.25507724  0.40068495  0.25498133
#> ATACCACTCTAAGC  0.48779493  0.399760931  0.48797841  0.31461330  0.45748814
#> CATATAGACTAAGC  0.18212952  0.303549205  0.27937032  0.01214197  0.23069740
#> TTTAGCTGTACTCT  0.48304589  0.571422709  0.62129262  0.47215012  0.48304589
#> GACATTCTCCACCT  0.28903666  0.433554985  0.46263261  0.37574765  0.28903666
#> ACGTGATGCCATGA  0.09946434  0.377964473  0.33830595  0.33817874  0.13925007
#> ATTGTAGATTCCCG  0.28142015  0.438726798  0.46343512  0.38099959  0.28430651
#> GATAGAGATCACGA  0.10774708  0.284500713  0.40692850  0.18886005  0.10411515
#> AATGCGTGGACGGA  0.47578204  0.560526937  0.61402604  0.46488627  0.47215012
#> GCGTAAACACGGTT  0.53279118  0.670563450  0.69881120  0.50480619  0.55108906
#> ATTCAGCTCATTGG -0.25770031  0.190030808  0.25501526 -0.17705309 -0.18724987
#> GGCATATGGGGAGT -0.23335817  0.211604443  0.24828362 -0.11766789 -0.09294775
#> ATCATCTGACACCA -0.23552514  0.224471618  0.21690073  0.03741194 -0.12839096
#> GTCATACTTCGCCT -0.21239972  0.201269606  0.18000457  0.01391265 -0.13819895
#> TTACGTACGTTCAG -0.21610976  0.291238048  0.32938980 -0.06678070 -0.08625840
#> GAGTTGTGGTAGCT -0.48497986  0.139153813  0.11449500 -0.05434409 -0.26595670
#> GACGCTCTCTCTCG -0.10672166  0.290520075  0.20166118  0.14624820 -0.02569225
#> AGTCTTACTTCGGA -0.38120649  0.117793733  0.18557172 -0.25970272 -0.23558747
#> GGAACACTTCAGAC -0.28393408  0.291894845  0.33801749 -0.08756845 -0.14329383
#> CTTGATTGATCTTC -0.19299074  0.375725933  0.32083955  0.33190678 -0.03076664
#>                      CD8 T           DC        Eryth Memory CD4 T          Mk
#> ATGCCAGAACGACT  0.28200261  0.431187221  0.303529523   0.44461201  0.24469875
#> CATGGCCTGTGCAT  0.46443383  0.271328052  0.091919195   0.45141476 -0.01453543
#> GAACCTGATGAACC  0.51245004  0.336025602  0.129461062   0.60173142  0.13397805
#> TGACTGGATTCTCA  0.45095603  0.602901241  0.436931086   0.64471223  0.19556474
#> AGTCAGACTGCACA  0.52525676  0.454902518  0.277697223   0.58228375 -0.01401358
#> TCTGATACACGTGT  0.55205245  0.403751165  0.062953349   0.57877995 -0.07153549
#> TGGTATCTAAACAG  0.63011078  0.357204033  0.045574050   0.68970124 -0.15875735
#> GCAGCTCTGTTTCT  0.68361075  0.303623428 -0.114925866   0.61491436 -0.26591856
#> GATATAACACGCAT  0.54909624  0.382766359  0.026608188   0.56553647 -0.09690288
#> AATGTTGACAGTCA  0.51449330  0.372976946  0.226032624   0.62742202 -0.04635263
#> AGGTCATGAGTGTC  0.16101530  0.173382633 -0.006053207   0.31350580 -0.14064605
#> AGAGATGATCTCGC  0.16352259  0.140506400  0.242496575   0.44905461  0.05489985
#> GGGTAACTCTAGTG  0.14029772 -0.036338578 -0.358001075   0.21302719 -0.28586348
#> CATGAGACACGGGA  0.16690591  0.152622029 -0.002418926   0.29671644 -0.15019946
#> TACGCCACTCCGAA  0.23002185  0.196419486  0.223968646   0.28938997  0.08972248
#> CTAAACCTGTGCAT  0.01078842  0.184760058 -0.086307375   0.05089833 -0.27227798
#> GTAAGCACTCATTC  0.24678382  0.320868890  0.059170390   0.14373989 -0.02023498
#> TTGGTACTGAATCC -0.11690735 -0.123036945 -0.219944329  -0.13503187 -0.12105248
#> CATCATACGGAGCA  0.05484085 -0.036133884 -0.187613433   0.14373989 -0.15754373
#> TACATCACGCTAAC  0.37251832 -0.003968934 -0.180314720   0.13918670 -0.27187196
#> TTACCATGAATCGC  0.34684399  0.463155650  0.289036657   0.48484848  0.17368337
#> ATAGGAGAAACAGA  0.23523838  0.518882574  0.460374502   0.42516747  0.08093990
#> GCGCACGACTTTAC  0.09946434  0.378534126  0.338178739   0.27112769  0.13945994
#> ACTCGCACGAAAGT  0.33626100  0.466849781  0.298738313   0.47963733  0.17488800
#> ATTACCTGCCTTAT  0.38094840  0.586727895  0.561973493   0.73725874  0.32972315
#> CCCAACTGCAATCG  0.09946434  0.378534126  0.338178739   0.27112769  0.13945994
#> AAATTCGAATCACG  0.09946434  0.378534126  0.338178739   0.27112769  0.13945994
#> CCATCCGATTCGCC          NA           NA           NA           NA          NA
#> TCCACTCTGAGCTT  0.09946434  0.378534126  0.338178739   0.27112769  0.13945994
#> CATCAGGATGCACA  0.51518492  0.650905127  0.390351647   0.77487519  0.03770487
#> CTAAACCTCTGACA  0.43825216  0.545609683  0.353507269   0.47977811  0.19278209
#> GATAGAGAAGGGTG -0.17903580 -0.258997034 -0.179035803  -0.22941573 -0.05976855
#> CTAACGGAACCGAT  0.11622157  0.246130590  0.215494157   0.11804065  0.10063467
#> AGATATACCCGTAA  0.31214029  0.380522722  0.251864923   0.33515169  0.13151210
#> TACTCTGAATCGAC  0.33914736  0.323759600 -0.038965867   0.09380919 -0.06648635
#> GCGCATCTTGCTCC  0.13925007  0.298842731  0.218821537   0.10427988  0.21915134
#> GTTGACGATATCGG  0.03144604  0.191383180  0.205608726   0.15977039  0.15019946
#> ACAGGTACTGGTGT  0.33914736  0.323759600 -0.038965867   0.09380919 -0.06648635
#> GGCATATGCTTATC -0.01731816  0.049142082  0.043295408  -0.07867868  0.12574592
#> CATTACACCAACTG  0.26386601  0.471766523  0.281519013   0.32826769  0.03256771
#> TAGGGACTGAACTC  0.35468500  0.676703196  0.517166396   0.59829506  0.12899035
#> GCTCCATGAGAAGT  0.47353676  0.425609383  0.157845587   0.54738221 -0.06080134
#> TACAATGATGCTAG  0.46851820  0.434062814  0.176753634   0.54704859 -0.05941083
#> CTTCATGACCGAAT  0.43825216  0.545609683  0.353507269   0.47977811  0.19278209
#> CTGCCAACAGGAGC  0.16452255  0.018789620 -0.077931734   0.14373989  0.03468853
#> TTGCATTGAGCTAC  0.43355498  0.434208421  0.260132991   0.51515152  0.05789446
#> AAGCAAGAGCTTAG  0.25688609  0.461068359  0.372340506   0.29504504  0.15320767
#> CGGCACGAACTCAG  0.56362148  0.521050106  0.173421994   0.56818182  0.04342084
#> GGTGGAGATTACTC  0.39178988  0.621987566  0.452065247   0.61049517  0.28350560
#> GGCCGATGTACTCT  0.57799771  0.403160051 -0.008610767   0.60598134 -0.17139692
#> CGTAGCCTGTATGC  0.34684399  0.463155650  0.289036657   0.48484848  0.17368337
#> TGAGCTGAATGCTG  0.42009865  0.651594896  0.508427081   0.72276945  0.16181992
#> CCTATAACGAGACG          NA           NA           NA           NA          NA
#> ATAAGTTGGTACGT          NA           NA           NA           NA          NA
#> AAGCGACTTTGACG  0.49136232  0.231577825 -0.144518328   0.45454545 -0.08684168
#> ACCAGTGAATACCG  0.41161805  0.598958185  0.433409597   0.59147249  0.13943359
#> ATTGCACTTGCTTT  0.36561368  0.575921332  0.387405226   0.47089333  0.27765471
#> CTAGGTGATGGTTG  0.41363638  0.598375258  0.437825639   0.59089685  0.12597374
#> GCACTAGACCTTTA  0.43704152  0.588045992  0.407986129   0.60289578  0.13579619
#> CATGCGCTAGTCAC  0.34685068  0.627861307  0.491192265   0.61661268  0.27401507
#> TTGAGGACTACGCA  0.20641346  0.425609383  0.303549205   0.49646294  0.01216027
#> ATACCACTCTAAGC  0.42718136  0.436497318  0.268431528   0.51141141  0.05347815
#> CATATAGACTAAGC  0.15784559  0.328327239  0.376401015   0.11456837  0.32832724
#> TTTAGCTGTACTCT  0.43704152  0.588045992  0.407986129   0.60289578  0.13579619
#> GACATTCTCCACCT  0.23122933  0.521050106  0.462458650   0.42424242  0.08684168
#> ACGTGATGCCATGA  0.09946434  0.378534126  0.338178739   0.27112769  0.13945994
#> ATTGTAGATTCCCG  0.22657930  0.521773285  0.463260862   0.42214137  0.09250274
#> GATAGAGATCACGA  0.03147667  0.434062814  0.613795155   0.23354279  0.07881029
#> AATGCGTGGACGGA  0.42251382  0.595320787  0.429777673   0.59528025  0.12488399
#> GCGTAAACACGGTT  0.47143947  0.686665649  0.482202930   0.61049517  0.22745126
#> ATTCAGCTCATTGG -0.23452583 -0.013925650  0.357814106  -0.12245424  0.70278113
#> GGCATATGGGGAGT -0.13151117  0.042582732  0.400466353   0.07153083  0.56050759
#> ATCATCTGACACCA -0.20916673 -0.021288819  0.431087539  -0.11321250  0.78938942
#> GTCATACTTCGCCT -0.20034210  0.029725044  0.462827343  -0.09918632  0.68460491
#> TTACGTACGTTCAG -0.13634393  0.134691604  0.521260455   0.08460010  0.61122121
#> GAGTTGTGGTAGCT -0.35735358 -0.110501268  0.399346741  -0.20545584  0.79577406
#> GACGCTCTCTCTCG -0.07905308  0.075213616  0.361667848   0.02072010  0.67494324
#> AGTCTTACTTCGGA -0.28752801 -0.105895468  0.332048475  -0.12738635  0.66045331
#> GGAACACTTCAGAC -0.20521092  0.080613530  0.465262693   0.01298296  0.70957623
#> CTTGATTGATCTTC -0.11933727  0.118583402  0.364538064   0.17887529  0.63306730
#>                Naive CD4 T           NK         pDCs
#> ATGCCAGAACGACT  0.33641919  0.386408152  0.334928712
#> CATGGCCTGTGCAT  0.45921267  0.179000538  0.361782653
#> GAACCTGATGAACC  0.52673269  0.433694559  0.452660978
#> TGACTGGATTCTCA  0.50523340  0.597678571  0.452660978
#> AGTCAGACTGCACA  0.53059403  0.373492002  0.474302273
#> TCTGATACACGTGT  0.57901926  0.399511640  0.323248106
#> TGGTATCTAAACAG  0.59542503  0.336851675  0.302323727
#> GCAGCTCTGTTTCT  0.75441286  0.283351703  0.230720739
#> GATATAACACGCAT  0.57591865  0.384609263  0.279228222
#> AATGTTGACAGTCA  0.46850324  0.328285477  0.422442344
#> AGGTCATGAGTGTC  0.26538383  0.123485416  0.346337256
#> AGAGATGATCTCGC  0.14325016  0.190466659  0.389368162
#> GGGTAACTCTAGTG  0.25878283 -0.065311007  0.087410574
#> CATGAGACACGGGA  0.25117157  0.149973423  0.201529934
#> TACGCCACTCCGAA  0.21713222  0.100483231  0.392515557
#> CTAAACCTGTGCAT  0.01584158  0.069045900  0.075263009
#> GTAAGCACTCATTC  0.14379933  0.180397532  0.522957788
#> TTGGTACTGAATCC -0.09767880 -0.257592458  0.371937743
#> CATCATACGGAGCA  0.14379933 -0.038965867  0.068810235
#> TACATCACGCTAAC  0.23276650 -0.081240698  0.318235502
#> TTACCATGAATCGC  0.39410228  0.346843988  0.435194140
#> ATAGGAGAAACAGA  0.27548924  0.435840437  0.349121404
#> GCGCACGACTTTAC  0.14605220  0.258607271  0.219648843
#> ACTCGCACGAAAGT  0.38447399  0.347806442  0.428796414
#> ATTACCTGCCTTAT  0.38495186  0.570925283  0.625000797
#> CCCAACTGCAATCG  0.14605220  0.258607271  0.219648843
#> AAATTCGAATCACG  0.14605220  0.258607271  0.219648843
#> CCATCCGATTCGCC          NA           NA           NA
#> TCCACTCTGAGCTT  0.14605220  0.258607271  0.219648843
#> CATCAGGATGCACA  0.53515513  0.566703407  0.427628956
#> CTAAACCTCTGACA  0.54346545  0.575054633  0.409528615
#> GATAGAGAAGGGTG -0.22951059  0.298393005 -0.319489226
#> CTAACGGAACCGAT  0.18157841  0.621059003  0.005468483
#> AGATATACCCGTAA  0.37818936  0.677021525  0.224726362
#> TACTCTGAATCGAC  0.27851659  0.219363399 -0.105750467
#> GCGCATCTTGCTCC  0.27123979  0.377964473  0.059904230
#> GTTGACGATATCGG  0.05581590  0.524906982 -0.074056181
#> ACAGGTACTGGTGT  0.27851659  0.219363399 -0.105750467
#> GGCATATGCTTATC  0.04843767  0.493567647 -0.174560702
#> CATTACACCAACTG  0.21584973  0.669885175  0.008859874
#> TAGGGACTGAACTC  0.51229414  0.634073742  0.294367840
#> GCTCCATGAGAAGT  0.59854888  0.376401015  0.353448338
#> TACAATGATGCTAG  0.58790771  0.387405226  0.358489441
#> CTTCATGACCGAAT  0.54346545  0.575054633  0.409528615
#> CTGCCAACAGGAGC  0.14379933  0.372340506  0.068810235
#> TTGCATTGAGCTAC  0.45473340  0.404651319  0.464207083
#> AAGCAAGAGCTTAG  0.36479619  0.520988072  0.224538663
#> CGGCACGAACTCAG  0.65936343  0.563621480  0.304635898
#> GGTGGAGATTACTC  0.47189001  0.564005213  0.425683589
#> GGCCGATGTACTCT  0.71009286  0.374568348  0.236610929
#> CGTAGCCTGTATGC  0.39410228  0.346843988  0.435194140
#> TGAGCTGAATGCTG  0.43497077  0.536433658  0.531974279
#> CCTATAACGAGACG          NA           NA           NA
#> ATAAGTTGGTACGT          NA           NA           NA
#> AAGCGACTTTGACG  0.51536452  0.231229325  0.174077656
#> ACCAGTGAATACCG  0.45966003  0.495152305  0.516467838
#> ATTGCACTTGCTTT  0.48378583  0.518154490  0.401022086
#> CTAGGTGATGGTTG  0.45921267  0.503136646  0.514751157
#> GCACTAGACCTTTA  0.48251605  0.491520381  0.531050459
#> CATGCGCTAGTCAC  0.41237489  0.551514122  0.404386952
#> TTGAGGACTACGCA  0.30564198  0.303549205  0.158442359
#> ATACCACTCTAAGC  0.44804843  0.408420012  0.459217781
#> CATATAGACTAAGC  0.16555607  0.303549205  0.304696843
#> TTTAGCTGTACTCT  0.48251605  0.491520381  0.531050459
#> GACATTCTCCACCT  0.27284004  0.433554985  0.348155312
#> ACGTGATGCCATGA  0.14605220  0.258607271  0.219648843
#> ATTGTAGATTCCCG  0.26943453  0.430067716  0.346224131
#> GATAGAGATCACGA  0.10539166  0.226389929  0.116053361
#> AATGCGTGGACGGA  0.46727870  0.502416153  0.520113493
#> GCGTAAACACGGTT  0.57462205  0.632891346  0.522920957
#> ATTCAGCTCATTGG -0.21632805 -0.106602648  0.173070063
#> GGCATATGGGGAGT -0.17734534 -0.038563427  0.253098776
#> ATCATCTGACACCA -0.20021035 -0.071422788  0.203129651
#> GTCATACTTCGCCT -0.19894107 -0.050085524  0.250443404
#> TTACGTACGTTCAG -0.15808276 -0.005565058  0.291873633
#> GAGTTGTGGTAGCT -0.38474128 -0.007410558  0.037606131
#> GACGCTCTCTCTCG -0.09742476  0.083005736  0.378905609
#> AGTCTTACTTCGGA -0.31373347 -0.165096728  0.062378097
#> GGAACACTTCAGAC -0.20827709 -0.031843074  0.213533698
#> CTTGATTGATCTTC -0.08849676  0.204178609  0.184361913