Compare scRNA-seq data to reference data.

Compare scRNA-seq data to reference data.

clustify(input, ...)

# S3 method for default
clustify(input, ref_mat, metadata = NULL,
  cluster_col = NULL, query_genes = NULL, per_cell = FALSE,
  n_perm = 0, compute_method = "spearman", verbose = FALSE,
  lookuptable = NULL, rm0 = FALSE, obj_out = FALSE,
  seurat_out = FALSE, rename_prefix = NULL, threshold = "auto", ...)

# S3 method for seurat
clustify(input, ref_mat, cluster_col = NULL,
  query_genes = NULL, per_cell = FALSE, n_perm = 0,
  compute_method = "spearman", use_var_genes = TRUE, dr = "umap",
  seurat_out = TRUE, obj_out = FALSE, threshold = "auto",
  verbose = FALSE, rm0 = FALSE, rename_prefix = NULL, ...)

# S3 method for Seurat
clustify(input, ref_mat, cluster_col = NULL,
  query_genes = NULL, per_cell = FALSE, n_perm = 0,
  compute_method = "spearman", use_var_genes = TRUE, dr = "umap",
  seurat_out = TRUE, obj_out = FALSE, threshold = "auto",
  verbose = FALSE, rm0 = FALSE, rename_prefix = NULL, ...)

# S3 method for SingleCellExperiment
clustify(input, ref_mat,
  cluster_col = NULL, query_genes = NULL, per_cell = FALSE,
  n_perm = 0, compute_method = "spearman", use_var_genes = TRUE,
  dr = "umap", seurat_out = TRUE, obj_out = TRUE,
  threshold = "auto", verbose = FALSE, rm0 = FALSE,
  rename_prefix = NULL, ...)

Arguments

input
single-cell expression matrix or Seurat object

...
additional arguments to pass to compute_method function

ref_mat
reference expression matrix

metadata
cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set. Not required if running correlation per cell.

cluster_col
column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes
A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

per_cell
if true run per cell, otherwise per cluster.

n_perm
number of permutations, set to 0 by default

compute_method
method(s) for computing similarity scores

verbose
whether to report certain variables chosen

lookuptable
if not supplied, will look in built-in table for object parsing

rm0
consider 0 as missing data, recommended for per_cell

obj_out
whether to output object instead of cor matrix

seurat_out
output cor matrix or called seurat object

rename_prefix
prefix to add to type and r column names

threshold
identity calling minimum correlation score threshold, only used when obj_out = T

use_var_genes
if providing a seurat object, use the variable genes (stored in seurat_object@var.genes) as the query_genes.

dr
stored dimension reduction

Value

single cell object with identity assigned in metadata, or matrix of correlation values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Annotate a matrix and metadata
clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = pbmc_bulk_matrix,
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  verbose = TRUE
)

#> using # of genes: 1688

#>              primary human B cells rep 1 primary human myeloid DC rep 1
#> B                             0.62161187                      0.5134249
#> CD14+ Mono                    0.38176061                      0.6585926
#> CD8 T                         0.43926595                      0.4648890
#> DC                            0.37326478                      0.5726146
#> FCGR3A+ Mono                  0.39707564                      0.6828512
#> Memory CD4 T                  0.47496679                      0.4835943
#> Naive CD4 T                   0.47847311                      0.4753289
#> NK                            0.41142231                      0.4600473
#> Platelet                      0.09448371                      0.2607297
#>              primary human monocytes rep 1 primary human neutrophils rep 1
#> B                                0.4858326                       0.4004996
#> CD14+ Mono                       0.6932047                       0.5548357
#> CD8 T                            0.4396090                       0.3759313
#> DC                               0.5619668                       0.3990601
#> FCGR3A+ Mono                     0.6464808                       0.5302247
#> Memory CD4 T                     0.4637761                       0.3883083
#> Naive CD4 T                      0.4615861                       0.3824808
#> NK                               0.4339593                       0.3824971
#> Platelet                         0.2607125                       0.2120760
#>              primary human NK cells rep 1 primary human PBMC rep 1
#> B                               0.4682206                0.4930223
#> CD14+ Mono                      0.4380704                0.5716609
#> CD8 T                           0.5988312                0.5230516
#> DC                              0.4031306                0.4926074
#> FCGR3A+ Mono                    0.4434886                0.5511974
#> Memory CD4 T                    0.5500075                0.5277524
#> Naive CD4 T                     0.5418622                0.5293089
#> NK                              0.6443927                0.5141864
#> Platelet                        0.1429789                0.2285756
#>              primary human T cells rep 1 primary human B cells rep 2
#> B                              0.4688555                   0.6409642
#> CD14+ Mono                     0.3525068                   0.4081323
#> CD8 T                          0.5533326                   0.4529453
#> DC                             0.3248138                   0.4027535
#> FCGR3A+ Mono                   0.3585348                   0.4240105
#> Memory CD4 T                   0.6017151                   0.4884443
#> Naive CD4 T                    0.6091448                   0.4899223
#> NK                             0.5180162                   0.4274134
#> Platelet                       0.0734803                   0.1147430
#>              primary human myeloid DC rep 2 primary human monocytes rep 2
#> B                                 0.5142264                     0.4817698
#> CD14+ Mono                        0.6585599                     0.7061875
#> CD8 T                             0.4657318                     0.4380312
#> DC                                0.5714870                     0.5624386
#> FCGR3A+ Mono                      0.6868239                     0.6588797
#> Memory CD4 T                      0.4841326                     0.4637714
#> Naive CD4 T                       0.4743983                     0.4603821
#> NK                                0.4652355                     0.4398213
#> Platelet                          0.2626675                     0.2685089
#>              primary human neutrophils rep 2 primary human NK cells rep 2
#> B                                  0.3882407                    0.4719677
#> CD14+ Mono                         0.5573050                    0.4408221
#> CD8 T                              0.3636086                    0.6007341
#> DC                                 0.3955419                    0.4074878
#> FCGR3A+ Mono                       0.5322509                    0.4450197
#> Memory CD4 T                       0.3760267                    0.5502540
#> Naive CD4 T                        0.3692816                    0.5415756
#> NK                                 0.3715250                    0.6474230
#> Platelet                           0.2165613                    0.1468565
#>              primary human PBMC rep 2 primary human T cells rep 2
#> B                           0.5179320                  0.48219201
#> CD14+ Mono                  0.6026460                  0.36879949
#> CD8 T                       0.5235911                  0.57153550
#> DC                          0.5077886                  0.33963841
#> FCGR3A+ Mono                0.5802021                  0.37464917
#> Memory CD4 T                0.5471137                  0.61932481
#> Naive CD4 T                 0.5498952                  0.62715073
#> NK                          0.5070531                  0.53661318
#> Platelet                    0.2523581                  0.08479795

# Annotate using a different method
res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  ref_mat = pbmc_bulk_matrix,
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  compute_method = "cosine"
)

# Annotate a Seurat object
clustify(
  s_small,
  pbmc_bulk_matrix,
  cluster_col = "res.1",
  seurat_out = TRUE,
  per_cell = FALSE,
  dr = "tsne"
)

#> using threshold of 0.5

#> seurat not loaded, returning cor_mat instead

#>   primary human B cells rep 1 primary human myeloid DC rep 1
#> 0                  0.56495413                      0.2905662
#> 1                  0.38339099                      0.6645159
#> 2                  0.01569994                      0.3132044
#> 3                 -0.42255352                     -0.1581610
#>   primary human monocytes rep 1 primary human neutrophils rep 1
#> 0                     0.2860993                      0.31382849
#> 1                     0.5913028                      0.54143473
#> 2                     0.2506601                      0.26686921
#> 3                    -0.1596368                     -0.04526416
#>   primary human NK cells rep 1 primary human PBMC rep 1
#> 0                    0.2521516                0.4810757
#> 1                    0.3960676                0.4553994
#> 2                    0.5476851                0.3873399
#> 3                   -0.2740145               -0.2484332
#>   primary human T cells rep 1 primary human B cells rep 2
#> 0                   0.5081027                   0.5427781
#> 1                   0.4296189                   0.3996944
#> 2                   0.3657692                   0.0329662
#> 3                  -0.3359033                  -0.3573163
#>   primary human myeloid DC rep 2 primary human monocytes rep 2
#> 0                      0.2512582                    0.21316182
#> 1                      0.6595903                    0.57611050
#> 2                      0.2938857                    0.22226395
#> 3                     -0.1374992                   -0.09252722
#>   primary human neutrophils rep 2 primary human NK cells rep 2
#> 0                      0.29267862                    0.3189304
#> 1                      0.55990915                    0.4737587
#> 2                      0.21612700                    0.5310227
#> 3                     -0.03016533                   -0.1999765
#>   primary human PBMC rep 2 primary human T cells rep 2
#> 0                0.4634318                   0.5432248
#> 1                0.4997303                   0.4406714
#> 2                0.2883316                   0.3448771
#> 3               -0.1891536                  -0.2745370

# Annotate (and return) a Seurat object per-cell
res <- clustify(
  input = s_small,
  ref_mat = pbmc_bulk_matrix,
  cluster_col = "res.1",
  seurat_out = TRUE,
  per_cell = TRUE,
  dr = "tsne"
)

#> Warning: the standard deviation is zero

#> using threshold of NA

#> seurat not loaded, returning cor_mat instead