Combined function to compare scRNA-seq data to bulk RNA-seq data and marker list

Combined function to compare scRNA-seq data to bulk RNA-seq data and marker list

clustify_nudge(input, ...)

# S3 method for default
clustify_nudge(
  input,
  ref_mat,
  marker,
  metadata = NULL,
  cluster_col = NULL,
  query_genes = NULL,
  compute_method = "spearman",
  weight = 1,
  seurat_out = FALSE,
  threshold = -Inf,
  dr = "umap",
  norm = "diff",
  call = TRUE,
  marker_inmatrix = TRUE,
  mode = "rank",
  obj_out = FALSE,
  rename_prefix = NULL,
  lookuptable = NULL,
  ...
)

# S3 method for seurat
clustify_nudge(
  input,
  ref_mat,
  marker,
  cluster_col = NULL,
  query_genes = NULL,
  compute_method = "spearman",
  weight = 1,
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = -Inf,
  dr = "umap",
  norm = "diff",
  marker_inmatrix = TRUE,
  mode = "rank",
  rename_prefix = NULL,
  ...
)

# S3 method for Seurat
clustify_nudge(
  input,
  ref_mat,
  marker,
  cluster_col = NULL,
  query_genes = NULL,
  compute_method = "spearman",
  weight = 1,
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = -Inf,
  dr = "umap",
  norm = "diff",
  marker_inmatrix = TRUE,
  mode = "rank",
  rename_prefix = NULL,
  ...
)

Arguments

input
express matrix or object

...
passed to matrixize_markers

ref_mat
reference expression matrix

marker
matrix of markers

metadata
cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set.

cluster_col
column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes
A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

compute_method
method(s) for computing similarity scores

weight
relative weight for the gene list scores, when added to correlation score

seurat_out
output cor matrix or called seurat object

threshold
identity calling minimum score threshold, only used when obj_out = T

dr
stored dimension reduction

norm
whether and how the results are normalized

call
make call or just return score matrix

marker_inmatrix
whether markers genes are already in preprocessed matrix form

mode
use marker expression pct or ranked cor score for nudging

obj_out
whether to output object instead of cor matrix

rename_prefix
prefix to add to type and r column names

lookuptable
if not supplied, will look in built-in table for object parsing

Value

single cell object, or matrix of numeric values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Seurat2
clustify_nudge(
  input = s_small,
  ref_mat = cbmc_ref,
  marker = cbmc_m,
  cluster_col = "res.1",
  threshold = 0.8,
  seurat_out = FALSE,
  mode = "pct",
  dr = "tsne"
)

#>              B   CD14+ Mono CD16+ Mono      CD34+       CD4 T      CD8 T
#> 0  0.522410963 -0.005572036  0.4119037  0.4172004  0.61062598  0.4431346
#> 1  0.396432770  0.725146450  0.6939952  0.4641469  0.42401600  0.1737119
#> 2  0.007020943  0.211394209  0.6094641  0.3178432  0.46805261  0.4812869
#> 3 -0.400393645 -0.243071872  0.1859978 -0.0854630 -0.08048411 -0.3731234
#>            DC     Eryth Memory CD4 T         Mk Naive CD4 T         NK
#> 0  0.36476066 0.1523530    0.7047302 -0.9967949   0.6398803 -0.2617596
#> 1  0.71146653 0.4359801    0.5349202 -0.7964748   0.4304222 -0.1248677
#> 2  0.67077257 0.4137920    0.6331816 -0.7115560   0.5864320  0.8099383
#> 3 -0.05333518 0.4352437    0.1083785  0.7453407  -0.1097800 -0.8598142
#>        pDCs
#> 0 0.3456669
#> 1 0.5116362
#> 2 0.2859966
#> 3 0.1901284

# Seurat3
 clustify_nudge(
  input = s_small3,
  ref_mat = cbmc_ref,
  marker = cbmc_m,
  cluster_col = "RNA_snn_res.1",
  threshold = 0.8,
  seurat_out = FALSE,
  mode = "pct",
  dr = "tsne"
)

#>            B CD14+ Mono CD16+ Mono      CD34+     CD4 T     CD8 T         DC
#> 0 -0.8386005 -0.8022027 -0.1290444 -0.1811809 0.3775856 0.4804364 -0.1773445
#> 1  0.1116745  0.7647078  0.7930478  0.3569654 0.4509427 0.2469878  0.6437572
#> 2  0.6997539 -0.2133902  0.5339625  0.3773183 0.2712622 0.1166450  0.6264979
#>        Eryth Memory CD4 T         Mk Naive CD4 T        NK      pDCs
#> 0  0.5092140    0.2683011  0.5851824  0.03142658 0.1767582 0.1596988
#> 1  0.2455553    0.5380182 -0.1502676  0.52633608 0.2681751 0.4804654
#> 2 -0.1592523    0.1233831 -0.5265590  0.31191120 0.1667222 0.4075066

# Matrix
clustify_nudge(
  input = pbmc_matrix_small,
  ref_mat = cbmc_ref,
  metadata = pbmc_meta,
  marker = as.matrix(cbmc_m),
  query_genes = pbmc_vargenes,
  cluster_col = "classified",
  threshold = 0.8,
  call = FALSE,
  marker_inmatrix = FALSE,
  mode = "pct"
)

#>                        B  CD14+ Mono CD16+ Mono     CD34+      CD4 T      CD8 T
#> B             0.89129128 -0.09482820 0.36430264 0.7317217 0.53933823 0.57819315
#> CD14+ Mono   -0.14739603  0.90814717 0.51856965 0.6710434 0.41250211 0.46484177
#> CD8 T        -0.11323147 -0.11229570 0.34225996 0.7136651 0.72384293 0.66175171
#> DC           -0.09890316  0.43792394 0.44916939 0.6814873 0.48842105 0.45893484
#> FCGR3A+ Mono -0.12660135  0.64538033 0.90590926 0.6726128 0.41690109 0.54635896
#> Memory CD4 T -0.06044487 -0.02870830 0.36003221 0.7427243 0.89489774 0.69324349
#> Naive CD4 T  -0.06874533 -0.07764912 0.35150020 0.7548765 0.82685564 0.74596560
#> NK           -0.13741949 -0.12259625 0.33838687 0.7067672 0.59374826 0.79378254
#> Platelet     -0.50747063 -0.25075312 0.02147571 0.3271915 0.08603668 0.08622349
#>                      DC     Eryth Memory CD4 T         Mk Naive CD4 T
#> B            0.47539346 0.6300171    0.5959394 -0.3538824   0.5953671
#> CD14+ Mono   0.56732763 0.6055132    0.5708809 -0.2338858   0.5384300
#> CD8 T        0.43400786 0.6525508    0.6542797 -0.3276053   0.6121911
#> DC           0.83270918 0.6049950    0.5657932 -0.2778113   0.5260392
#> FCGR3A+ Mono 0.54084520 0.6097144    0.5710313 -0.2341562   0.5300822
#> Memory CD4 T 0.44862102 0.6643080    0.6860640 -0.3457256   0.6554955
#> Naive CD4 T  0.44731357 0.6677468    0.6913118 -0.3406683   0.6622109
#> NK           0.42414982 0.6420863    0.6259194 -0.3341079   0.5880876
#> Platelet     0.09763352 0.3599792    0.2945712  0.6304742   0.2899873
#>                        NK      pDCs
#> B            -0.154620036 0.5866840
#> CD14+ Mono   -0.096549233 0.5377967
#> CD8 T         0.598665387 0.5300752
#> DC           -0.011303596 0.7012226
#> FCGR3A+ Mono -0.039743952 0.5332615
#> Memory CD4 T -0.002563329 0.5406439
#> Naive CD4 T  -0.069020452 0.5415440
#> NK            0.896029075 0.5264843
#> Platelet     -0.534417903 0.2308274