Combined function to compare scRNA-seq data to bulk RNA-seq data and marker list

Combined function to compare scRNA-seq data to bulk RNA-seq data and marker list

clustify_nudge(input, ...)

# S3 method for default
clustify_nudge(
  input,
  ref_mat,
  marker,
  metadata = NULL,
  cluster_col = NULL,
  query_genes = NULL,
  compute_method = "spearman",
  weight = 1,
  seurat_out = FALSE,
  threshold = -Inf,
  dr = "umap",
  norm = "diff",
  call = TRUE,
  marker_inmatrix = TRUE,
  mode = "rank",
  obj_out = FALSE,
  rename_prefix = NULL,
  lookuptable = NULL,
  ...
)

# S3 method for Seurat
clustify_nudge(
  input,
  ref_mat,
  marker,
  cluster_col = NULL,
  query_genes = NULL,
  compute_method = "spearman",
  weight = 1,
  seurat_out = TRUE,
  obj_out = FALSE,
  threshold = -Inf,
  dr = "umap",
  norm = "diff",
  marker_inmatrix = TRUE,
  mode = "rank",
  rename_prefix = NULL,
  ...
)

Arguments

input
express matrix or object

...
passed to matrixize_markers

ref_mat
reference expression matrix

marker
matrix of markers

metadata
cell cluster assignments, supplied as a vector or data.frame. If data.frame is supplied then cluster_col needs to be set.

cluster_col
column in metadata that contains cluster ids per cell. Will default to first column of metadata if not supplied. Not required if running correlation per cell.

query_genes
A vector of genes of interest to compare. If NULL, then common genes between the expr_mat and ref_mat will be used for comparision.

compute_method
method(s) for computing similarity scores

weight
relative weight for the gene list scores, when added to correlation score

seurat_out
output cor matrix or called seurat object

threshold
identity calling minimum score threshold, only used when obj_out = T

dr
stored dimension reduction

norm
whether and how the results are normalized

call
make call or just return score matrix

marker_inmatrix
whether markers genes are already in preprocessed matrix form

mode
use marker expression pct or ranked cor score for nudging

obj_out
whether to output object instead of cor matrix

rename_prefix
prefix to add to type and r column names

lookuptable
if not supplied, will look in built-in table for object parsing

Value

single cell object, or matrix of numeric values, clusters from input as row names, cell types from ref_mat as column names

Examples

# Seurat3
clustify_nudge(
    input = s_small3,
    ref_mat = cbmc_ref,
    marker = cbmc_m,
    cluster_col = "RNA_snn_res.1",
    threshold = 0.8,
    seurat_out = FALSE,
    mode = "pct",
    dr = "tsne"
)
#> object data retrieval complete, moving to similarity computation
#> using # of genes: 19
#> similarity computation completed, matrix of 3 x 13, preparing output
#>             B CD14+ Mono CD16+ Mono      CD34+     CD4 T      CD8 T         DC
#> 0 -0.94493203 -0.9156671 -0.2220272 -0.2825193 0.2887232 0.37971905 -0.3168822
#> 1  0.04838758  0.7219124  0.7316084  0.3238704 0.3472936 0.19775419  0.5865428
#> 2  0.62480323 -0.2550175  0.4973646  0.3895667 0.1769981 0.01448416  0.5146717
#>        Eryth Memory CD4 T         Mk Naive CD4 T         NK      pDCs
#> 0  0.4203598    0.1987308  0.6204476 -0.08992247 0.08688022 0.1134067
#> 1  0.1736468    0.5311026 -0.2163932  0.51050337 0.25334722 0.4370882
#> 2 -0.1716881    0.1094072 -0.6697973  0.23853307 0.14695522 0.3758960

# Matrix
clustify_nudge(
    input = pbmc_matrix_small,
    ref_mat = cbmc_ref,
    metadata = pbmc_meta,
    marker = as.matrix(cbmc_m),
    query_genes = pbmc_vargenes,
    cluster_col = "classified",
    threshold = 0.8,
    call = FALSE,
    marker_inmatrix = FALSE,
    mode = "pct"
)
#> using # of genes: 599
#> similarity computation completed, matrix of 9 x 13, preparing output
#>                       B CD14+ Mono  CD16+ Mono     CD34+       CD4 T
#> B             0.9093577 -0.2452939  0.22341215 0.6248639  0.39898892
#> CD14+ Mono   -0.3184633  0.9145428  0.52579175 0.5103056  0.20913853
#> CD8 T        -0.2866333 -0.2920874  0.17944732 0.5663773  0.64950686
#> DC           -0.1927865  0.3971091  0.41292804 0.5694074  0.31768980
#> FCGR3A+ Mono -0.2924419  0.6387074  0.92949138 0.5331927  0.22686950
#> Memory CD4 T -0.2094668 -0.1734003  0.22245854 0.6265186  0.86108263
#> Naive CD4 T  -0.2115884 -0.2355533  0.19645924 0.6480900  0.80013677
#> NK           -0.3523479 -0.3353237  0.15173936 0.5350808  0.47524944
#> Platelet     -0.6782452 -0.3759687 -0.08362827 0.2455947 -0.05758481
#>                    CD8 T          DC     Eryth Memory CD4 T         Mk
#> B             0.43921857  0.35367232 0.4817829    0.4803367 -0.6748463
#> CD14+ Mono    0.25084659  0.51955311 0.4559350    0.4553454 -0.4660850
#> CD8 T         0.59418188  0.26878200 0.5007841    0.6323214 -0.5987391
#> DC            0.28543102  0.84933995 0.4470039    0.4764929 -0.5578229
#> FCGR3A+ Mono  0.35105547  0.48594006 0.4604308    0.4614474 -0.4760099
#> Memory CD4 T  0.65776299  0.30625770 0.5395170    0.7128879 -0.6336165
#> Naive CD4 T   0.72313292  0.29335212 0.5528441    0.7277404 -0.6269043
#> NK            0.69267415  0.23172856 0.4578726    0.5665250 -0.6199834
#> Platelet     -0.07272109 -0.06019224 0.3068445    0.1457308  0.7319575
#>              Naive CD4 T         NK      pDCs
#> B              0.5407149 -0.3536418 0.4801131
#> CD14+ Mono     0.4564707 -0.2621333 0.3710598
#> CD8 T          0.6036230  0.5582283 0.3519523
#> DC             0.4475783 -0.1659100 0.6825388
#> FCGR3A+ Mono   0.4574654 -0.1921798 0.3755701
#> Memory CD4 T   0.6871755 -0.1208282 0.3668161
#> Naive CD4 T    0.6967073 -0.1890644 0.3594580
#> NK             0.5505877  0.8940904 0.3186506
#> Platelet       0.1609963 -0.7349565 0.1110947