clustifyr: Classifier for Single-cell RNA-seq Using Cell Clusters

Package designed to aid in classifying cells from single-cell RNA sequencing data using external reference data (e.g., bulk RNA-seq, scRNA-seq, microarray, gene lists). A variety of correlation based methods and gene list enrichment methods are provided to assist cell type assignment.

See also

Author

Maintainer: Kent Riemondy kent.riemondy@cuanschutz.edu

Authors:

Rui Fu raysinensis@gmail.com

Other contributors:

Austin Gillen austin.gillen@ucdenver.edu [contributor]

Chengzhe Tian Chengzhe.Tian@colorado.edu [contributor]

Jay Hesselberth jay.hesselberth@gmail.com [contributor]

Yue Hao haoyuethink@gmail.com [contributor]

Michelle Daya michelle.daya@ucdenver.edu [contributor]

Sidhant Puntambekar sidhantnp@yahoo.com [contributor]

RNA Bioscience Initiative [funder, copyright holder]