get ranked calls for each cluster

get ranked calls for each cluster

cor_to_call_rank(
  cor_mat,
  metadata = NULL,
  cluster_col = "cluster",
  collapse_to_cluster = FALSE,
  threshold = 0,
  rename_prefix = NULL,
  top_n = NULL
)

Arguments

cor_mat
input similarity matrix

metadata
input metadata with tsne or umap coordinates and cluster ids

cluster_col
metadata column, can be cluster or cellid

collapse_to_cluster
if a column name is provided, takes the most frequent call of entire cluster to color in plot

threshold
minimum correlation coefficent cutoff for calling clusters

rename_prefix
prefix to add to type and r column names

top_n
the number of ranks to keep, the rest will be set to 100

Value

dataframe of cluster, new ident, and r info

Examples

res <- clustify(
  input = pbmc_matrix_small,
  metadata = pbmc_meta,
  cluster_col = "classified",
  ref_mat = cbmc_ref
)

#> var.features not found, using all genes instead

res2 <- cor_to_call_rank(res, threshold = "auto")

#> using threshold of 0.69