Returns a list of variable genes based on PCA

Extract genes, i.e. "features", based on the top loadings of principal components formed from the bulk expression data set

Usage

feature_select_PCA(
  mat = NULL,
  pcs = NULL,
  n_pcs = 10,
  percentile = 0.99,
  if_log = TRUE
)

Arguments

mat

Expression matrix. Rownames are genes, colnames are single cell cluster name, and values are average single cell expression (log transformed).

pcs

Precalculated pcs if available, will skip over processing on mat.

n_pcs

Number of PCs to selected gene loadings from. See the explore_PCA_corr.Rmd vignette for details.

percentile

Select the percentile of absolute values of PCA loadings to select genes from. E.g. 0.999 would select the top point 1 percent of genes with the largest loadings.

if_log

whether the data is already log transformed

Value

vector of genes

Examples

feature_select_PCA(
    cbmc_ref,
    if_log = FALSE
)
#>   [1] "HLA-DRA"       "CST3"          "HLA-DPB1"      "HLA-DRB1"     
#>   [5] "HLA-DPA1"      "TYROBP"        "IL32"          "HLA-DQB1"     
#>   [9] "HLA-DRB5"      "IFI30"         "RP11-1143G9.4" "FCN1"         
#>  [13] "TRAC"          "SPI1"          "MS4A6A"        "LGALS2"       
#>  [17] "CD68"          "CSTA"          "CEBPD"         "CLEC12A"      
#>  [21] "HBG1"          "HBA2"          "KLRB1"         "IL32"         
#>  [25] "HBM"           "AHSP"          "CD7"           "IL7R"         
#>  [29] "TRBC2"         "CD3D"          "CA1"           "EVL"          
#>  [33] "IFITM1"        "GYPA"          "ALAS2"         "CD3G"         
#>  [37] "MT2A"          "FYB"           "LINC00861"     "GIMAP4"       
#>  [41] "HBG2"          "HBG1"          "HBA1"          "HLA-DPB1"     
#>  [45] "IGJ"           "HLA-DQB1"      "HLA-DQA1"      "ITM2C"        
#>  [49] "PLD4"          "FCER1A"        "ALOX5AP"       "IRF7"         
#>  [53] "IRF8"          "HBD"           "SERPINF1"      "MZB1"         
#>  [57] "LILRA4"        "CA2"           "SPIB"          "STOM"         
#>  [61] "PPBP"          "GZMB"          "PF4"           "CCL5"         
#>  [65] "PTGDS"         "GNG11"         "SDPR"          "HIST1H2AC"    
#>  [69] "CLIC3"         "TUBB1"         "PTCRA"         "ACRBP"        
#>  [73] "C12orf75"      "GZMH"          "GP9"           "MYL9"         
#>  [77] "CLEC1B"        "TREML1"        "CMTM5"         "MAP3K7CL"     
#>  [81] "IGHM"          "TCL1A"         "CD79A"         "FABP4"        
#>  [85] "RHOC"          "MS4A1"         "IGHD"          "TXN"          
#>  [89] "PRSS57"        "MPO"           "MATK"          "IGFBP7"       
#>  [93] "TXK"           "SERPINB6"      "NDUFS3"        "UQCRFS1"      
#>  [97] "RASSF1"        "STOML2"        "ILF2"          "MACF1"        
#> [101] "GZMB"          "GNLY"          "NKG7"          "GNG11"        
#> [105] "FCGR3A"        "HOPX"          "PRF1"          "KLRF1"        
#> [109] "GZMH"          "CMC1"          "SPON2"         "FABP4"        
#> [113] "KLRD1"         "NGFRAP1"       "CCL4"          "FGFBP2"       
#> [117] "CD160"         "KLRC1"         "EGFL7"         "TTC38"        
#> [121] "HBG2"          "CD7"           "IGHM"          "CD69"         
#> [125] "CLIC3"         "HOPX"          "CD79B"         "MS4A1"        
#> [129] "SPINK2"        "IGHD"          "CD36"          "YBX3"         
#> [133] "C1QTNF4"       "KIAA0125"      "VPREB3"        "AVP"          
#> [137] "BASP1"         "LAT2"          "PARP1"         "RCSD1"        
#> [141] "HBG1"          "IGJ"           "TCL1A"         "CEBPB"        
#> [145] "IGLC2"         "CD79A"         "CD79B"         "ID2"          
#> [149] "IGLC3"         "MS4A1"         "SPINK2"        "C1QTNF4"      
#> [153] "VPREB3"        "LINC00926"     "RP5-887A10.1"  "ANKRD28"      
#> [157] "AVP"           "FCRLA"         "IGLC7"         "ZFP36L1"      
#> [161] "KLRB1"         "IGHM"          "SERPINF1"      "MS4A1"        
#> [165] "TOB1"          "IGHD"          "TRDC"          "CD63"         
#> [169] "LEF1"          "BST2"          "S100B"         "CD8B"         
#> [173] "ACTN1"         "LAT2"          "OXNAD1"        "C1orf228"     
#> [177] "LDLRAP1"       "RP4-594I10.3"  "NELL2"         "PAG1"         
#> [181] "FCGR3A"        "FCER1A"        "SERPINA1"      "GPR183"       
#> [185] "CLEC10A"       "TOB1"          "RP11-290F20.3" "CD1C"         
#> [189] "C5AR1"         "PPP1R14A"      "RGS1"          "BASP1"        
#> [193] "C1QA"          "CDKN1C"        "ENHO"          "APOBEC3A"     
#> [197] "ITGB7"         "CD1E"          "NDRG2"         "FCGR2B"