Convert candidate genes list into matrix

Convert candidate genes list into matrix

matrixize_markers(
  marker_df,
  ranked = FALSE,
  n = NULL,
  step_weight = 1,
  background_weight = 0,
  unique = FALSE,
  metadata = NULL,
  cluster_col = "classified",
  remove_rp = FALSE
)

Arguments

marker_df
dataframe of candidate genes, must contain "gene" and "cluster" columns, or a matrix of gene names to convert to ranked

ranked
unranked gene list feeds into hyperp, the ranked gene list feeds into regular corr_coef

n
number of genes to use

step_weight
ranked genes are tranformed into pseudo expression by descending weight

background_weight
ranked genes are tranformed into pseudo expression with added weight

unique
whether to use only unique markers to 1 cluster

metadata
vector or dataframe of cluster names, should have column named cluster

cluster_col
column for cluster names to replace original cluster, if metadata is dataframe

remove_rp
do not include rps, rpl, rp1-9 in markers

Value

matrix of unranked gene marker names, or matrix of ranked expression

Examples

matrixize_markers(pbmc_markers)

#>            0        1        2         3        4             5        6
#> 1      RPS12     IL32   S100A9     CD79A     CCL5        CDKN1C     GZMB
#> 2      RPS27      LTB   S100A8     MS4A1     NKG7          HES4   FGFBP2
#> 3       RPS6     CD3D   LGALS2     CD79B     GZMA RP11-290F20.3    SPON2
#> 4      RPL32     IL7R     FCN1 LINC00926     CST7         MS4A7     PRF1
#> 5      RPS14     LDHB     CD14     TCL1A     GZMK        FCGR3A     GNLY
#> 6      RPS25      CD2   TYROBP  HLA-DQA1     CTSW           CKB   AKR1C3
#> 7       RPL9     AQP3   MS4A6A    VPREB3     CD8A        LILRA3     XCL2
#> 8      RPL13     CD3E     CST3  HLA-DQB1    KLRG1        IFITM3    CLIC3
#> 9      RPL31     TPT1      LYZ      CD74     GZMH        MS4A4A     CST7
#> 10      RPS3    TRAT1     TYMP   HLA-DRA      B2M         PILRA    KLRD1
#> 11      RPL3    RPS12      CFD     FCER2     LYAR        LRRC25     GZMA
#> 12     RPL21   SPOCK2     LST1  HLA-DPB1     IL32        LILRB1     NKG7
#> 13     RPS3A   CD40LG   LGALS1     BANK1    HLA-C          RHOC     CCL4
#> 14     RPL30   GIMAP7     AIF1  HLA-DRB1     PRF1      SIGLEC10    TTC38
#> 15      LDHB    TRADD    GSTP1  HLA-DPA1     GZMM         HMOX1   PRSS23
#> 16    RPS15A    RPS25      GRN   TSPAN13    HLA-A          LST1     XCL1
#> 17     RPLP2     RPSA     GPX1  HLA-DQA2     CD3D          CTSL     CTSW
#> 18    RPS27A    PRDX2      FTL     FCRLA  PTPRCAP           HCK   SH2D1B
#> 19    MALAT1   TTC39C   S100A6      CD37     CCL4         CSF1R   FCGR3A
#> 20    EEF1A1    RPLP0     FTH1  HLA-DRB5     HOPX      SERPINA1   IGFBP7
#> 21     RPS13     CD27   FCER1G     HVCN1     CD8B        FCER1G     GZMH
#> 22    RPL13A      AES     CTSS      PKIG     MATK          CD68    IL2RB
#> 23    RPL27A      JUN     CD68   HLA-DOB     HCST          AIF1    KLRF1
#> 24     RPS18   TMEM66    CEBPD       BLK    SAMD3           CFD    S1PR5
#> 25     RPL11   MYL12A   S100A4   HLA-DMA    HLA-B         IFI30    CD247
#> 26     RPS29    RPS29   LGALS3    PDLIM1     NCR3          SPI1    APMAP
#> 27     RPS28     JUNB    AP1S2      GNG7 APOBEC3G        LILRB2     HOPX
#> 28    RPL35A    ANXA1     NPC2      CD72   FGFBP2         BATF3    GPR56
#> 29    RPL23A ARHGAP15     OAZ1   HLA-DMB      LCK        LILRA5     MATK
#> 30     RPS23    HINT1      CDA     P2RX5   MALAT1        STXBP2     GZMM
#> 31      CCR7      EVL  S100A11      CD40 C12orf75          WARS  C1orf21
#> 32      TPT1      LCK     PSAP      EAF2     CD99        CXCL16  PLEKHF1
#> 33     RPS20      LAT     SAT1  PPAPDC1B     GZMB         TPPP3    HLA-C
#> 34      CD3D    NOSIP     NCF2      IRF8   MYL12A         C5AR1      B2M
#> 35     RPS4X     CD3G      CFP    RPL18A  RARRES3         COTL1    TBX21
#> 36      RPL5      VIM    CSF3R  CYB561A3      CD2         CEBPB    PTGDR
#> 37     RPS16    ITM2A    BLVRB     QRSL1    DUSP2          MAFB      CD7
#> 38     RPL19     OPTN    COTL1    MARCH1     CD3E          TESC   HAVCR2
#> 39    RPL10A      MAL SERPINA1     MEF2C   CHST12       SLC31A2    TPST2
#> 40      RPSA     PLP2    ASGR1     RPS23    HLA-E         TIMP1    HLA-A
#> 41      RPS5   FLT3LG   NUP214    SMIM14    APMAP        IFITM2    PLAC8
#> 42     RPS10     SIT1  GABARAP     RPS11 SLC9A3R1           BID    FCRL6
#> 43     RPL14  TNFAIP8    ALDH2      RPS5    CD247          ABI3    HLA-B
#> 44     RPS15     RGCC    FCGRT      LY86    ZAP70      C19orf38    ITGB2
#> 45     RPL18    GSTK1    IFI30   PLEKHF2    LITAF          SAT1     CTSC
#> 46      NPM1    FXYD5     CYBA      SNX2   MYL12B          BRI3     CD63
#> 47     RPLP0    CALM1     FPR1       LTB    TPST2          FTH1     NCR3
#> 48     RPL36  PPP2R5C    FOLR3     ISG20   MT-CYB         CAMK1      ID2
#> 49      CD3E  ZFP36L2   PYCARD      NCF1     PFN1          PSAP      UBB
#> 50      RPL4  ALOX5AP      TKT     CXCR4      UBC CTD-2006K23.1 C12orf75
#> 51     NOSIP   CORO1B     SPI1    RPL13A     CTSC          CTSS     SRGN
#> 52      LEF1    SUSD3     RAC1     SNHG7     DDX5           LYN     RHOC
#> 53    RPL36A   PBXIP1     TSPO    LAPTM5     GNLY       S100A11  RARRES3
#> 54    EEF1B2   GIMAP5   SLC7A7      RPS8      FYN           CFP   ARPC5L
#> 55     RPL22   IFITM1     RBP7     RPL15    ANXA1          CST3     PFN1
#> 56      RPL7    IL2RG     BRI3      RPL8   ARPC5L          TYMP    SAMD3
#> 57      RPS8    HSPA8  S100A12     FAIM3    YWHAQ        TCF7L2     CCL5
#> 58 PRKCQ-AS1    YWHAQ TNFSF13B     RPL18      ID2           FTL APOBEC3G
#> 59     RPL17  RPSAP58    IGSF6     RPS27    CALM1        PTP4A3     CD99
#> 60   PIK3IP1  TNFAIP3   IFITM3    RPL10A    IL2RG        TYROBP    RAP1B
#> 61      JUNB     CD69   RNF130   PTPRCAP    MYO1F       FAM110A   CX3CR1
#> 62    TMEM66   GIMAP4    GAPDH      STX7    SEPT7          FGL2   FCER1G
#> 63   GLTSCR2    CRIP1   TALDO1      CD52 C9orf142      APOBEC3A  PTPRCAP
#> 64     RPS26  ARL6IP5     VCAN    GNB2L1      LAT           FGR   CHST12
#> 65      BTG1    MZT2A     BST1     SYPL1     CD53          NPC2    EFHD2
#> 66       CD7    LIMS1     MNDA      RPSA     SUN2        FAM26F   MYL12A
#> 67     RPL38     LEF1    PLBD1     LIMD2     PLEK          OAZ1    HLA-E
#> 68     RPL35      FYB APOBEC3A     BIRC3     BIN2       CD300LF     HCST
#> 69     RPL24    SEPT1     CTSB     RPS21    HLA-F        SLC7A7     PLEK
#> 70     RPS21      TXN     CTSH     PRKCB    RAP1B        CPPED1    ZAP70
#> 71      IL7R    PEBP1     IFI6      SNX3     CD3G        UNC119   CD300A
#> 72       MAL   OCIAD2    NEAT1      EEF2    ANXA6        S100A4    CLIC1
#> 73     RPL37  PIK3IP1     CPVL    FAM26F HSP90AA1         NINJ1    ASCL2
#> 74      TCF7    TSTD1    ODF3B     POLD4     PIM1           CDA   IFITM1
#> 75   C6orf48     SOD1    AP2S1     PLAC8   TUBA4A        FCGR2A    PDIA3
#> 76      CD27 ILF3-AS1    TGFBI    EEF1B2   MT-ND2         HSBP1  ABHD17A
#> 77     RPL34     TCF7    SMCO4   RPL22L1    PTPRC         AP2A1  ALOX5AP
#> 78    RPS4Y1    SOCS3   LILRA5     RCSD1    ARL4C         PLAUR     CCL3
#> 79   RPSAP58  ANKRD12 TMEM176B     SEC62  PPP1R18        LGALS1     AOAH
#> 80  LEPROTL1   DNAJB1 SH3BGRL3     PNISR    DHRS7      TNFRSF1B   IFITM2
#> 81       AES   HNRNPF  HLA-DRA   BLOC1S2   IFITM1          CUX1  C5orf56
#> 82      SELL     KLF6     CD33     HMGN1  HSP90B1          EMR2     BST2
#> 83   TMEM123     PPA1    RAB32   TMEM243   STK17A        FAM49A     CFL1
#> 84      CD3G LEPROTL1   ARPC1B  RNASEH2B    HSPA5          ACTB  PPP1R18
#> 85      RGCC    PDCD5 C1orf162    POU2F2  PPP2R5C         ARPC3     RAC2
#> 86    GIMAP5     SKP1   TMSB10      CD69    PDIA3        GPBAR1     XBP1
#> 87       LCK    RWDD1    ANXA2     DRAM2      UBB          LYST     ABI3
#> 88    NDFIP1   SLC2A3    STX11      SELL     GUK1      MAPKAPK3   TYROBP
#> 89  C12orf57    PSMC5      IL8    EIF4A2     PPIB       SLC11A1    MYO1F
#> 90     TOMM7   NSMCE1     CAPG   TSC22D3   HNRNPK           ADA    ARL4C
#> 91    FLT3LG    DDIT4     RGS2   ZFP36L1     XBP1          NAAA     CD53
#> 92      RHOH    TRAM1    PLAUR  ATP6V1G1    CCND3        TMSB4X    CALM1
#> 93    DNAJB1    EWSR1    CEBPB     SP100     RPA2         CASP1    UBE2F
#> 94      GYPC      FUS      VIM      CD53     PPIA        PYCARD    LITAF
#> 95     TRAT1     NSA2     SRGN       DCK   RASSF5          YBX1    DHRS7
#>            7             8
#> 1     FCER1A           GP9
#> 2   SERPINF1        ITGA2B
#> 3    CLEC10A        TMEM40
#> 4       ENHO    AP001189.4
#> 5      CLIC2        LY6G6F
#> 6       CD1C         SEPT5
#> 7   HLA-DQB2         PTCRA
#> 8   CACNA2D3        TREML1
#> 9      UPK3A         CLDN5
#> 10   PHACTR1           HGD
#> 11  HLA-DQA2 RP11-879F14.2
#> 12     BASP1       C2orf88
#> 13      PLD4         SPARC
#> 14      LGMN           PF4
#> 15     NDRG2         CMTM5
#> 16      CPVL         GNG11
#> 17   CCDC88A         GP1BA
#> 18  HLA-DQA1         ITGB3
#> 19      CD33          SDPR
#> 20       GSN       SCGB1C1
#> 21   RPS6KA4        CLEC1B
#> 22     SULF2    AC147651.3
#> 23   HLA-DOA         PVALB
#> 24      CD1D         SCN1B
#> 25    RNASE6       ARHGAP6
#> 26      LMO2         ACRBP
#> 27  HLA-DQB1        FAXDC2
#> 28  HLA-DRB5          MYL9
#> 29  PPP1R14A           CLU
#> 30    CSF2RA      C19orf33
#> 31   HLA-DMA      PDZK1IP1
#> 32    SEMA4A         TUBA8
#> 33  HLA-DRB1          CTTN
#> 34     ALDH2          PPBP
#> 35     CD302          ESAM
#> 36  HLA-DPA1  RP11-367G6.3
#> 37      CST3          NRGN
#> 38  HLA-DPB1         TUBB1
#> 39     GNA15         F13A1
#> 40   SPATS2L         FSTL1
#> 41       SYK         RGS18
#> 42      IL1B        ALOX12
#> 43      KLF4           CA2
#> 44   HLA-DMB         PTGS1
#> 45     YIF1B           CD9
#> 46   HLA-DRA       C6orf25
#> 47   ZNF385A     HIST1H2AC
#> 48    LILRB4          SNCA
#> 49       GRN        GAS2L1
#> 50      CD74     HIST1H2BJ
#> 51    TRIM33           MMD
#> 52  C12orf45          PCP2
#> 53    HAVCR2       TSC22D1
#> 54    CCDC50          DMTN
#> 55     IGSF6        ANKRD9
#> 56      LAP3         ABCC3
#> 57      BLNK         RUFY1
#> 58      CTSH        PGRMC1
#> 59      CAPG          TPM1
#> 60    RNF130      MAP3K7CL
#> 61      ARF3         SENCR
#> 62       ID1          MPP1
#> 63     AP1S2       NGFRAP1
#> 64    SPINT2        TPTEP1
#> 65      RGS1        PDLIM1
#> 66   FAM105A          NEXN
#> 67       VCL          MLH3
#> 68    LGALS2          TPM4
#> 69   TNFAIP2         ACTN1
#> 70      TPM4         SCFD2
#> 71     MGST2           ILK
#> 72     H2AFY        FAM63A
#> 73      LY86           SNN
#> 74     FABP5         PARVB
#> 75      UBL7         GRAP2
#> 76     DDAH2        TMEM91
#> 77     TGFBI        MARCH2
#> 78     IFI30     HIST1H2BD
#> 79      GPX1           FAH
#> 80     FDFT1      PLA2G12A
#> 81      FUOM         PNMA1
#> 82    IGFLR1          ODC1
#> 83      LSP1         P2RX1
#> 84      MNDA           GSN
#> 85     PTPRE         NCOA4
#> 86     FCGRT         DAAM1
#> 87     GSTP1          CTSA
#> 88      SPI1        FERMT3
#> 89     TAF10          H1F0
#> 90    RAB22A        TUBA1C
#> 91 NME1-NME2      KIAA0513
#> 92       LYZ          MGLL
#> 93       CFP         CALM3
#> 94    SCAMP4         UBL4A
#> 95      CD86         CAPN1