Function to convert labelled object to avg expression matrix

Function to convert labelled object to avg expression matrix

object_ref(input, ...)

# S3 method for default
object_ref(
  input,
  cluster_col = NULL,
  var_genes_only = FALSE,
  assay_name = NULL,
  method = "mean",
  lookuptable = NULL,
  if_log = TRUE,
  ...
)

# S3 method for Seurat
object_ref(
  input,
  cluster_col = NULL,
  var_genes_only = FALSE,
  assay_name = NULL,
  method = "mean",
  lookuptable = NULL,
  if_log = TRUE,
  ...
)

# S3 method for seurat
object_ref(
  input,
  cluster_col = NULL,
  var_genes_only = FALSE,
  assay_name = NULL,
  method = "mean",
  lookuptable = NULL,
  if_log = TRUE,
  ...
)

# S3 method for SingleCellExperiment
object_ref(
  input,
  cluster_col = NULL,
  var_genes_only = FALSE,
  assay_name = NULL,
  method = "mean",
  lookuptable = NULL,
  if_log = TRUE,
  ...
)

Arguments

input
object after tsne or umap projections and clustering

...
additional arguments

cluster_col
column name where classified cluster names are stored in seurat meta data, cannot be "rn"

var_genes_only
whether to keep only var.genes in the final matrix output, could also look up genes used for PCA

assay_name
any additional assay data, such as ADT, to include. If more than 1, pass a vector of names

method
whether to take mean (default) or median

lookuptable
if not supplied, will look in built-in table for object parsing

if_log
input data is natural log, averaging will be done on unlogged data

Value

reference expression matrix, with genes as row names, and cell types as column names

Examples

object_ref(
    s_small3,
    cluster_col = "RNA_snn_res.1"
)

#>                       0         1         2
#> MS4A1         0.0000000 1.1260469 5.1510649
#> CD79B         2.4693409 2.9204074 5.0313160
#> CD79A         0.0000000 2.5351508 5.3756805
#> HLA-DRA       3.6403680 6.0084459 7.0553864
#> TCL1A         0.0000000 1.4958668 4.9633675
#> HLA-DQB1      1.6030685 3.8362900 5.1374216
#> HVCN1         0.0000000 1.8793600 4.3278881
#> HLA-DMB       0.4867611 3.0346761 4.3213139
#> LTB           5.5719249 3.0846795 5.5860803
#> LINC00926     0.0000000 0.0000000 3.4639360
#> FCER2         0.0000000 0.0000000 3.2406102
#> SP100         2.8094328 1.9706145 4.0812464
#> NCF1          0.0000000 2.0574919 3.7768261
#> PPP3CC        2.4699359 0.0000000 3.4057277
#> EAF2          0.0000000 1.2578967 3.7232671
#> PPAPDC1B      0.0000000 0.6236904 3.5008153
#> CD19          0.0000000 0.0000000 3.5319470
#> KIAA0125      0.0000000 0.0000000 3.3125954
#> CYB561A3      0.0000000 0.0000000 3.3523912
#> CD180         0.0000000 0.0000000 3.1935946
#> RP11-693J15.5 0.0000000 0.6236904 3.4557536
#> FAM96A        2.4991563 1.6970316 3.5812560
#> CXCR4         4.5984867 2.9386397 5.1914542
#> STX10         2.6030281 1.9213321 3.8770784
#> SNHG7         2.5883236 1.4600407 3.7651494
#> NT5C          1.6349875 1.5284050 4.1948110
#> BANK1         1.5675876 0.7071870 3.1933573
#> IGLL5         0.0000000 0.0000000 5.1779746
#> CD200         0.0000000 0.0000000 3.1657922
#> FCRLA         0.0000000 1.0048882 3.2310352
#> CD3D          5.0406142 0.6226876 4.8522894
#> NOSIP         4.3078382 2.3120289 2.1555293
#> SAFB2         2.8645701 0.0000000 0.0000000
#> CD2           3.7654580 0.4473585 2.4301393
#> IL7R          4.9599790 2.2971940 3.3514195
#> PIK3IP1       4.0010004 1.1105669 2.6513010
#> MPHOSPH6      3.2731282 1.1943493 1.5657879
#> KHDRBS1       4.2916181 1.5710342 1.2997552
#> MAL           2.8366930 0.7098036 0.0000000
#> CCR7          3.4508524 0.0000000 1.7627743
#> THYN1         3.5605783 1.0791229 1.6627255
#> TAF7          3.9279682 2.3779859 2.9831085
#> LDHB          4.8265044 3.1555651 4.2613947
#> TMEM123       4.3108107 2.7595383 3.8997968
#> CCDC104       2.5802723 0.0000000 1.6627255
#> EPC1          3.2862465 1.0710579 2.7795258
#> EIF4A2        4.8610986 2.9792186 4.2968188
#> CD3E          4.5564124 0.8511256 3.5851637
#> TMUB1         2.9399784 1.6693303 0.0000000
#> BLOC1S4       3.1292930 1.5102194 0.0000000
#> ACSM3         2.5733641 0.0000000 0.0000000
#> TMEM204       2.4513547 0.0000000 0.0000000
#> SRSF7         5.2073487 3.5966346 3.6690405
#> ACAP1         3.8480931 0.0000000 1.4722778
#> TNFAIP8       3.7427554 2.0034328 3.2266266
#> CD7           4.5355550 0.0000000 2.7376000
#> TAGAP         3.3378346 1.8810018 2.9432233
#> DNAJB1        3.7233133 2.2667350 2.1486048
#> ASNSD1        2.9052521 0.9994320 0.0000000
#> S1PR4         4.2951141 1.7979555 2.9518075
#> CTSW          4.4740282 0.7071870 3.9647706
#> GZMK          3.3682403 0.0000000 2.7453500
#> NKG7          6.2420120 3.1384237 5.5586985
#> IL32          5.1842108 1.5577485 4.0940805
#> DNAJC2        2.7028669 1.4087208 2.6316102
#> LYAR          3.8720115 1.8862886 5.2466439
#> CST7          4.6412651 1.0669347 3.8032034
#> LCK           4.2022629 0.5860144 3.1393764
#> CCL5          6.0106843 2.4880438 5.2814928
#> HNRNPH1       2.7089064 0.0000000 2.4100626
#> SSR2          4.4929659 3.2898973 3.9903973
#> DLGAP1-AS1    1.9384709 0.0000000 1.5271215
#> GIMAP1        4.2516757 2.5636774 1.8015084
#> MMADHC        2.8931953 1.7834336 3.3828356
#> ZNF76         1.9702458 0.6857306 1.5271215
#> CD8A          2.9991018 0.0000000 3.0196988
#> PTPN22        2.3913659 0.7457554 2.0692999
#> GYPC          4.2992024 0.7457554 3.2527556
#> HNRNPF        4.0475485 2.2347297 3.6580028
#> RPL7L1        2.6072371 1.4608337 2.7553779
#> KLRG1         3.5659317 0.0000000 1.5657879
#> CRBN          3.1398594 1.5836832 3.3746219
#> SATB1         3.8218037 0.0000000 1.6965705
#> SIT1          2.3367048 0.0000000 3.1448126
#> PMPCB         2.1622737 1.0962175 2.9151059
#> NRBP1         2.8967740 0.7575972 2.2038812
#> TCF7          3.2108187 0.4297757 2.4539445
#> HNRNPA3       3.5535247 2.4968528 3.6278586
#> S100A8        1.7859132 5.8229607 0.4722344
#> S100A9        2.9842995 6.4878861 1.8911084
#> LYZ           3.8136749 6.8958254 5.5725155
#> CD14          0.0000000 3.3033428 0.7917004
#> FCN1          3.1456147 4.8012315 3.5595998
#> TYROBP        4.7025251 5.8844609 3.4805484
#> ASGR1         0.0000000 2.5377164 0.5294697
#> NFKBIA        3.7221964 4.5567472 3.8382252
#> TYMP          2.2608012 4.8000897 3.0144503
#> CTSS          3.9069989 5.3329072 4.3967249
#> TSPO          3.3427492 4.8047808 3.3677127
#> RBP7          0.0000000 2.8941633 2.0363024
#> CTSB          1.8068070 3.7664341 1.6965705
#> LGALS1        4.5012015 5.7810064 4.7091092
#> FPR1          0.4867611 3.0293458 0.0000000
#> VSTM1         0.4867611 2.2700517 0.0000000
#> BLVRA         2.4163949 3.5189677 2.6475970
#> MPEG1         0.4233656 2.8003527 2.2255376
#> BID           2.2317326 4.1298828 2.4159639
#> SMCO4         0.0000000 2.9618122 1.5657879
#> CFD           1.8163037 4.1952415 2.5620920
#> LINC00936     2.2144252 3.3835009 2.2451806
#> LGALS2        2.0684178 4.2025588 2.3837466
#> MS4A6A        0.0000000 3.5842731 2.4584348
#> FCGRT         2.0376802 4.0528464 1.8417437
#> LGALS3        2.2453707 3.9846461 1.8658930
#> NUP214        1.4511392 3.1883697 1.0438974
#> SCO2          0.8127400 2.8570737 1.6647029
#> IL17RA        0.0000000 2.3130623 0.0000000
#> IFI6          3.0315004 4.0703482 1.5307061
#> HLA-DPA1      2.6017167 5.4406246 5.8418204
#> FCER1A        1.0226764 3.1908453 3.0061433
#> CLEC10A       0.0000000 3.0585437 1.8849018
#> HLA-DMA       1.7873288 3.7176150 3.9739660
#> RGS1          1.6752777 2.1703532 1.8313991
#> HLA-DPB1      1.9568190 5.5042352 6.4019650
#> HLA-DQA1      1.1591511 4.0771719 4.9841036
#> RNF130        1.3505966 3.4070770 1.2917444
#> HLA-DRB5      1.8862138 4.5942708 5.2075721
#> HLA-DRB1      1.9738282 5.2989581 6.1165475
#> CST3          3.8445027 6.3972718 5.5136635
#> IL1B          0.9758566 3.8165354 0.8208081
#> POP7          1.8011671 3.6127395 1.5962229
#> HLA-DQA2      0.0000000 2.7757140 3.7136018
#> CD1C          0.0000000 1.7851541 1.5782104
#> GSTP1         3.2010286 4.5645496 4.0945985
#> EIF3G         3.9932235 3.6772649 3.8487774
#> VPS28         3.2781671 3.8822700 3.0065124
#> LY86          0.8373655 2.3705157 4.0272398
#> ZFP36L1       2.5433199 2.7796699 2.7430907
#> ZNF330        1.0226764 3.5588959 1.9335748
#> ANXA2         3.7624813 4.2363047 3.7797193
#> GRN           2.0495073 3.8966757 2.7283129
#> CFP           2.2382154 4.0989617 2.9965578
#> HSP90AA1      4.4349415 3.0769802 4.7311639
#> FUOM          1.3291359 1.5756518 1.2365866
#> LST1          3.5075083 5.6295119 4.2177054
#> AIF1          3.8110868 5.6135533 4.0967432
#> PSAP          3.9559775 4.8981360 4.0489287
#> YWHAB         3.6793221 4.3038964 3.8411272
#> MYO1G         3.1351910 3.9974149 3.3368753
#> SAT1          4.7748986 5.7911395 4.3456705
#> RGS2          2.8153806 3.8514532 4.3140983
#> SERPINA1      1.7949444 4.0892610 2.2285155
#> IFITM3        2.3478669 4.5703041 3.1061576
#> FCGR3A        3.8298907 4.2203261 2.8151298
#> LILRA3        0.4867611 1.9941152 2.5144115
#> S100A11       3.9133848 5.1743051 3.6051760
#> FCER1G        4.4362056 5.4779208 3.4817679
#> TNFRSF1B      1.6269903 2.8131341 2.1903217
#> IFITM2        5.1716444 5.0628057 4.7230454
#> WARS          2.2003492 2.9326544 2.4996139
#> IFI30         1.5607155 3.7042918 2.6339778
#> MS4A7         1.4909477 2.9009766 2.5522460
#> C5AR1         0.0000000 3.1105490 2.1486048
#> HCK           1.4195331 3.3298567 1.4806266
#> COTL1         4.4406965 5.7008031 4.9137513
#> LGALS9        2.1203927 3.0626285 1.2636454
#> CD68          2.3846961 3.8223409 2.3669400
#> RP11-290F20.3 1.9022567 3.0422327 1.5889567
#> RHOC          3.3538790 3.5123180 2.1421543
#> CARD16        2.7519244 3.8387294 2.8785233
#> LRRC25        0.8127400 3.0161754 2.9092311
#> COPS6         3.0552960 3.6137904 1.6845487
#> ADAR          2.4482380 3.6456973 3.1996590
#> PPBP          6.3361534 2.4985104 1.3232427
#> GPX1          5.3975295 4.9344494 3.9999043
#> TPM4          3.9650277 2.4310243 2.4376107
#> PF4           5.2628190 0.0000000 1.0438974
#> SDPR          4.6187509 1.4891840 0.5294697
#> NRGN          3.5519248 2.2159806 0.0000000
#> SPARC         3.8611297 1.7167882 0.0000000
#> GNG11         4.7290162 0.9994320 0.0000000
#> CLU           4.5716300 1.4891840 0.0000000
#> HIST1H2AC     4.0955950 0.9994320 1.5889567
#> NCOA4         4.2857970 2.0987713 1.8015084
#> GP9           3.8473106 0.0000000 0.0000000
#> FERMT3        3.8984137 2.3881841 1.7019216
#> ODC1          3.5813379 1.6348387 2.3661512
#> CD9           4.4128429 1.8237111 0.0000000
#> RUFY1         3.2944391 0.7098036 2.4692572
#> TUBB1         4.3406225 0.0000000 0.0000000
#> TALDO1        4.4870900 3.9153687 2.8007831
#> TREML1        3.6457763 1.4891840 0.0000000
#> NGFRAP1       3.4490426 0.0000000 0.0000000
#> PGRMC1        3.2643962 0.0000000 0.0000000
#> CA2           3.8930121 0.7972874 0.0000000
#> ITGA2B        3.4370738 0.7575972 0.0000000
#> MYL9          3.3849450 0.9994320 0.0000000
#> TMEM40        2.9969736 0.0000000 0.0000000
#> PARVB         3.1059711 1.4622407 1.5657879
#> PTCRA         3.3644245 0.0000000 0.0000000
#> ACRBP         3.3417653 0.9994320 0.0000000
#> TSC22D1       3.3077263 0.9994320 0.0000000
#> VDAC3         4.5096924 1.7487485 4.4015913
#> GZMB          4.9108339 1.4445172 2.3178006
#> GZMA          4.8910087 0.8628322 3.9922168
#> GNLY          5.7198145 0.5860144 1.2636454
#> FGFBP2        4.6123510 1.8347624 3.5423195
#> AKR1C3        3.4104783 0.0000000 0.0000000
#> CCL4          3.3503552 0.8511256 2.6898847
#> PRF1          4.9437954 0.8628322 4.0367279
#> GZMH          4.2888180 0.0000000 2.6898847
#> XBP1          3.8863898 1.4543988 2.6614702
#> GZMM          4.0156492 0.0000000 3.3093286
#> PTGDR         3.9660372 0.0000000 2.2171973
#> IGFBP7        3.4925100 2.5527317 1.0334405
#> TTC38         2.5124157 0.6433972 1.7185899
#> KLRD1         3.0392705 0.7457554 0.0000000
#> ARHGDIA       3.6343771 2.9091133 3.1032747
#> IL2RB         2.7771354 0.0000000 1.5889567
#> CLIC3         3.0638541 0.7098036 1.8263295
#> PPP1R18       3.5865558 1.9511920 3.6241193
#> CD247         3.6905393 0.7575972 2.2367768
#> ALOX5AP       3.6098397 1.6641347 3.6493268
#> XCL2          2.6177158 0.0000000 2.2367768
#> C12orf75      3.2237835 0.4297757 2.3057666
#> RARRES3       4.3416989 1.7550597 3.7945663
#> PCMT1         4.5763104 2.8893691 2.2231406
#> LAMP1         3.9191289 0.9441022 1.5657879
#> SPON2         3.6704971 1.3182933 3.2007957
#> S100B         3.0692104 0.0000000 0.0000000